Module Wayland

module Display : sig ... end

Connect to a Wayland server and interact with the initial display object.

module Registry : sig ... end

A simple wrapper for the Wayland_client.Wl_registry API.

val callback : (int32 -> unit) -> ([ `Wl_callback ], 'v, 'v, [ `V1 ]) Proxy.handler

callback fn is a simple wrapper for the Wayland_client.Wl_callback API. It calls fn data when the callback's "done" signal is received. Wl_callback seems to be an exception to the usual Wayland versioning rules (a wl_callback can be created by multiple objects).

module Fixed : sig ... end

Wayland's 24.8 fixed-point type.

module Log : sig ... end

Generated bindings for the core Wayland protocol

module Wayland_client : sig ... end

Bindings for use by Wayland clients.

module Wayland_server : sig ... end

Bindings for use by Wayland servers.

module Wayland_proto : sig ... end

Type metadata for use by bindings.

Low-level API

module Proxy : sig ... end

Low-level operations on objects, used by the generated wrappers.

module Metadata : sig ... end

Describing interfaces.

module Msg : sig ... end

Building and parsing messages.