Class type Handler.t

An ('a, 'v, 'role) t handles incoming messages for an object of type 'a.

Typically, a constructor will let the user pick from a range of versions for 'v, which will then be constrained by the spawn call. The generated bindings provide implementations of this where dispatch collects the arguments and passes them to an appopriately-typed virtual method, which must be implmented by the application.

method user_data : ( 'a, 'role ) Wayland__S.user_data

method metadata : (module Wayland__Metadata.S with type t = 'a)

method dispatch : ( 'a, 'v, 'role ) proxy -> ( 'a, [ `R ] ) Wayland__Msg.t -> unit