Class Wl_keyboard.h4

inherit h3

method virtual on_repeat_info : 'v t -> rate:int32 -> delay:int32 -> unit

Repeat rate and delay.

Informs the client about the keyboard's repeat rate and delay.

This event is sent as soon as the wl_keyboard object has been created, and is guaranteed to be received by the client before any key press event.

Negative values for either rate or delay are illegal. A rate of zero will disable any repeating (regardless of the value of delay).

This event can be sent later on as well with a new value if necessary, so clients should continue listening for the event past the creation of wl_keyboard.