Module Wayland_client.Wl_output

Compositor output region.

An output describes part of the compositor geometry. The compositor works in the 'compositor coordinate system' and an output corresponds to a rectangular area in that space that is actually visible. This typically corresponds to a monitor that displays part of the compositor space. This object is published as global during start up, or when a monitor is hotplugged.

type 'v t = ([ `Wl_output ], 'v, [ `Client ]) Proxy.t

module Subpixel = Wayland_proto.Wl_output.Subpixel

module Transform = Wayland_proto.Wl_output.Transform

module Mode = Wayland_proto.Wl_output.Mode

Version 1

Version 2

Version 3

val release : [< `V3 ] t -> unit

Release the output object.

Using this request a client can tell the server that it is not going to use the output object anymore.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.