Class Wl_pointer.h1

method virtual on_axis : 'v t -> time:int32 -> axis:int32 -> value:Fixed.t -> unit

Axis event.

Scroll and other axis notifications.

For scroll events (vertical and horizontal scroll axes), the value parameter is the length of a vector along the specified axis in a coordinate space identical to those of motion events, representing a relative movement along the specified axis.

For devices that support movements non-parallel to axes multiple axis events will be emitted.

When applicable, for example for touch pads, the server can choose to emit scroll events where the motion vector is equivalent to a motion event vector.

When applicable, a client can transform its content relative to the scroll distance.

method virtual on_button : 'v t -> serial:int32 -> time:int32 -> button:int32 -> state:int32 -> unit

Pointer button event.

Mouse button click and release notifications.

The location of the click is given by the last motion or enter event. The time argument is a timestamp with millisecond granularity, with an undefined base.

The button is a button code as defined in the Linux kernel's linux/input-event-codes.h header file, e.g. BTN_LEFT.

Any 16-bit button code value is reserved for future additions to the kernel's event code list. All other button codes above 0xFFFF are currently undefined but may be used in future versions of this protocol.

method virtual on_motion : 'v t -> time:int32 -> surface_x:Fixed.t -> surface_y:Fixed.t -> unit

Pointer motion event.

Notification of pointer location change. The arguments surface_x and surface_y are the location relative to the focused surface.

method virtual on_leave : 'v t -> serial:int32 -> surface:int32 -> unit

Leave event.

Notification that this seat's pointer is no longer focused on a certain surface.

The leave notification is sent before the enter notification for the new focus.

method virtual on_enter : 'v t -> serial:int32 -> surface:int32 -> surface_x:Fixed.t -> surface_y:Fixed.t -> unit

Enter event.

Notification that this seat's pointer is focused on a certain surface.

When a seat's focus enters a surface, the pointer image is undefined and a client should respond to this event by setting an appropriate pointer image with the set_cursor request.