Class Wl_registry.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_global : [> `V1 ] t -> name:int32 -> interface:string -> version:int32 -> unit

Announce global object.

Notify the client of global objects.

The event notifies the client that a global object with the given name is now available, and it implements the given version of the given interface.

method private virtual on_global_remove : [> `V1 ] t -> name:int32 -> unit

Announce removal of global object.

Notify the client of removed global objects.

This event notifies the client that the global identified by name is no longer available. If the client bound to the global using the bind request, the client should now destroy that object.

The object remains valid and requests to the object will be ignored until the client destroys it, to avoid races between the global going away and a client sending a request to it.

method min_version : 'a v1 -> int32