Module Wayland_client.Wl_seat

Group of input devices.

A seat is a group of keyboards, pointer and touch devices. This object is published as a global during start up, or when such a device is hot plugged. A seat typically has a pointer and maintains a keyboard focus and a pointer focus.

type 'v t = ( [ `Wl_seat ], 'v, [ `Client ] ) Proxy.t

module Capability = Wayland_proto.Wl_seat.Capability

module Error = Wayland_proto.Wl_seat.Error

Version 1

val get_touch : [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 ] as 'a t -> [ [ `Wl_touch ], 'b, [ `Client ] ] Proxy.Handler.t -> ( [ `Wl_touch ], 'c, [ `Client ] ) Proxy.t

Return touch object.

The ID provided will be initialized to the wl_touch interface for this seat.

This request only takes effect if the seat has the touch capability, or has had the touch capability in the past. It is a protocol violation to issue this request on a seat that has never had the touch capability. The missing_capability error will be sent in this case.

val get_keyboard : [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 ] as 'a t -> [ [ `Wl_keyboard ], 'b, [ `Client ] ] Proxy.Handler.t -> ( [ `Wl_keyboard ], 'c, [ `Client ] ) Proxy.t

Return keyboard object.

The ID provided will be initialized to the wl_keyboard interface for this seat.

This request only takes effect if the seat has the keyboard capability, or has had the keyboard capability in the past. It is a protocol violation to issue this request on a seat that has never had the keyboard capability. The missing_capability error will be sent in this case.

val get_pointer : [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 ] as 'a t -> [ [ `Wl_pointer ], 'b, [ `Client ] ] Proxy.Handler.t -> ( [ `Wl_pointer ], 'c, [ `Client ] ) Proxy.t

Return pointer object.

The ID provided will be initialized to the wl_pointer interface for this seat.

This request only takes effect if the seat has the pointer capability, or has had the pointer capability in the past. It is a protocol violation to issue this request on a seat that has never had the pointer capability. The missing_capability error will be sent in this case.

Version 2, 3, 4

Version 5, 6, 7, 8

val release : [< `V5 | `V6 | `V7 | `V8 ] t -> unit

Release the seat object.

Using this request a client can tell the server that it is not going to use the seat object anymore.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.

class virtual +'a v4 : object ... end

Handler for a proxy with version >= 4.

class virtual +'a v5 : object ... end

Handler for a proxy with version >= 5.

class virtual +'a v6 : object ... end

Handler for a proxy with version >= 6.

class virtual +'a v7 : object ... end

Handler for a proxy with version >= 7.

class virtual +'a v8 : object ... end

Handler for a proxy with version >= 8.