Class Wl_seat._handlers_unsafe

method user_data : [Wayland_proto.Wl_seat.t, 'v, [ `Client ]] Proxy.Handler.t -> (Wayland_proto.Wl_seat.t, [ `Client ]) S.user_data

method metadata : [Wayland_proto.Wl_seat.t, 'v, [ `Client ]] Proxy.Handler.t -> (module Metadata.S with type t = Wayland_proto.Wl_seat.t)

method private virtual on_capabilities : 'v t -> capabilities:$Imports.Wl_seat.Capability.t -> unit

Seat capabilities changed.

This is emitted whenever a seat gains or loses the pointer, keyboard or touch capabilities. The argument is a capability enum containing the complete set of capabilities this seat has.

When the pointer capability is added, a client may create a wl_pointer object using the wl_seat.get_pointer request. This object will receive pointer events until the capability is removed in the future.

When the pointer capability is removed, a client should destroy the wl_pointer objects associated with the seat where the capability was removed, using the wl_pointer.release request. No further pointer events will be received on these objects.

In some compositors, if a seat regains the pointer capability and a client has a previously obtained wl_pointer object of version 4 or less, that object may start sending pointer events again. This behavior is considered a misinterpretation of the intended behavior and must not be relied upon by the client. wl_pointer objects of version 5 or later must not send events if created before the most recent event notifying the client of an added pointer capability.

The above behavior also applies to wl_keyboard and wl_touch with the keyboard and touch capabilities, respectively.

method private virtual on_name : 'v t -> name:string -> unit

Unique identifier for this seat.

In a multiseat configuration this can be used by the client to help identify which physical devices the seat represents. Based on the seat configuration used by the compositor.

method dispatch : [Wayland_proto.Wl_seat.t, 'v, [ `Client ]] Proxy.Handler.t -> (Wayland_proto.Wl_seat.t, 'v, [ `Client ]) Proxy.t -> (Wayland_proto.Wl_seat.t, [ `R ]) Msg.t -> unit