Class Wl_seat.v2

Handler for a proxy with version >= 2.

constraint 'a = [< `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ]

method private virtual on_capabilities : [> `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> capabilities:{Imports}1.Wl_seat.Capability.t -> unit

Seat capabilities changed.

This is emitted whenever a seat gains or loses the pointer, keyboard or touch capabilities. The argument is a capability enum containing the complete set of capabilities this seat has.

When the pointer capability is added, a client may create a wl_pointer object using the wl_seat.get_pointer request. This object will receive pointer events until the capability is removed in the future.

When the pointer capability is removed, a client should destroy the wl_pointer objects associated with the seat where the capability was removed, using the wl_pointer.release request. No further pointer events will be received on these objects.

In some compositors, if a seat regains the pointer capability and a client has a previously obtained wl_pointer object of version 4 or less, that object may start sending pointer events again. This behavior is considered a misinterpretation of the intended behavior and must not be relied upon by the client. wl_pointer objects of version 5 or later must not send events if created before the most recent event notifying the client of an added pointer capability.

The above behavior also applies to wl_keyboard and wl_touch with the keyboard and touch capabilities, respectively.

method private virtual on_name : [> `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> name:string -> unit

Unique identifier for this seat.

In a multi-seat configuration the seat name can be used by clients to help identify which physical devices the seat represents.

The seat name is a UTF-8 string with no convention defined for its contents. Each name is unique among all wl_seat globals. The name is only guaranteed to be unique for the current compositor instance.

The same seat names are used for all clients. Thus, the name can be shared across processes to refer to a specific wl_seat global.

The name event is sent after binding to the seat global. This event is only sent once per seat object, and the name does not change over the lifetime of the wl_seat global.

Compositors may re-use the same seat name if the wl_seat global is destroyed and re-created later.

method min_version : int32

method bind_version : [ `V2 ]