Module Wayland_client.Wl_shm_pool

A shared memory pool.

The wl_shm_pool object encapsulates a piece of memory shared between the compositor and client. Through the wl_shm_pool object, the client can allocate shared memory wl_buffer objects. All objects created through the same pool share the same underlying mapped memory. Reusing the mapped memory avoids the setup/teardown overhead and is useful when interactively resizing a surface or for many small buffers.

type 'v t = ([ `Wl_shm_pool ], 'v, [ `Client ]) Proxy.t

Version 1, 2

val resize : [< `V1 | `V2 ] t -> size:int32 -> unit

Change the size of the pool mapping.

This request will cause the server to remap the backing memory for the pool from the file descriptor passed when the pool was created, but using the new size. This request can only be used to make the pool bigger.

This request only changes the amount of bytes that are mmapped by the server and does not touch the file corresponding to the file descriptor passed at creation time. It is the client's responsibility to ensure that the file is at least as big as the new pool size.

val destroy : [< `V1 | `V2 ] t -> unit

Destroy the pool.

Destroy the shared memory pool.

The mmapped memory will be released when all buffers that have been created from this pool are gone.

val create_buffer : [< `V1 | `V2 ] t -> [[ `Wl_buffer ], [ `V1 ], [ `Client ]] Proxy.Handler.t -> offset:int32 -> width:int32 -> height:int32 -> stride:int32 -> format:{Imports}1.Wl_shm.Format.t -> ([ `Wl_buffer ], [ `V1 ], [ `Client ]) Proxy.t

Create a buffer from the pool.

Create a wl_buffer object from the pool.

The buffer is created offset bytes into the pool and has width and height as specified. The stride argument specifies the number of bytes from the beginning of one row to the beginning of the next. The format is the pixel format of the buffer and must be one of those advertised through the wl_shm.format event.

A buffer will keep a reference to the pool it was created from so it is valid to destroy the pool immediately after creating a buffer from it.

Handlers

Note: Servers will always want to use v1.

class +'a v1 : object ... end

Handler for a proxy with version >= 1.

class +'a v2 : object ... end

Handler for a proxy with version >= 2.