Class Wl_surface.v4

Handler for a proxy with version >= 4.

constraint 'a = [< `V4 | `V5 ]

method private virtual on_enter : [> `V4 | `V5 ] t -> output: ( [ `Wl_output ], [> Wayland__Wayland_proto.Wl_output.versions ], [ `Client ] ) Wayland__Proxy.t -> unit

Surface enters an output.

This is emitted whenever a surface's creation, movement, or resizing results in some part of it being within the scanout region of an output.

Note that a surface may be overlapping with zero or more outputs.

method private virtual on_leave : [> `V4 | `V5 ] t -> output: ( [ `Wl_output ], [> Wayland__Wayland_proto.Wl_output.versions ], [ `Client ] ) Wayland__Proxy.t -> unit

Surface leaves an output.

This is emitted whenever a surface's creation, movement, or resizing results in it no longer having any part of it within the scanout region of an output.

Clients should not use the number of outputs the surface is on for frame throttling purposes. The surface might be hidden even if no leave event has been sent, and the compositor might expect new surface content updates even if no enter event has been sent. The frame event should be used instead.

method min_version : 'a v4 -> int32