Class Wl_touch.h1

method virtual on_cancel : 'v t -> unit

Touch session cancelled.

Sent if the compositor decides the touch stream is a global gesture. No further events are sent to the clients from that particular gesture. Touch cancellation applies to all touch points currently active on this client's surface. The client is responsible for finalizing the touch points, future touch points on this surface may reuse the touch point ID.

method virtual on_frame : 'v t -> unit

End of touch frame event.

Indicates the end of a set of events that logically belong together. A client is expected to accumulate the data in all events within the frame before proceeding.

A wl_touch.frame terminates at least one event but otherwise no guarantee is provided about the set of events within a frame. A client must assume that any state not updated in a frame is unchanged from the previously known state.

method virtual on_motion : 'v t -> time:int32 -> id:int32 -> x:Fixed.t -> y:Fixed.t -> unit

Update of touch point coordinates.

A touch point has changed coordinates.

method virtual on_up : 'v t -> serial:int32 -> time:int32 -> id:int32 -> unit

End of a touch event sequence.

The touch point has disappeared. No further events will be sent for this touch point and the touch point's ID is released and may be reused in a future touch down event.

method virtual on_down : 'v t -> serial:int32 -> time:int32 -> surface:int32 -> id:int32 -> x:Fixed.t -> y:Fixed.t -> unit

Touch down event and beginning of a touch sequence.

A new touch point has appeared on the surface. This touch point is assigned a unique ID. Future events from this touch point reference this ID. The ID ceases to be valid after a touch up event and may be reused in the future.