Class Wl_touch.v3

Handler for a proxy with version >= 3.

constraint 'a = [< `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ]

method private virtual on_down : [> `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> serial:int32 -> time:int32 -> surface: ([ `Wl_surface ], [> {Imports}1.Wl_surface.versions ], [ `Client ]) Wayland__.Proxy.t -> id:int32 -> x:Wayland__.Fixed.t -> y:Wayland__.Fixed.t -> unit

Touch down event and beginning of a touch sequence.

A new touch point has appeared on the surface. This touch point is assigned a unique ID. Future events from this touch point reference this ID. The ID ceases to be valid after a touch up event and may be reused in the future.

method private virtual on_up : [> `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> serial:int32 -> time:int32 -> id:int32 -> unit

End of a touch event sequence.

The touch point has disappeared. No further events will be sent for this touch point and the touch point's ID is released and may be reused in a future touch down event.

method private virtual on_motion : [> `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> time:int32 -> id:int32 -> x:Wayland__.Fixed.t -> y:Wayland__.Fixed.t -> unit

Update of touch point coordinates.

A touch point has changed coordinates.

method private virtual on_frame : [> `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> unit

End of touch frame event.

Indicates the end of a set of events that logically belong together. A client is expected to accumulate the data in all events within the frame before proceeding.

A wl_touch.frame terminates at least one event but otherwise no guarantee is provided about the set of events within a frame. A client must assume that any state not updated in a frame is unchanged from the previously known state.

method private virtual on_cancel : [> `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> unit

Touch session cancelled.

Sent if the compositor decides the touch stream is a global gesture. No further events are sent to the clients from that particular gesture. Touch cancellation applies to all touch points currently active on this client's surface. The client is responsible for finalizing the touch points, future touch points on this surface may reuse the touch point ID.

No frame event is required after the cancel event.

method private virtual on_shape : [> `V6 | `V7 | `V8 | `V9 ] t -> id:int32 -> major:Wayland__.Fixed.t -> minor:Wayland__.Fixed.t -> unit

Update shape of touch point.

Sent when a touchpoint has changed its shape.

This event does not occur on its own. It is sent before a wl_touch.frame event and carries the new shape information for any previously reported, or new touch points of that frame.

Other events describing the touch point such as wl_touch.down, wl_touch.motion or wl_touch.orientation may be sent within the same wl_touch.frame. A client should treat these events as a single logical touch point update. The order of wl_touch.shape, wl_touch.orientation and wl_touch.motion is not guaranteed. A wl_touch.down event is guaranteed to occur before the first wl_touch.shape event for this touch ID but both events may occur within the same wl_touch.frame.

A touchpoint shape is approximated by an ellipse through the major and minor axis length. The major axis length describes the longer diameter of the ellipse, while the minor axis length describes the shorter diameter. Major and minor are orthogonal and both are specified in surface-local coordinates. The center of the ellipse is always at the touchpoint location as reported by wl_touch.down or wl_touch.move.

This event is only sent by the compositor if the touch device supports shape reports. The client has to make reasonable assumptions about the shape if it did not receive this event.

method private virtual on_orientation : [> `V6 | `V7 | `V8 | `V9 ] t -> id:int32 -> orientation:Wayland__.Fixed.t -> unit

Update orientation of touch point.

Sent when a touchpoint has changed its orientation.

This event does not occur on its own. It is sent before a wl_touch.frame event and carries the new shape information for any previously reported, or new touch points of that frame.

Other events describing the touch point such as wl_touch.down, wl_touch.motion or wl_touch.shape may be sent within the same wl_touch.frame. A client should treat these events as a single logical touch point update. The order of wl_touch.shape, wl_touch.orientation and wl_touch.motion is not guaranteed. A wl_touch.down event is guaranteed to occur before the first wl_touch.orientation event for this touch ID but both events may occur within the same wl_touch.frame.

The orientation describes the clockwise angle of a touchpoint's major axis to the positive surface y-axis and is normalized to the -180 to +180 degree range. The granularity of orientation depends on the touch device, some devices only support binary rotation values between 0 and 90 degrees.

This event is only sent by the compositor if the touch device supports orientation reports.

method min_version : int32