Module Wl_output.Transform

Transform from framebuffer to output.

This describes the transform that a compositor will apply to a surface to compensate for the rotation or mirroring of an output device.

The flipped values correspond to an initial flip around a vertical axis followed by rotation.

The purpose is mainly to allow clients to render accordingly and tell the compositor, so that for fullscreen surfaces, the compositor will still be able to scan out directly from client surfaces.

type t =

| Normal : t

| V90 : t

| V180 : t

| V270 : t

| Flipped : t

| Flipped_90 : t

| Flipped_180 : t

| Flipped_270 : t

val to_int32 : t -> int32

val of_int32 : int32 -> t