Module Wl_seat.Capability

Seat capability bitmask.

This is a bitmask of capabilities this seat has; if a member is set, then it is present on the seat.

type t = int32

val pointer : int32

val keyboard : int32

val touch : int32

val to_int32 : 'a -> 'a

val of_int32 : 'a -> 'a