Class Wl_buffer.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_destroy : [> `V1 ] t -> unit

Destroy a buffer.

Destroy a buffer. If and how you need to release the backing storage is defined by the buffer factory interface.

For possible side-effects to a surface, see wl_surface.attach.

method min_version : int32