Module Wayland_server.Wl_callback

Callback object.

Clients can handle the 'done' event to get notified when the related request is done.

Note, because wl_callback objects are created from multiple independent factory interfaces, the wl_callback interface is frozen at version 1.

type 'v t = ([ `Wl_callback ], 'v, [ `Server ]) Proxy.t

Version 1

val done_ : [< `V1 ] t -> callback_data:int32 -> unit

Done event.

Notify the client when the related request is done.

Handlers

Note: Servers will always want to use v1.

class +'a v1 : object ... end

Handler for a proxy with version >= 1.