Class Wl_data_device.handlers

method user_data : [Wayland_proto.Wl_data_device.t, 'v, [ `Server ]] Proxy.Handler.t -> (Wayland_proto.Wl_data_device.t, [ `Server ]) S.user_data

method metadata : [Wayland_proto.Wl_data_device.t, 'v, [ `Server ]] Proxy.Handler.t -> (module Metadata.S with type t = Wayland_proto.Wl_data_device.t)

method private virtual on_start_drag : 'v t -> source:([ `Wl_data_source ], 'v, [ `Server ]) Proxy.t option -> origin:([ `Wl_surface ], 'v, [ `Server ]) Proxy.t -> icon:([ `Wl_surface ], 'v, [ `Server ]) Proxy.t option -> serial:int32 -> unit

Start drag-and-drop operation.

This request asks the compositor to start a drag-and-drop operation on behalf of the client.

The source argument is the data source that provides the data for the eventual data transfer. If source is NULL, enter, leave and motion events are sent only to the client that initiated the drag and the client is expected to handle the data passing internally. If source is destroyed, the drag-and-drop session will be cancelled.

The origin surface is the surface where the drag originates and the client must have an active implicit grab that matches the serial.

The icon surface is an optional (can be NULL) surface that provides an icon to be moved around with the cursor. Initially, the top-left corner of the icon surface is placed at the cursor hotspot, but subsequent wl_surface.attach request can move the relative position. Attach requests must be confirmed with wl_surface.commit as usual. The icon surface is given the role of a drag-and-drop icon. If the icon surface already has another role, it raises a protocol error.

The current and pending input regions of the icon wl_surface are cleared, and wl_surface.set_input_region is ignored until the wl_surface is no longer used as the icon surface. When the use as an icon ends, the current and pending input regions become undefined, and the wl_surface is unmapped.

method private virtual on_set_selection : 'v t -> source:([ `Wl_data_source ], 'v, [ `Server ]) Proxy.t option -> serial:int32 -> unit

Copy data to the selection.

This request asks the compositor to set the selection to the data from the source on behalf of the client.

To unset the selection, set the source to NULL.

method private virtual on_release : 'v t -> unit

Destroy data device.

This request destroys the data device.

method dispatch : [Wayland_proto.Wl_data_device.t, 'v, [ `Server ]] Proxy.Handler.t -> (Wayland_proto.Wl_data_device.t, 'v, [ `Server ]) Proxy.t -> (Wayland_proto.Wl_data_device.t, [ `R ]) Msg.t -> unit