Class Wl_data_device.v3

Handler for a proxy with version >= 3.

constraint 'a = [< `V3 ]

method private virtual on_start_drag : [> `V3 ] t -> source: ([ `Wl_data_source ], [> {Imports}1.Wl_data_source.versions ], [ `Server ]) Wayland__.Proxy.t option -> origin: ([ `Wl_surface ], [> {Imports}1.Wl_surface.versions ], [ `Server ]) Wayland__.Proxy.t -> icon: ([ `Wl_surface ], [> {Imports}1.Wl_surface.versions ], [ `Server ]) Wayland__.Proxy.t option -> serial:int32 -> unit

Start drag-and-drop operation.

This request asks the compositor to start a drag-and-drop operation on behalf of the client.

The source argument is the data source that provides the data for the eventual data transfer. If source is NULL, enter, leave and motion events are sent only to the client that initiated the drag and the client is expected to handle the data passing internally. If source is destroyed, the drag-and-drop session will be cancelled.

The origin surface is the surface where the drag originates and the client must have an active implicit grab that matches the serial.

The icon surface is an optional (can be NULL) surface that provides an icon to be moved around with the cursor. Initially, the top-left corner of the icon surface is placed at the cursor hotspot, but subsequent wl_surface.attach request can move the relative position. Attach requests must be confirmed with wl_surface.commit as usual. The icon surface is given the role of a drag-and-drop icon. If the icon surface already has another role, it raises a protocol error.

The input region is ignored for wl_surfaces with the role of a drag-and-drop icon.

The given source may not be used in any further set_selection or start_drag requests. Attempting to reuse a previously-used source may send a used_source error.

method private virtual on_set_selection : [> `V3 ] t -> source: ([ `Wl_data_source ], [> {Imports}1.Wl_data_source.versions ], [ `Server ]) Wayland__.Proxy.t option -> serial:int32 -> unit

Copy data to the selection.

This request asks the compositor to set the selection to the data from the source on behalf of the client.

To unset the selection, set the source to NULL.

The given source may not be used in any further set_selection or start_drag requests. Attempting to reuse a previously-used source may send a used_source error.

method private virtual on_release : [> `V3 ] t -> unit

Destroy data device.

This request destroys the data device.

method min_version : int32