Class Wl_data_source.v3

Handler for a proxy with version >= 3.

constraint 'b = [< `V3 ]

method private virtual on_offer : [> `V3 ] t -> mime_type:string -> unit

Add an offered mime type.

This request adds a mime type to the set of mime types advertised to targets. Can be called several times to offer multiple types.

method private virtual on_destroy : [> `V3 ] t -> unit

Destroy the data source.

Destroy the data source.

method private virtual on_set_actions : [> `V3 ] t -> dnd_actions:Wayland__Wayland_proto.Wl_data_device_manager.Dnd_action.t -> unit

Set the available drag-and-drop actions.

Sets the actions that the source side client supports for this operation. This request may trigger wl_data_source.action and wl_data_offer.action events if the compositor needs to change the selected action.

The dnd_actions argument must contain only values expressed in the wl_data_device_manager.dnd_actions enum, otherwise it will result in a protocol error.

This request must be made once only, and can only be made on sources used in drag-and-drop, so it must be performed before wl_data_device.start_drag. Attempting to use the source other than for drag-and-drop will raise a protocol error.

method min_version : [Wayland__Wayland_proto.Wl_data_source.t, 'b, [ `Server ]] Wayland__.Proxy.Service_handler.t -> int32