Module Wayland_server.Wl_pointer

Pointer input device.

The wl_pointer interface represents one or more input devices, such as mice, which control the pointer location and pointer_focus of a seat.

The wl_pointer interface generates motion, enter and leave events for the surfaces that the pointer is located over, and button and axis events for button presses, button releases and scrolling.

type 'v t = ([ `Wl_pointer ], 'v, [ `Server ]) Proxy.t

module Error = Wayland_proto.Wl_pointer.Error

module Button_state = Wayland_proto.Wl_pointer.Button_state

module Axis = Wayland_proto.Wl_pointer.Axis

module Axis_source = Wayland_proto.Wl_pointer.Axis_source

module Axis_relative_direction = Wayland_proto.Wl_pointer.Axis_relative_direction

Version 1, 2

val axis : [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> time:int32 -> axis:{Imports}1.Wl_pointer.Axis.t -> value:Fixed.t -> unit

Axis event.

Scroll and other axis notifications.

For scroll events (vertical and horizontal scroll axes), the value parameter is the length of a vector along the specified axis in a coordinate space identical to those of motion events, representing a relative movement along the specified axis.

For devices that support movements non-parallel to axes multiple axis events will be emitted.

When applicable, for example for touch pads, the server can choose to emit scroll events where the motion vector is equivalent to a motion event vector.

When applicable, a client can transform its content relative to the scroll distance.

val button : [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> serial:int32 -> time:int32 -> button:int32 -> state:{Imports}1.Wl_pointer.Button_state.t -> unit

Pointer button event.

Mouse button click and release notifications.

The location of the click is given by the last motion or enter event. The time argument is a timestamp with millisecond granularity, with an undefined base.

The button is a button code as defined in the Linux kernel's linux/input-event-codes.h header file, e.g. BTN_LEFT.

Any 16-bit button code value is reserved for future additions to the kernel's event code list. All other button codes above 0xFFFF are currently undefined but may be used in future versions of this protocol.

val motion : [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> time:int32 -> surface_x:Fixed.t -> surface_y:Fixed.t -> unit

Pointer motion event.

Notification of pointer location change. The arguments surface_x and surface_y are the location relative to the focused surface.

val leave : [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> serial:int32 -> surface:([ `Wl_surface ], 'a, [ `Server ]) Proxy.t -> unit

Leave event.

Notification that this seat's pointer is no longer focused on a certain surface.

The leave notification is sent before the enter notification for the new focus.

val enter : [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 | `V8 | `V9 ] t -> serial:int32 -> surface:([ `Wl_surface ], 'a, [ `Server ]) Proxy.t -> surface_x:Fixed.t -> surface_y:Fixed.t -> unit

Enter event.

Notification that this seat's pointer is focused on a certain surface.

When a seat's focus enters a surface, the pointer image is undefined and a client should respond to this event by setting an appropriate pointer image with the set_cursor request.

Version 3, 4

Version 5, 6, 7

val axis_discrete : [< `V5 | `V6 | `V7 | `V8 | `V9 ] t -> axis:{Imports}1.Wl_pointer.Axis.t -> discrete:int32 -> unit

Axis click event.

Discrete step information for scroll and other axes.

This event carries the axis value of the wl_pointer.axis event in discrete steps (e.g. mouse wheel clicks).

This event is deprecated with wl_pointer version 8 - this event is not sent to clients supporting version 8 or later.

This event does not occur on its own, it is coupled with a wl_pointer.axis event that represents this axis value on a continuous scale. The protocol guarantees that each axis_discrete event is always followed by exactly one axis event with the same axis number within the same wl_pointer.frame. Note that the protocol allows for other events to occur between the axis_discrete and its coupled axis event, including other axis_discrete or axis events. A wl_pointer.frame must not contain more than one axis_discrete event per axis type.

This event is optional; continuous scrolling devices like two-finger scrolling on touchpads do not have discrete steps and do not generate this event.

The discrete value carries the directional information. e.g. a value of -2 is two steps towards the negative direction of this axis.

The axis number is identical to the axis number in the associated axis event.

The order of wl_pointer.axis_discrete and wl_pointer.axis_source is not guaranteed.

val axis_stop : [< `V5 | `V6 | `V7 | `V8 | `V9 ] t -> time:int32 -> axis:{Imports}1.Wl_pointer.Axis.t -> unit

Axis stop event.

Stop notification for scroll and other axes.

For some wl_pointer.axis_source types, a wl_pointer.axis_stop event is sent to notify a client that the axis sequence has terminated. This enables the client to implement kinetic scrolling. See the wl_pointer.axis_source documentation for information on when this event may be generated.

Any wl_pointer.axis events with the same axis_source after this event should be considered as the start of a new axis motion.

The timestamp is to be interpreted identical to the timestamp in the wl_pointer.axis event. The timestamp value may be the same as a preceding wl_pointer.axis event.

val axis_source : [< `V5 | `V6 | `V7 | `V8 | `V9 ] t -> axis_source:{Imports}1.Wl_pointer.Axis_source.t -> unit

Axis source event.

Source information for scroll and other axes.

This event does not occur on its own. It is sent before a wl_pointer.frame event and carries the source information for all events within that frame.

The source specifies how this event was generated. If the source is wl_pointer.axis_source.finger, a wl_pointer.axis_stop event will be sent when the user lifts the finger off the device.

If the source is wl_pointer.axis_source.wheel, wl_pointer.axis_source.wheel_tilt or wl_pointer.axis_source.continuous, a wl_pointer.axis_stop event may or may not be sent. Whether a compositor sends an axis_stop event for these sources is hardware-specific and implementation-dependent; clients must not rely on receiving an axis_stop event for these scroll sources and should treat scroll sequences from these scroll sources as unterminated by default.

This event is optional. If the source is unknown for a particular axis event sequence, no event is sent. Only one wl_pointer.axis_source event is permitted per frame.

The order of wl_pointer.axis_discrete and wl_pointer.axis_source is not guaranteed.

val frame : [< `V5 | `V6 | `V7 | `V8 | `V9 ] t -> unit

End of a pointer event sequence.

Indicates the end of a set of events that logically belong together. A client is expected to accumulate the data in all events within the frame before proceeding.

All wl_pointer events before a wl_pointer.frame event belong logically together. For example, in a diagonal scroll motion the compositor will send an optional wl_pointer.axis_source event, two wl_pointer.axis events (horizontal and vertical) and finally a wl_pointer.frame event. The client may use this information to calculate a diagonal vector for scrolling.

When multiple wl_pointer.axis events occur within the same frame, the motion vector is the combined motion of all events. When a wl_pointer.axis and a wl_pointer.axis_stop event occur within the same frame, this indicates that axis movement in one axis has stopped but continues in the other axis. When multiple wl_pointer.axis_stop events occur within the same frame, this indicates that these axes stopped in the same instance.

A wl_pointer.frame event is sent for every logical event group, even if the group only contains a single wl_pointer event. Specifically, a client may get a sequence: motion, frame, button, frame, axis, frame, axis_stop, frame.

The wl_pointer.enter and wl_pointer.leave events are logical events generated by the compositor and not the hardware. These events are also grouped by a wl_pointer.frame. When a pointer moves from one surface to another, a compositor should group the wl_pointer.leave event within the same wl_pointer.frame. However, a client must not rely on wl_pointer.leave and wl_pointer.enter being in the same wl_pointer.frame. Compositor-specific policies may require the wl_pointer.leave and wl_pointer.enter event being split across multiple wl_pointer.frame groups.

Version 8

val axis_value120 : [< `V8 | `V9 ] t -> axis:{Imports}1.Wl_pointer.Axis.t -> value120:int32 -> unit

Axis high-resolution scroll event.

Discrete high-resolution scroll information.

This event carries high-resolution wheel scroll information, with each multiple of 120 representing one logical scroll step (a wheel detent). For example, an axis_value120 of 30 is one quarter of a logical scroll step in the positive direction, a value120 of -240 are two logical scroll steps in the negative direction within the same hardware event. Clients that rely on discrete scrolling should accumulate the value120 to multiples of 120 before processing the event.

The value120 must not be zero.

This event replaces the wl_pointer.axis_discrete event in clients supporting wl_pointer version 8 or later.

Where a wl_pointer.axis_source event occurs in the same wl_pointer.frame, the axis source applies to this event.

The order of wl_pointer.axis_value120 and wl_pointer.axis_source is not guaranteed.

Version 9

val axis_relative_direction : [< `V9 ] t -> axis:{Imports}1.Wl_pointer.Axis.t -> direction:{Imports}1.Wl_pointer.Axis_relative_direction.t -> unit

Axis relative physical direction event.

Relative directional information of the entity causing the axis motion.

For a wl_pointer.axis event, the wl_pointer.axis_relative_direction event specifies the movement direction of the entity causing the wl_pointer.axis event. For example:

if a user's fingers on a touchpad move down and this causes a wl_pointer.axis vertical_scroll down event, the physical direction is 'identical'

if a user's fingers on a touchpad move down and this causes a wl_pointer.axis vertical_scroll up scroll up event ('natural scrolling'), the physical direction is 'inverted'.

A client may use this information to adjust scroll motion of components. Specifically, enabling natural scrolling causes the content to change direction compared to traditional scrolling. Some widgets like volume control sliders should usually match the physical direction regardless of whether natural scrolling is active. This event enables clients to match the scroll direction of a widget to the physical direction.

This event does not occur on its own, it is coupled with a wl_pointer.axis event that represents this axis value. The protocol guarantees that each axis_relative_direction event is always followed by exactly one axis event with the same axis number within the same wl_pointer.frame. Note that the protocol allows for other events to occur between the axis_relative_direction and its coupled axis event.

The axis number is identical to the axis number in the associated axis event.

The order of wl_pointer.axis_relative_direction, wl_pointer.axis_discrete and wl_pointer.axis_source is not guaranteed.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.

class virtual +'a v4 : object ... end

Handler for a proxy with version >= 4.

class virtual +'a v5 : object ... end

Handler for a proxy with version >= 5.

class virtual +'a v6 : object ... end

Handler for a proxy with version >= 6.

class virtual +'a v7 : object ... end

Handler for a proxy with version >= 7.

class virtual +'a v8 : object ... end

Handler for a proxy with version >= 8.

class virtual +'a v9 : object ... end

Handler for a proxy with version >= 9.