Class Wl_seat.handlers

constraint 'b = [> `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 ]

method user_data : [Wayland_proto.Wl_seat.t, 'b, [ `Server ]] Proxy.Handler.t -> (Wayland_proto.Wl_seat.t, [ `Server ]) S.user_data

method metadata : [Wayland_proto.Wl_seat.t, 'b, [ `Server ]] Proxy.Handler.t -> (module Metadata.S with type t = Wayland_proto.Wl_seat.t)

method min_version : [Wayland_proto.Wl_seat.t, 'b, [ `Server ]] Proxy.Handler.t -> int32

method max_version : [Wayland_proto.Wl_seat.t, 'b, [ `Server ]] Proxy.Handler.t -> int32

method private virtual on_get_pointer : 'b t -> ([ `Wl_pointer ], 'b, [ `Server ]) Proxy.t -> unit

Return pointer object.

The ID provided will be initialized to the wl_pointer interface for this seat.

This request only takes effect if the seat has the pointer capability, or has had the pointer capability in the past. It is a protocol violation to issue this request on a seat that has never had the pointer capability. The missing_capability error will be sent in this case.

method private virtual on_get_keyboard : 'b t -> ([ `Wl_keyboard ], 'b, [ `Server ]) Proxy.t -> unit

Return keyboard object.

The ID provided will be initialized to the wl_keyboard interface for this seat.

This request only takes effect if the seat has the keyboard capability, or has had the keyboard capability in the past. It is a protocol violation to issue this request on a seat that has never had the keyboard capability. The missing_capability error will be sent in this case.

method private virtual on_get_touch : 'b t -> ([ `Wl_touch ], 'b, [ `Server ]) Proxy.t -> unit

Return touch object.

The ID provided will be initialized to the wl_touch interface for this seat.

This request only takes effect if the seat has the touch capability, or has had the touch capability in the past. It is a protocol violation to issue this request on a seat that has never had the touch capability. The missing_capability error will be sent in this case.

method private virtual on_release : 'b t -> unit

Release the seat object.

Using this request a client can tell the server that it is not going to use the seat object anymore.

method dispatch : [Wayland_proto.Wl_seat.t, 'b, [ `Server ]] Proxy.Handler.t -> (Wayland_proto.Wl_seat.t, 'b, [ `Server ]) Proxy.t -> (Wayland_proto.Wl_seat.t, [ `R ]) Msg.t -> unit