Class Wl_seat.v7

Handler for a proxy with version >= 7.

constraint 'a = [< `V7 | `V8 | `V9 ]

method private virtual on_get_pointer : [> `V7 | `V8 | `V9 ] t -> ([ `Wl_pointer ], 'b, [ `Server ]) Wayland__.Proxy.t -> unit

Return pointer object.

The ID provided will be initialized to the wl_pointer interface for this seat.

This request only takes effect if the seat has the pointer capability, or has had the pointer capability in the past. It is a protocol violation to issue this request on a seat that has never had the pointer capability. The missing_capability error will be sent in this case.

method private virtual on_get_keyboard : [> `V7 | `V8 | `V9 ] t -> ([ `Wl_keyboard ], 'b, [ `Server ]) Wayland__.Proxy.t -> unit

Return keyboard object.

The ID provided will be initialized to the wl_keyboard interface for this seat.

This request only takes effect if the seat has the keyboard capability, or has had the keyboard capability in the past. It is a protocol violation to issue this request on a seat that has never had the keyboard capability. The missing_capability error will be sent in this case.

method private virtual on_get_touch : [> `V7 | `V8 | `V9 ] t -> ([ `Wl_touch ], 'b, [ `Server ]) Wayland__.Proxy.t -> unit

Return touch object.

The ID provided will be initialized to the wl_touch interface for this seat.

This request only takes effect if the seat has the touch capability, or has had the touch capability in the past. It is a protocol violation to issue this request on a seat that has never had the touch capability. The missing_capability error will be sent in this case.

method private virtual on_release : [> `V7 | `V8 | `V9 ] t -> unit

Release the seat object.

Using this request a client can tell the server that it is not going to use the seat object anymore.

method min_version : int32

method bind_version : [ `V7 ]