Class Wl_shell.h1

method virtual on_get_shell_surface : 'v t -> int32 -> surface:int32 -> unit

Create a shell surface from a surface.

Create a shell surface for an existing surface. This gives the wl_surface the role of a shell surface. If the wl_surface already has another role, it raises a protocol error.

Only one shell surface can be associated with a given surface.