Class Wl_shell.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_get_shell_surface : [> `V1 ] t -> ([ `Wl_shell_surface ], 'b, [ `Server ]) Wayland__.Proxy.t -> surface: ([ `Wl_surface ], [> {Imports}1.Wl_surface.versions ], [ `Server ]) Wayland__.Proxy.t -> unit

Create a shell surface from a surface.

Create a shell surface for an existing surface. This gives the wl_surface the role of a shell surface. If the wl_surface already has another role, it raises a protocol error.

Only one shell surface can be associated with a given surface.

method min_version : int32

method bind_version : [ `V1 ]