Class Wl_shm.handlers

constraint 'b = [> `V1 ]

method user_data : [Wayland_proto.Wl_shm.t, 'b, [ `Server ]] Proxy.Handler.t -> (Wayland_proto.Wl_shm.t, [ `Server ]) S.user_data

method metadata : [Wayland_proto.Wl_shm.t, 'b, [ `Server ]] Proxy.Handler.t -> (module Metadata.S with type t = Wayland_proto.Wl_shm.t)

method min_version : [Wayland_proto.Wl_shm.t, 'b, [ `Server ]] Proxy.Handler.t -> int32

method max_version : [Wayland_proto.Wl_shm.t, 'b, [ `Server ]] Proxy.Handler.t -> int32

method private virtual on_create_pool : 'b t -> ([ `Wl_shm_pool ], 'b, [ `Server ]) Proxy.t -> fd:Unix.file_descr -> size:int32 -> unit

Create a shm pool.

Create a new wl_shm_pool object.

The pool can be used to create shared memory based buffer objects. The server will mmap size bytes of the passed file descriptor, to use as backing memory for the pool.

method dispatch : [Wayland_proto.Wl_shm.t, 'b, [ `Server ]] Proxy.Handler.t -> (Wayland_proto.Wl_shm.t, 'b, [ `Server ]) Proxy.t -> (Wayland_proto.Wl_shm.t, [ `R ]) Msg.t -> unit