Class Wl_shm.v1

Handler for a proxy with version >= 1.

constraint 'b = [< `V1 ]

method private virtual on_create_pool : [> `V1 ] t -> ([ `Wl_shm_pool ], 'b, [ `Server ]) Wayland__Proxy.t -> fd:Unix.file_descr -> size:int32 -> unit

Create a shm pool.

Create a new wl_shm_pool object.

The pool can be used to create shared memory based buffer objects. The server will mmap size bytes of the passed file descriptor, to use as backing memory for the pool.

method min_version : [Wayland__Wayland_proto.Wl_shm.t, 'b, [ `Server ]] Wayland__.Proxy.Service_handler.t -> int32

method bind_version : 'c -> [ `V1 ]