Class Wl_shm.v2

Handler for a proxy with version >= 2.

constraint 'a = [< `V2 ]

method private virtual on_create_pool : [> `V2 ] t -> ([ `Wl_shm_pool ], 'b, [ `Server ]) Wayland__.Proxy.t -> fd:Unix.file_descr -> size:int32 -> unit

Create a shm pool.

Create a new wl_shm_pool object.

The pool can be used to create shared memory based buffer objects. The server will mmap size bytes of the passed file descriptor, to use as backing memory for the pool.

method private virtual on_release : [> `V2 ] t -> unit

Release the shm object.

Using this request a client can tell the server that it is not going to use the shm object anymore.

Objects created via this interface remain unaffected.

method min_version : int32

method bind_version : [ `V2 ]