Class Wl_shm_pool.h1

method virtual on_resize : 'v t -> size:int32 -> unit

Change the size of the pool mapping.

This request will cause the server to remap the backing memory for the pool from the file descriptor passed when the pool was created, but using the new size. This request can only be used to make the pool bigger.

method virtual on_destroy : 'v t -> unit

Destroy the pool.

Destroy the shared memory pool.

The mmapped memory will be released when all buffers that have been created from this pool are gone.

method virtual on_create_buffer : 'v t -> int32 -> offset:int32 -> width:int32 -> height:int32 -> stride:int32 -> format:int32 -> unit

Create a buffer from the pool.

Create a wl_buffer object from the pool.

The buffer is created offset bytes into the pool and has width and height as specified. The stride argument specifies the number of bytes from the beginning of one row to the beginning of the next. The format is the pixel format of the buffer and must be one of those advertised through the wl_shm.format event.

A buffer will keep a reference to the pool it was created from so it is valid to destroy the pool immediately after creating a buffer from it.