Class Wl_subcompositor.h1

method virtual on_get_subsurface : 'v t -> int32 -> surface:int32 -> parent:int32 -> unit

Give a surface the role sub-surface.

Create a sub-surface interface for the given surface, and associate it with the given parent surface. This turns a plain wl_surface into a sub-surface.

The to-be sub-surface must not already have another role, and it must not have an existing wl_subsurface object. Otherwise a protocol error is raised.

Adding sub-surfaces to a parent is a double-buffered operation on the parent (see wl_surface.commit). The effect of adding a sub-surface becomes visible on the next time the state of the parent surface is applied.

This request modifies the behaviour of wl_surface.commit request on the sub-surface, see the documentation on wl_subsurface interface.

method virtual on_destroy : 'v t -> unit

Unbind from the subcompositor interface.

Informs the server that the client will not be using this protocol object anymore. This does not affect any other objects, wl_subsurface objects included.