Class Wl_subcompositor.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_destroy : [> `V1 ] t -> unit

Unbind from the subcompositor interface.

Informs the server that the client will not be using this protocol object anymore. This does not affect any other objects, wl_subsurface objects included.

method private virtual on_get_subsurface : [> `V1 ] t -> ( [ `Wl_subsurface ], 'b, [ `Server ] ) Wayland__Proxy.t -> surface: ( [ `Wl_surface ], [> Wayland__Wayland_proto.Wl_surface.versions ], [ `Server ] ) Wayland__Proxy.t -> parent: ( [ `Wl_surface ], [> Wayland__Wayland_proto.Wl_surface.versions ], [ `Server ] ) Wayland__Proxy.t -> unit

Give a surface the role sub-surface.

Create a sub-surface interface for the given surface, and associate it with the given parent surface. This turns a plain wl_surface into a sub-surface.

The to-be sub-surface must not already have another role, and it must not have an existing wl_subsurface object. Otherwise the bad_surface protocol error is raised.

Adding sub-surfaces to a parent is a double-buffered operation on the parent (see wl_surface.commit). The effect of adding a sub-surface becomes visible on the next time the state of the parent surface is applied.

The parent surface must not be one of the child surface's descendants, and the parent must be different from the child surface, otherwise the bad_parent protocol error is raised.

This request modifies the behaviour of wl_surface.commit request on the sub-surface, see the documentation on wl_subsurface interface.

method min_version : 'a v1 -> int32

method bind_version : 'c -> [ `V1 ]