Class Wl_subsurface.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_destroy : [> `V1 ] t -> unit

Remove sub-surface interface.

The sub-surface interface is removed from the wl_surface object that was turned into a sub-surface with a wl_subcompositor.get_subsurface request. The wl_surface's association to the parent is deleted. The wl_surface is unmapped immediately.

method private virtual on_set_position : [> `V1 ] t -> x:int32 -> y:int32 -> unit

Reposition the sub-surface.

This schedules a sub-surface position change. The sub-surface will be moved so that its origin (top left corner pixel) will be at the location x, y of the parent surface coordinate system. The coordinates are not restricted to the parent surface area. Negative values are allowed.

The scheduled coordinates will take effect whenever the state of the parent surface is applied.

If more than one set_position request is invoked by the client before the commit of the parent surface, the position of a new request always replaces the scheduled position from any previous request.

The initial position is 0, 0.

method private virtual on_place_above : [> `V1 ] t -> sibling: ([ `Wl_surface ], [> {Imports}1.Wl_surface.versions ], [ `Server ]) Wayland__.Proxy.t -> unit

Restack the sub-surface.

This sub-surface is taken from the stack, and put back just above the reference surface, changing the z-order of the sub-surfaces. The reference surface must be one of the sibling surfaces, or the parent surface. Using any other surface, including this sub-surface, will cause a protocol error.

The z-order is double-buffered. Requests are handled in order and applied immediately to a pending state. The final pending state is copied to the active state the next time the state of the parent surface is applied.

A new sub-surface is initially added as the top-most in the stack of its siblings and parent.

method private virtual on_place_below : [> `V1 ] t -> sibling: ([ `Wl_surface ], [> {Imports}1.Wl_surface.versions ], [ `Server ]) Wayland__.Proxy.t -> unit

Restack the sub-surface.

The sub-surface is placed just below the reference surface. See wl_subsurface.place_above.

method private virtual on_set_sync : [> `V1 ] t -> unit

Set sub-surface to synchronized mode.

Change the commit behaviour of the sub-surface to synchronized mode, also described as the parent dependent mode.

In synchronized mode, wl_surface.commit on a sub-surface will accumulate the committed state in a cache, but the state will not be applied and hence will not change the compositor output. The cached state is applied to the sub-surface immediately after the parent surface's state is applied. This ensures atomic updates of the parent and all its synchronized sub-surfaces. Applying the cached state will invalidate the cache, so further parent surface commits do not (re-)apply old state.

See wl_subsurface for the recursive effect of this mode.

method private virtual on_set_desync : [> `V1 ] t -> unit

Set sub-surface to desynchronized mode.

Change the commit behaviour of the sub-surface to desynchronized mode, also described as independent or freely running mode.

In desynchronized mode, wl_surface.commit on a sub-surface will apply the pending state directly, without caching, as happens normally with a wl_surface. Calling wl_surface.commit on the parent surface has no effect on the sub-surface's wl_surface state. This mode allows a sub-surface to be updated on its own.

If cached state exists when wl_surface.commit is called in desynchronized mode, the pending state is added to the cached state, and applied as a whole. This invalidates the cache.

Note: even if a sub-surface is set to desynchronized, a parent sub-surface may override it to behave as synchronized. For details, see wl_subsurface.

If a surface's parent surface behaves as desynchronized, then the cached state is applied on set_desync.

method min_version : int32