Module Wayland_client.Wl_buffer

Content for a wl_surface.

A buffer provides the content for a wl_surface. Buffers are created through factory interfaces such as wl_drm, wl_shm or similar. It has a width and a height and can be attached to a wl_surface, but the mechanism by which a client provides and updates the contents is defined by the buffer factory interface.

type 'v t = ([ `Wl_buffer ], 'v, [ `Client ]) Wayland.Proxy.t

Version 1

val destroy : [< `V1 ] t -> unit

Destroy a buffer.

Destroy a buffer. If and how you need to release the backing storage is defined by the buffer factory interface.

For possible side-effects to a surface, see wl_surface.attach.

class virtual +'a handlers : object ... end