Class Wl_buffer.handlers

method user_data : [Wayland.Wayland_proto.Wl_buffer.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_buffer.t, [ `Client ]) Wayland.S.user_data

method metadata : [Wayland.Wayland_proto.Wl_buffer.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wayland.Wayland_proto.Wl_buffer.t)

method private virtual on_release : 'v t -> unit

Compositor releases buffer.

Sent when this wl_buffer is no longer used by the compositor. The client is now free to reuse or destroy this buffer and its backing storage.

If a client receives a release event before the frame callback requested in the same wl_surface.commit that attaches this wl_buffer to a surface, then the client is immediately free to reuse the buffer and its backing storage, and does not need a second buffer for the next surface content update. Typically this is possible, when the compositor maintains a copy of the wl_surface contents, e.g. as a GL texture. This is an important optimization for GL(ES) compositors with wl_shm clients.

method dispatch : [Wayland.Wayland_proto.Wl_buffer.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_buffer.t, 'v, [ `Client ]) Wayland.Proxy.t -> (Wayland.Wayland_proto.Wl_buffer.t, [ `R ]) Wayland.Msg.t -> unit