Class Wl_data_device.handlers

method user_data : [Wayland.Wayland_proto.Wl_data_device.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_data_device.t, [ `Client ]) Wayland.S.user_data

method metadata : [Wayland.Wayland_proto.Wl_data_device.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wayland.Wayland_proto.Wl_data_device.t)

method private virtual on_data_offer : 'v t -> ([ `Wl_data_offer ], 'v, [ `Client ]) Wayland.Proxy.t -> unit

Introduce a new wl_data_offer.

The data_offer event introduces a new wl_data_offer object, which will subsequently be used in either the data_device.enter event (for drag-and-drop) or the data_device.selection event (for selections). Immediately following the data_device_data_offer event, the new data_offer object will send out data_offer.offer events to describe the mime types it offers.

method private virtual on_enter : 'v t -> serial:int32 -> surface:([ `Wl_surface ], 'v, [ `Client ]) Wayland.Proxy.t -> x:Wayland.Fixed.t -> y:Wayland.Fixed.t -> ([ `Wl_data_offer ], 'v, [ `Client ]) Wayland.Proxy.t option -> unit

Initiate drag-and-drop session.

This event is sent when an active drag-and-drop pointer enters a surface owned by the client. The position of the pointer at enter time is provided by the x and y arguments, in surface-local coordinates.

method private virtual on_leave : 'v t -> unit

End drag-and-drop session.

This event is sent when the drag-and-drop pointer leaves the surface and the session ends. The client must destroy the wl_data_offer introduced at enter time at this point.

method private virtual on_motion : 'v t -> time:int32 -> x:Wayland.Fixed.t -> y:Wayland.Fixed.t -> unit

Drag-and-drop session motion.

This event is sent when the drag-and-drop pointer moves within the currently focused surface. The new position of the pointer is provided by the x and y arguments, in surface-local coordinates.

method private virtual on_drop : 'v t -> unit

End drag-and-drop session successfully.

The event is sent when a drag-and-drop operation is ended because the implicit grab is removed.

The drag-and-drop destination is expected to honor the last action received through wl_data_offer.action, if the resulting action is "copy" or "move", the destination can still perform wl_data_offer.receive requests, and is expected to end all transfers with a wl_data_offer.finish request.

If the resulting action is "ask", the action will not be considered final. The drag-and-drop destination is expected to perform one last wl_data_offer.set_actions request, or wl_data_offer.destroy in order to cancel the operation.

method private virtual on_selection : 'v t -> ([ `Wl_data_offer ], 'v, [ `Client ]) Wayland.Proxy.t option -> unit

Advertise new selection.

The selection event is sent out to notify the client of a new wl_data_offer for the selection for this device. The data_device.data_offer and the data_offer.offer events are sent out immediately before this event to introduce the data offer object. The selection event is sent to a client immediately before receiving keyboard focus and when a new selection is set while the client has keyboard focus. The data_offer is valid until a new data_offer or NULL is received or until the client loses keyboard focus. The client must destroy the previous selection data_offer, if any, upon receiving this event.

method dispatch : [Wayland.Wayland_proto.Wl_data_device.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_data_device.t, 'v, [ `Client ]) Wayland.Proxy.t -> (Wayland.Wayland_proto.Wl_data_device.t, [ `R ]) Wayland.Msg.t -> unit