Class Wl_data_offer.handlers

method user_data : [Wayland.Wayland_proto.Wl_data_offer.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_data_offer.t, [ `Client ]) Wayland.S.user_data

method metadata : [Wayland.Wayland_proto.Wl_data_offer.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wayland.Wayland_proto.Wl_data_offer.t)

method private virtual on_offer : 'v t -> mime_type:string -> unit

Advertise offered mime type.

Sent immediately after creating the wl_data_offer object. One event per offered mime type.

method private virtual on_source_actions : 'v t -> source_actions:$Imports.Wl_data_device_manager.Dnd_action.t -> unit

Notify the source-side available actions.

This event indicates the actions offered by the data source. It will be sent right after wl_data_device.enter, or anytime the source side changes its offered actions through wl_data_source.set_actions.

method private virtual on_action : 'v t -> dnd_action:$Imports.Wl_data_device_manager.Dnd_action.t -> unit

Notify the selected action.

This event indicates the action selected by the compositor after matching the source/destination side actions. Only one action (or none) will be offered here.

This event can be emitted multiple times during the drag-and-drop operation in response to destination side action changes through wl_data_offer.set_actions.

This event will no longer be emitted after wl_data_device.drop happened on the drag-and-drop destination, the client must honor the last action received, or the last preferred one set through wl_data_offer.set_actions when handling an "ask" action.

Compositors may also change the selected action on the fly, mainly in response to keyboard modifier changes during the drag-and-drop operation.

The most recent action received is always the valid one. Prior to receiving wl_data_device.drop, the chosen action may change (e.g. due to keyboard modifiers being pressed). At the time of receiving wl_data_device.drop the drag-and-drop destination must honor the last action received.

Action changes may still happen after wl_data_device.drop, especially on "ask" actions, where the drag-and-drop destination may choose another action afterwards. Action changes happening at this stage are always the result of inter-client negotiation, the compositor shall no longer be able to induce a different action.

Upon "ask" actions, it is expected that the drag-and-drop destination may potentially choose a different action and/or mime type, based on wl_data_offer.source_actions and finally chosen by the user (e.g. popping up a menu with the available options). The final wl_data_offer.set_actions and wl_data_offer.accept requests must happen before the call to wl_data_offer.finish.

method dispatch : [Wayland.Wayland_proto.Wl_data_offer.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_data_offer.t, 'v, [ `Client ]) Wayland.Proxy.t -> (Wayland.Wayland_proto.Wl_data_offer.t, [ `R ]) Wayland.Msg.t -> unit