Module Wayland_client.Wl_display

Core global object.

The core global object. This is a special singleton object. It is used for internal Wayland protocol features.

type 'v t = ([ `Wl_display ], 'v) Wayland.Proxy.t

Version 1

val get_registry : 'v t -> ([ `Wl_registry ], 'v, 'v, 'a) Wayland.Proxy.handler -> ([ `Wl_registry ], 'v) Wayland.Proxy.t

Get global registry object.

This request creates a registry object that allows the client to list and bind the global objects available from the compositor.

It should be noted that the server side resources consumed in response to a get_registry request can only be released when the client disconnects, not when the client side proxy is destroyed. Therefore, clients should invoke get_registry as infrequently as possible to avoid wasting memory.

val sync : 'v t -> ([ `Wl_callback ], 'v, 'v, 'a) Wayland.Proxy.handler -> ([ `Wl_callback ], 'v) Wayland.Proxy.t

Asynchronous roundtrip.

The sync request asks the server to emit the 'done' event on the returned wl_callback object. Since requests are handled in-order and events are delivered in-order, this can be used as a barrier to ensure all previous requests and the resulting events have been handled.

The object returned by this request will be destroyed by the compositor after the callback is fired and as such the client must not attempt to use it after that point.

The callback_data passed in the callback is the event serial.

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_display ], 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler