Class Wl_display.h1

method virtual on_delete_id : 'v t -> id:int32 -> unit

Acknowledge object ID deletion.

This event is used internally by the object ID management logic. When a client deletes an object that it had created, the server will send this event to acknowledge that it has seen the delete request. When the client receives this event, it will know that it can safely reuse the object ID.

method virtual on_error : 'v t -> object_id:int32 -> code:int32 -> message:string -> unit

Fatal error event.

The error event is sent out when a fatal (non-recoverable) error has occurred. The object_id argument is the object where the error occurred, most often in response to a request to that object. The code identifies the error and is defined by the object interface. As such, each interface defines its own set of error codes. The message is a brief description of the error, for (debugging) convenience.