Class Wl_seat.h2

inherit h1

method virtual on_name : 'v t -> name:string -> unit

Unique identifier for this seat.

In a multiseat configuration this can be used by the client to help identify which physical devices the seat represents. Based on the seat configuration used by the compositor.