Module Wayland_client.Wl_shell

Create desktop-style surfaces.

This interface is implemented by servers that provide desktop-style user interfaces.

It allows clients to associate a wl_shell_surface with a basic surface.

Note! This protocol is deprecated and not intended for production use. For desktop-style user interfaces, use xdg_shell.

type 'v t = ([ `Wl_shell ], 'v) Wayland.Proxy.t

Version 1

val get_shell_surface : 'v t -> ([ `Wl_shell_surface ], 'v, 'v, 'a) Wayland.Proxy.handler -> surface:([ `Wl_surface ], 'b) Wayland.Proxy.t -> ([ `Wl_shell_surface ], 'v) Wayland.Proxy.t

Create a shell surface from a surface.

Create a shell surface for an existing surface. This gives the wl_surface the role of a shell surface. If the wl_surface already has another role, it raises a protocol error.

Only one shell surface can be associated with a given surface.

class virtual 'v h1 : object ... end

val v1 : unit -> ('a, 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler