Class Wl_shell_surface.handlers

method user_data : [Wayland.Wayland_proto.Wl_shell_surface.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_shell_surface.t, [ `Client ]) Wayland.S.user_data

method metadata : [Wayland.Wayland_proto.Wl_shell_surface.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wayland.Wayland_proto.Wl_shell_surface.t)

method private virtual on_ping : 'v t -> serial:int32 -> unit

Ping client.

Ping a client to check if it is receiving events and sending requests. A client is expected to reply with a pong request.

method private virtual on_configure : 'v t -> edges:$Imports.Wl_shell_surface.Resize.t -> width:int32 -> height:int32 -> unit

Suggest resize.

The configure event asks the client to resize its surface.

The size is a hint, in the sense that the client is free to ignore it if it doesn't resize, pick a smaller size (to satisfy aspect ratio or resize in steps of NxM pixels).

The edges parameter provides a hint about how the surface was resized. The client may use this information to decide how to adjust its content to the new size (e.g. a scrolling area might adjust its content position to leave the viewable content unmoved).

The client is free to dismiss all but the last configure event it received.

The width and height arguments specify the size of the window in surface-local coordinates.

method private virtual on_popup_done : 'v t -> unit

Popup interaction is done.

The popup_done event is sent out when a popup grab is broken, that is, when the user clicks a surface that doesn't belong to the client owning the popup surface.

method dispatch : [Wayland.Wayland_proto.Wl_shell_surface.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_shell_surface.t, 'v, [ `Client ]) Wayland.Proxy.t -> (Wayland.Wayland_proto.Wl_shell_surface.t, [ `R ]) Wayland.Msg.t -> unit