Module Wayland_client.Wl_shm

Shared memory support.

A singleton global object that provides support for shared memory.

Clients can create wl_shm_pool objects using the create_pool request.

At connection setup time, the wl_shm object emits one or more format events to inform clients about the valid pixel formats that can be used for buffers.

type 'v t = ([ `Wl_shm ], 'v, [ `Client ]) Wayland.Proxy.t

module Error = Wayland.Wayland_proto.Wl_shm.Error

module Format = Wayland.Wayland_proto.Wl_shm.Format

Version 1

val create_pool : [< `V1 ] as 'a t -> [[ `Wl_shm_pool ], 'a, [ `Client ]] Wayland.Proxy.Handler.t -> fd:Unix.file_descr -> size:int32 -> ([ `Wl_shm_pool ], 'a, [ `Client ]) Wayland.Proxy.t

Create a shm pool.

Create a new wl_shm_pool object.

The pool can be used to create shared memory based buffer objects. The server will mmap size bytes of the passed file descriptor, to use as backing memory for the pool.

class virtual +'a _handlers_unsafe : object ... end

Requesting a minimum version

class virtual v1 : object ... end