Module Wayland_client.Wl_touch

Touchscreen input device.

The wl_touch interface represents a touchscreen associated with a seat.

Touch interactions can consist of one or more contacts. For each contact, a series of events is generated, starting with a down event, followed by zero or more motion events, and ending with an up event. Events relating to the same contact point can be identified by the ID of the sequence.

type 'v t = ([ `Wl_touch ], 'v, [ `Client ]) Wayland.Proxy.t

Version 1, 2

Version 3, 4, 5

val release : [< `V3 | `V4 | `V5 | `V6 | `V7 ] t -> unit

Release the touch object.

Version 6, 7

class virtual +'a handlers : object ... end