Module Wayland_server.Wl_data_source

Offer to transfer data.

The wl_data_source object is the source side of a wl_data_offer. It is created by the source client in a data transfer and provides a way to describe the offered data and a way to respond to requests to transfer the data.

type 'v t = ([ `Wl_data_source ], 'v) Wayland.Proxy.t

Version 1, 2

val cancelled : 'v t -> unit

Selection was cancelled.

This data source is no longer valid. There are several reasons why this could happen:

The data source has been replaced by another data source.

The drag-and-drop operation was performed, but the drop destination did not accept any of the mime types offered through wl_data_source.target.

The drag-and-drop operation was performed, but the drop destination did not select any of the actions present in the mask offered through wl_data_source.action.

The drag-and-drop operation was performed but didn't happen over a surface.

The compositor cancelled the drag-and-drop operation (e.g. compositor dependent timeouts to avoid stale drag-and-drop transfers).

The client should clean up and destroy this data source.

For objects of version 2 or older, wl_data_source.cancelled will only be emitted if the data source was replaced by another data source.

val send : 'v t -> mime_type:string -> fd:Unix.file_descr -> unit

Send the data.

Request for data from the client. Send the data as the specified mime type over the passed file descriptor, then close it.

val target : 'v t -> mime_type:string -> unit

A target accepts an offered mime type.

Sent when a target accepts pointer_focus or motion events. If a target does not accept any of the offered types, type is NULL.

Used for feedback during drag-and-drop.

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_data_source ], 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler

val v2 : 'v h1 -> ([ `Wl_data_source ], 'v, [< `V1 | `V2 ], [> `V2 ]) Wayland.Proxy.handler

Version 3

val action : 'v t -> dnd_action:int32 -> unit

Notify the selected action.

This event indicates the action selected by the compositor after matching the source/destination side actions. Only one action (or none) will be offered here.

This event can be emitted multiple times during the drag-and-drop operation, mainly in response to destination side changes through wl_data_offer.set_actions, and as the data device enters/leaves surfaces.

It is only possible to receive this event after wl_data_source.dnd_drop_performed if the drag-and-drop operation ended in an "ask" action, in which case the final wl_data_source.action event will happen immediately before wl_data_source.dnd_finished.

Compositors may also change the selected action on the fly, mainly in response to keyboard modifier changes during the drag-and-drop operation.

The most recent action received is always the valid one. The chosen action may change alongside negotiation (e.g. an "ask" action can turn into a "move" operation), so the effects of the final action must always be applied in wl_data_offer.dnd_finished.

Clients can trigger cursor surface changes from this point, so they reflect the current action.

val dnd_finished : 'v t -> unit

The drag-and-drop operation concluded.

The drop destination finished interoperating with this data source, so the client is now free to destroy this data source and free all associated data.

If the action used to perform the operation was "move", the source can now delete the transferred data.

val dnd_drop_performed : 'v t -> unit

The drag-and-drop operation physically finished.

The user performed the drop action. This event does not indicate acceptance, wl_data_source.cancelled may still be emitted afterwards if the drop destination does not accept any mime type.

However, this event might however not be received if the compositor cancelled the drag-and-drop operation before this event could happen.

Note that the data_source may still be used in the future and should not be destroyed here.

class virtual +'a h3 : object ... end

val v3 : 'v h3 -> ([ `Wl_data_source ], 'v, [< `V1 | `V2 | `V3 ], [> `V3 ]) Wayland.Proxy.handler