Module Wayland_server.Wl_display

Core global object.

The core global object. This is a special singleton object. It is used for internal Wayland protocol features.

type 'v t = ([ `Wl_display ], 'v, [ `Server ]) Wayland.Proxy.t

module Error = Wayland.Wayland_proto.Wl_display.Error

Version 1

val delete_id : [< `V1 ] t -> id:int32 -> unit

Acknowledge object ID deletion.

This event is used internally by the object ID management logic. When a client deletes an object that it had created, the server will send this event to acknowledge that it has seen the delete request. When the client receives this event, it will know that it can safely reuse the object ID.

val error : [< `V1 ] t -> object_id:('a, 'b, 'c) Wayland.Proxy.t -> code:int32 -> message:string -> unit

Fatal error event.

The error event is sent out when a fatal (non-recoverable) error has occurred. The object_id argument is the object where the error occurred, most often in response to a request to that object. The code identifies the error and is defined by the object interface. As such, each interface defines its own set of error codes. The message is a brief description of the error, for (debugging) convenience.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.