Module Wayland_server.Wl_shell

Create desktop-style surfaces.

This interface is implemented by servers that provide desktop-style user interfaces.

It allows clients to associate a wl_shell_surface with a basic surface.

Note! This protocol is deprecated and not intended for production use. For desktop-style user interfaces, use xdg_shell.

type 'v t = ([ `Wl_shell ], 'v, [ `Server ]) Wayland.Proxy.t

module Error = Wayland.Wayland_proto.Wl_shell.Error

Version 1

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.