Module Wayland_server.Wl_shm

Shared memory support.

A singleton global object that provides support for shared memory.

Clients can create wl_shm_pool objects using the create_pool request.

At connection setup time, the wl_shm object emits one or more format events to inform clients about the valid pixel formats that can be used for buffers.

type 'v t = ([ `Wl_shm ], 'v) Wayland.Proxy.t

Version 1

val format : 'v t -> format:int32 -> unit

Pixel format description.

Informs the client about a valid pixel format that can be used for buffers. Known formats include argb8888 and xrgb8888.

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_shm ], 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler