Class Wl_shm.h1

method virtual on_create_pool : 'v t -> int32 -> fd:Unix.file_descr -> size:int32 -> unit

Create a shm pool.

Create a new wl_shm_pool object.

The pool can be used to create shared memory based buffer objects. The server will mmap size bytes of the passed file descriptor, to use as backing memory for the pool.