Module Wayland_server.Wl_subcompositor

Sub-surface compositing.

The global interface exposing sub-surface compositing capabilities. A wl_surface, that has sub-surfaces associated, is called the parent surface. Sub-surfaces can be arbitrarily nested and create a tree of sub-surfaces.

The root surface in a tree of sub-surfaces is the main surface. The main surface cannot be a sub-surface, because sub-surfaces must always have a parent.

A main surface with its sub-surfaces forms a (compound) window. For window management purposes, this set of wl_surface objects is to be considered as a single window, and it should also behave as such.

The aim of sub-surfaces is to offload some of the compositing work within a window from clients to the compositor. A prime example is a video player with decorations and video in separate wl_surface objects. This should allow the compositor to pass YUV video buffer processing to dedicated overlay hardware when possible.

type 'v t = ([ `Wl_subcompositor ], 'v) Wayland.Proxy.t

Version 1

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_subcompositor ], 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler