Class Wl_subcompositor.handlers

constraint 'b = [> `V1 ]

method user_data : [Wayland.Wayland_proto.Wl_subcompositor.t, 'b, [ `Server ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_subcompositor.t, [ `Server ]) Wayland.S.user_data

method metadata : [Wayland.Wayland_proto.Wl_subcompositor.t, 'b, [ `Server ]] Wayland.Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wayland.Wayland_proto.Wl_subcompositor.t)

method min_version : [Wayland.Wayland_proto.Wl_subcompositor.t, 'b, [ `Server ]] Wayland.Proxy.Handler.t -> int32

method max_version : [Wayland.Wayland_proto.Wl_subcompositor.t, 'b, [ `Server ]] Wayland.Proxy.Handler.t -> int32

method private virtual on_destroy : 'b t -> unit

Unbind from the subcompositor interface.

Informs the server that the client will not be using this protocol object anymore. This does not affect any other objects, wl_subsurface objects included.

method private virtual on_get_subsurface : 'b t -> ([ `Wl_subsurface ], 'b, [ `Server ]) Wayland.Proxy.t -> surface:([ `Wl_surface ], 'b, [ `Server ]) Wayland.Proxy.t -> parent:([ `Wl_surface ], 'b, [ `Server ]) Wayland.Proxy.t -> unit

Give a surface the role sub-surface.

Create a sub-surface interface for the given surface, and associate it with the given parent surface. This turns a plain wl_surface into a sub-surface.

The to-be sub-surface must not already have another role, and it must not have an existing wl_subsurface object. Otherwise a protocol error is raised.

Adding sub-surfaces to a parent is a double-buffered operation on the parent (see wl_surface.commit). The effect of adding a sub-surface becomes visible on the next time the state of the parent surface is applied.

This request modifies the behaviour of wl_surface.commit request on the sub-surface, see the documentation on wl_subsurface interface.

method dispatch : [Wayland.Wayland_proto.Wl_subcompositor.t, 'b, [ `Server ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_subcompositor.t, 'b, [ `Server ]) Wayland.Proxy.t -> (Wayland.Wayland_proto.Wl_subcompositor.t, [ `R ]) Wayland.Msg.t -> unit