Module Wayland__Iface_reg

val register : (module Wayland.Metadata.S) -> unit

Called by the generated code to register each Wayland interface.

val lookup : string -> (module Wayland.Metadata.S)

Called by the generated code to convert a string interface to a variant.