Class type Handler.t

An ('a, 'v) t handles incoming messages for an object of type 'a. Typically, a constructor will let the user pick from a range of versions for 'v, which will then be constrained by the spawn call.

method user_data : ('a, 'role) Wayland.S.user_data

method metadata : (module Wayland.Metadata.S with type t = 'a)

method dispatch : ('a, 'v, 'role) proxy -> ('a, [ `R ]) Wayland.Msg.t -> unit