Class Wl_data_source.h1

method virtual on_cancelled : 'v t -> unit

Selection was cancelled.

This data source is no longer valid. There are several reasons why this could happen:

The data source has been replaced by another data source.

The drag-and-drop operation was performed, but the drop destination did not accept any of the mime types offered through wl_data_source.target.

The drag-and-drop operation was performed, but the drop destination did not select any of the actions present in the mask offered through wl_data_source.action.

The drag-and-drop operation was performed but didn't happen over a surface.

The compositor cancelled the drag-and-drop operation (e.g. compositor dependent timeouts to avoid stale drag-and-drop transfers).

The client should clean up and destroy this data source.

For objects of version 2 or older, wl_data_source.cancelled will only be emitted if the data source was replaced by another data source.

method virtual on_send : 'v t -> mime_type:string -> fd:Unix.file_descr -> unit

Send the data.

Request for data from the client. Send the data as the specified mime type over the passed file descriptor, then close it.

method virtual on_target : 'v t -> mime_type:string -> unit

A target accepts an offered mime type.

Sent when a target accepts pointer_focus or motion events. If a target does not accept any of the offered types, type is NULL.

Used for feedback during drag-and-drop.