Class Wl_keyboard.h1

method virtual on_modifiers : 'v t -> serial:int32 -> mods_depressed:int32 -> mods_latched:int32 -> mods_locked:int32 -> group:int32 -> unit

Modifier and group state.

Notifies clients that the modifier and/or group state has changed, and it should update its local state.

method virtual on_key : 'v t -> serial:int32 -> time:int32 -> key:int32 -> state:int32 -> unit

Key event.

A key was pressed or released. The time argument is a timestamp with millisecond granularity, with an undefined base.

The key is a platform-specific key code that can be interpreted by feeding it to the keyboard mapping (see the keymap event).

If this event produces a change in modifiers, then the resulting wl_keyboard.modifiers event must be sent after this event.

method virtual on_leave : 'v t -> serial:int32 -> surface:int32 -> unit

Leave event.

Notification that this seat's keyboard focus is no longer on a certain surface.

The leave notification is sent before the enter notification for the new focus.

After this event client must assume that all keys, including modifiers, are lifted and also it must stop key repeating if there's some going on.

method virtual on_enter : 'v t -> serial:int32 -> surface:int32 -> keys:string -> unit

Enter event.

Notification that this seat's keyboard focus is on a certain surface.

The compositor must send the wl_keyboard.modifiers event after this event.

method virtual on_keymap : 'v t -> format:int32 -> fd:Unix.file_descr -> size:int32 -> unit

Keyboard mapping.

This event provides a file descriptor to the client which can be memory-mapped to provide a keyboard mapping description.

From version 7 onwards, the fd must be mapped with MAP_PRIVATE by the recipient, as MAP_SHARED may fail.