Class Wl_keyboard.handlers

method user_data : [Wayland.Wayland_proto.Wl_keyboard.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_keyboard.t, [ `Client ]) Wayland.S.user_data

method metadata : [Wayland.Wayland_proto.Wl_keyboard.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wayland.Wayland_proto.Wl_keyboard.t)

method private virtual on_keymap : 'v t -> format:$Imports.Wl_keyboard.Keymap_format.t -> fd:Unix.file_descr -> size:int32 -> unit

Keyboard mapping.

This event provides a file descriptor to the client which can be memory-mapped to provide a keyboard mapping description.

From version 7 onwards, the fd must be mapped with MAP_PRIVATE by the recipient, as MAP_SHARED may fail.

method private virtual on_enter : 'v t -> serial:int32 -> surface:([ `Wl_surface ], 'v, [ `Client ]) Wayland.Proxy.t -> keys:string -> unit

Enter event.

Notification that this seat's keyboard focus is on a certain surface.

The compositor must send the wl_keyboard.modifiers event after this event.

method private virtual on_leave : 'v t -> serial:int32 -> surface:([ `Wl_surface ], 'v, [ `Client ]) Wayland.Proxy.t -> unit

Leave event.

Notification that this seat's keyboard focus is no longer on a certain surface.

The leave notification is sent before the enter notification for the new focus.

After this event client must assume that all keys, including modifiers, are lifted and also it must stop key repeating if there's some going on.

method private virtual on_key : 'v t -> serial:int32 -> time:int32 -> key:int32 -> state:$Imports.Wl_keyboard.Key_state.t -> unit

Key event.

A key was pressed or released. The time argument is a timestamp with millisecond granularity, with an undefined base.

The key is a platform-specific key code that can be interpreted by feeding it to the keyboard mapping (see the keymap event).

If this event produces a change in modifiers, then the resulting wl_keyboard.modifiers event must be sent after this event.

method private virtual on_modifiers : 'v t -> serial:int32 -> mods_depressed:int32 -> mods_latched:int32 -> mods_locked:int32 -> group:int32 -> unit

Modifier and group state.

Notifies clients that the modifier and/or group state has changed, and it should update its local state.

method private virtual on_repeat_info : 'v t -> rate:int32 -> delay:int32 -> unit

Repeat rate and delay.

Informs the client about the keyboard's repeat rate and delay.

This event is sent as soon as the wl_keyboard object has been created, and is guaranteed to be received by the client before any key press event.

Negative values for either rate or delay are illegal. A rate of zero will disable any repeating (regardless of the value of delay).

This event can be sent later on as well with a new value if necessary, so clients should continue listening for the event past the creation of wl_keyboard.

method dispatch : [Wayland.Wayland_proto.Wl_keyboard.t, 'v, [ `Client ]] Wayland.Proxy.Handler.t -> (Wayland.Wayland_proto.Wl_keyboard.t, 'v, [ `Client ]) Wayland.Proxy.t -> (Wayland.Wayland_proto.Wl_keyboard.t, [ `R ]) Wayland.Msg.t -> unit