Module Wayland__Wayland_client.Wl_output

Compositor output region.

An output describes part of the compositor geometry. The compositor works in the 'compositor coordinate system' and an output corresponds to a rectangular area in that space that is actually visible. This typically corresponds to a monitor that displays part of the compositor space. This object is published as global during start up, or when a monitor is hotplugged.

type 'v t = ([ `Wl_output ], 'v) Wayland.Proxy.t

Version 1

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_output ], 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler

Version 2

class virtual +'a h2 : object ... end

val v2 : 'v h2 -> ([ `Wl_output ], 'v, [< `V1 | `V2 ], [> `V2 ]) Wayland.Proxy.handler

Version 3

val release : 'v t -> unit

Release the output object.

Using this request a client can tell the server that it is not going to use the output object anymore.

class virtual +'a h3 : object ... end

val v3 : 'v h3 -> ([ `Wl_output ], 'v, [< `V1 | `V2 | `V3 ], [> `V3 ]) Wayland.Proxy.handler