Module Wayland__Wayland_client.Wl_pointer

Pointer input device.

The wl_pointer interface represents one or more input devices, such as mice, which control the pointer location and pointer_focus of a seat.

The wl_pointer interface generates motion, enter and leave events for the surfaces that the pointer is located over, and button and axis events for button presses, button releases and scrolling.

type 'v t = ([ `Wl_pointer ], 'v) Wayland.Proxy.t

Version 1, 2

val set_cursor : 'v t -> serial:int32 -> surface:([ `Wl_surface ], 'a) Wayland.Proxy.t -> hotspot_x:int32 -> hotspot_y:int32 -> unit

Set the pointer surface.

Set the pointer surface, i.e., the surface that contains the pointer image (cursor). This request gives the surface the role of a cursor. If the surface already has another role, it raises a protocol error.

The cursor actually changes only if the pointer focus for this device is one of the requesting client's surfaces or the surface parameter is the current pointer surface. If there was a previous surface set with this request it is replaced. If surface is NULL, the pointer image is hidden.

The parameters hotspot_x and hotspot_y define the position of the pointer surface relative to the pointer location. Its top-left corner is always at (x, y) - (hotspot_x, hotspot_y), where (x, y) are the coordinates of the pointer location, in surface-local coordinates.

On surface.attach requests to the pointer surface, hotspot_x and hotspot_y are decremented by the x and y parameters passed to the request. Attach must be confirmed by wl_surface.commit as usual.

The hotspot can also be updated by passing the currently set pointer surface to this request with new values for hotspot_x and hotspot_y.

The current and pending input regions of the wl_surface are cleared, and wl_surface.set_input_region is ignored until the wl_surface is no longer used as the cursor. When the use as a cursor ends, the current and pending input regions become undefined, and the wl_surface is unmapped.

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_pointer ], 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler

val v2 : 'v h1 -> ([ `Wl_pointer ], 'v, [< `V1 | `V2 ], [> `V2 ]) Wayland.Proxy.handler

Version 3, 4

val release : 'v t -> unit

Release the pointer object.

Using this request a client can tell the server that it is not going to use the pointer object anymore.

This request destroys the pointer proxy object, so clients must not call wl_pointer_destroy() after using this request.

class virtual +'a h3 : object ... end

val v3 : 'v h3 -> ([ `Wl_pointer ], 'v, [< `V1 | `V2 | `V3 ], [> `V3 ]) Wayland.Proxy.handler

val v4 : 'v h3 -> ([ `Wl_pointer ], 'v, [< `V1 | `V2 | `V3 | `V4 ], [> `V4 ]) Wayland.Proxy.handler

Version 5, 6, 7

class virtual +'a h5 : object ... end

val v5 : 'v h5 -> ([ `Wl_pointer ], 'v, [< `V1 | `V2 | `V3 | `V4 | `V5 ], [> `V5 ]) Wayland.Proxy.handler

val v6 : 'v h5 -> ([ `Wl_pointer ], 'v, [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 ], [> `V6 ]) Wayland.Proxy.handler

val v7 : 'v h5 -> ([ `Wl_pointer ], 'v, [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 ], [> `V7 ]) Wayland.Proxy.handler