Module Wayland__Wayland_client.Wl_seat

Group of input devices.

A seat is a group of keyboards, pointer and touch devices. This object is published as a global during start up, or when such a device is hot plugged. A seat typically has a pointer and maintains a keyboard focus and a pointer focus.

type 'v t = ([ `Wl_seat ], 'v) Wayland.Proxy.t

Version 1

val get_touch : 'v t -> ([ `Wl_touch ], 'v, 'v, 'a) Wayland.Proxy.handler -> ([ `Wl_touch ], 'v) Wayland.Proxy.t

Return touch object.

The ID provided will be initialized to the wl_touch interface for this seat.

This request only takes effect if the seat has the touch capability, or has had the touch capability in the past. It is a protocol violation to issue this request on a seat that has never had the touch capability. The missing_capability error will be sent in this case.

val get_keyboard : 'v t -> ([ `Wl_keyboard ], 'v, 'v, 'a) Wayland.Proxy.handler -> ([ `Wl_keyboard ], 'v) Wayland.Proxy.t

Return keyboard object.

The ID provided will be initialized to the wl_keyboard interface for this seat.

This request only takes effect if the seat has the keyboard capability, or has had the keyboard capability in the past. It is a protocol violation to issue this request on a seat that has never had the keyboard capability. The missing_capability error will be sent in this case.

val get_pointer : 'v t -> ([ `Wl_pointer ], 'v, 'v, 'a) Wayland.Proxy.handler -> ([ `Wl_pointer ], 'v) Wayland.Proxy.t

Return pointer object.

The ID provided will be initialized to the wl_pointer interface for this seat.

This request only takes effect if the seat has the pointer capability, or has had the pointer capability in the past. It is a protocol violation to issue this request on a seat that has never had the pointer capability. The missing_capability error will be sent in this case.

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_seat ], 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler

Version 2, 3, 4

class virtual +'a h2 : object ... end

val v2 : 'v h2 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 ], [> `V2 ]) Wayland.Proxy.handler

val v3 : 'v h2 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 ], [> `V3 ]) Wayland.Proxy.handler

val v4 : 'v h2 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 | `V4 ], [> `V4 ]) Wayland.Proxy.handler

Version 5, 6, 7

val release : 'v t -> unit

Release the seat object.

Using this request a client can tell the server that it is not going to use the seat object anymore.

class virtual +'a h5 : object ... end

val v5 : 'v h5 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 | `V4 | `V5 ], [> `V5 ]) Wayland.Proxy.handler

val v6 : 'v h5 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 ], [> `V6 ]) Wayland.Proxy.handler

val v7 : 'v h5 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 ], [> `V7 ]) Wayland.Proxy.handler