Class Wl_buffer.h1

method virtual on_destroy : 'v t -> unit

Destroy a buffer.

Destroy a buffer. If and how you need to release the backing storage is defined by the buffer factory interface.

For possible side-effects to a surface, see wl_surface.attach.