Module Wayland__Wayland_server.Wl_callback

Callback object.

Clients can handle the 'done' event to get notified when the related request is done.

type 'v t = ([ `Wl_callback ], 'v) Wayland.Proxy.t

Version 1

val done_ : 'v t -> callback_data:int32 -> unit

Done event.

Notify the client when the related request is done.

class virtual 'v h1 : object ... end

val v1 : unit -> ('a, 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler