Class Wl_data_source.h1

method virtual on_destroy : 'v t -> unit

Destroy the data source.

Destroy the data source.

method virtual on_offer : 'v t -> mime_type:string -> unit

Add an offered mime type.

This request adds a mime type to the set of mime types advertised to targets. Can be called several times to offer multiple types.