Module Wayland__Wayland_server.Wl_seat

Group of input devices.

A seat is a group of keyboards, pointer and touch devices. This object is published as a global during start up, or when such a device is hot plugged. A seat typically has a pointer and maintains a keyboard focus and a pointer focus.

type 'v t = ([ `Wl_seat ], 'v) Wayland.Proxy.t

Version 1

val capabilities : 'v t -> capabilities:int32 -> unit

Seat capabilities changed.

This is emitted whenever a seat gains or loses the pointer, keyboard or touch capabilities. The argument is a capability enum containing the complete set of capabilities this seat has.

When the pointer capability is added, a client may create a wl_pointer object using the wl_seat.get_pointer request. This object will receive pointer events until the capability is removed in the future.

When the pointer capability is removed, a client should destroy the wl_pointer objects associated with the seat where the capability was removed, using the wl_pointer.release request. No further pointer events will be received on these objects.

In some compositors, if a seat regains the pointer capability and a client has a previously obtained wl_pointer object of version 4 or less, that object may start sending pointer events again. This behavior is considered a misinterpretation of the intended behavior and must not be relied upon by the client. wl_pointer objects of version 5 or later must not send events if created before the most recent event notifying the client of an added pointer capability.

The above behavior also applies to wl_keyboard and wl_touch with the keyboard and touch capabilities, respectively.

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_seat ], 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler

Version 2, 3, 4

val name : 'v t -> name:string -> unit

Unique identifier for this seat.

In a multiseat configuration this can be used by the client to help identify which physical devices the seat represents. Based on the seat configuration used by the compositor.

class virtual +'a h2 : object ... end

val v2 : 'v h2 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 ], [> `V2 ]) Wayland.Proxy.handler

val v3 : 'v h2 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 ], [> `V3 ]) Wayland.Proxy.handler

val v4 : 'v h2 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 | `V4 ], [> `V4 ]) Wayland.Proxy.handler

Version 5, 6, 7

class virtual +'a h5 : object ... end

val v5 : 'v h5 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 | `V4 | `V5 ], [> `V5 ]) Wayland.Proxy.handler

val v6 : 'v h5 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 ], [> `V6 ]) Wayland.Proxy.handler

val v7 : 'v h5 -> ([ `Wl_seat ], 'v, [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 | `V7 ], [> `V7 ]) Wayland.Proxy.handler