Module Wayland__Wayland_server.Wl_shm_pool

A shared memory pool.

The wl_shm_pool object encapsulates a piece of memory shared between the compositor and client. Through the wl_shm_pool object, the client can allocate shared memory wl_buffer objects. All objects created through the same pool share the same underlying mapped memory. Reusing the mapped memory avoids the setup/teardown overhead and is useful when interactively resizing a surface or for many small buffers.

type 'v t = ([ `Wl_shm_pool ], 'v) Wayland.Proxy.t

Version 1

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_shm_pool ], 'v, [< `V1 ], [> `V1 ]) Wayland.Proxy.handler