Module Gtk_primary_selection_client.Gtk_primary_selection_device

type 'v t = ([ `Gtk_primary_selection_device ], 'v, [ `Client ]) Proxy.t

Version 1

val destroy : [< `V1 ] t -> unit

Destroy the primary selection device.

Destroy the primary selection device.

val set_selection : [< `V1 ] t -> source:([ `Gtk_primary_selection_source ], 'a, [ `Client ]) Proxy.t option -> serial:int32 -> unit

Set the primary selection.

Replaces the current selection. The previous owner of the primary selection will receive a wp_primary_selection_source.cancelled event.

To unset the selection, set the source to NULL.

class virtual +'a handlers : object ... end