Class Gtk_primary_selection_device.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_data_offer : [> `V1 ] t -> ([ `Gtk_primary_selection_offer ], 'b, [ `Client ]) {Proxy}2.t -> unit

Introduce a new wp_primary_selection_offer.

Introduces a new wp_primary_selection_offer object that may be used to receive the current primary selection. Immediately following this event, the new wp_primary_selection_offer object will send wp_primary_selection_offer.offer events to describe the offered mime types.

method private virtual on_selection : [> `V1 ] t -> ([ `Gtk_primary_selection_offer ], [> {Imports}1.Gtk_primary_selection_offer.versions ], [ `Client ]) {Proxy}2.t option -> unit

Advertise a new primary selection.

The wp_primary_selection_device.selection event is sent to notify the client of a new primary selection. This event is sent after the wp_primary_selection.data_offer event introducing this object, and after the offer has announced its mimetypes through wp_primary_selection_offer.offer.

The data_offer is valid until a new offer or NULL is received or until the client loses keyboard focus. The client must destroy the previous selection data_offer, if any, upon receiving this event.

method min_version : int32