Module Gtk_primary_selection_client.Gtk_primary_selection_source

Offer to replace the contents of the primary selection.

The source side of a wp_primary_selection_offer, it provides a way to describe the offered data and respond to requests to transfer the requested contents of the primary selection clipboard.

type 'v t = ([ `Gtk_primary_selection_source ], 'v, [ `Client ]) Proxy.t

Version 1

val destroy : [< `V1 ] t -> unit

Destroy the primary selection source.

Destroy the primary selection source.

val offer : [< `V1 ] t -> mime_type:string -> unit

Add an offered mime type.

This request adds a mime type to the set of mime types advertised to targets. Can be called several times to offer multiple types.

class virtual +'a handlers : object ... end