Class Gtk_primary_selection_source.handlers

method user_data : [Gtk_primary_selection_proto.Gtk_primary_selection_source.t, 'v, [ `Client ]] Proxy.Handler.t -> (Gtk_primary_selection_proto.Gtk_primary_selection_source.t, [ `Client ]) Wayland.S.user_data

method metadata : [Gtk_primary_selection_proto.Gtk_primary_selection_source.t, 'v, [ `Client ]] Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Gtk_primary_selection_proto.Gtk_primary_selection_source.t)

method private virtual on_send : 'v t -> mime_type:string -> fd:Unix.file_descr -> unit

Send the primary selection contents.

Request for the current primary selection contents from the client. Send the specified mime type over the passed file descriptor, then close it.

method private virtual on_cancelled : 'v t -> unit

Request for primary selection contents was canceled.

This primary selection source is no longer valid. The client should clean up and destroy this primary selection source.

method dispatch : [Gtk_primary_selection_proto.Gtk_primary_selection_source.t, 'v, [ `Client ]] Proxy.Handler.t -> (Gtk_primary_selection_proto.Gtk_primary_selection_source.t, 'v, [ `Client ]) Proxy.t -> (Gtk_primary_selection_proto.Gtk_primary_selection_source.t, [ `R ]) Wayland.Msg.t -> unit