Class Gtk_primary_selection_device.handlers

method user_data : [Gtk_primary_selection_proto.Gtk_primary_selection_device.t, 'v, [ `Server ]] Proxy.Handler.t -> (Gtk_primary_selection_proto.Gtk_primary_selection_device.t, [ `Server ]) Wayland.S.user_data

method metadata : [Gtk_primary_selection_proto.Gtk_primary_selection_device.t, 'v, [ `Server ]] Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Gtk_primary_selection_proto.Gtk_primary_selection_device.t)

method private virtual on_set_selection : 'v t -> source:([ `Gtk_primary_selection_source ], 'v, [ `Server ]) Proxy.t option -> serial:int32 -> unit

Set the primary selection.

Replaces the current selection. The previous owner of the primary selection will receive a wp_primary_selection_source.cancelled event.

To unset the selection, set the source to NULL.

method private virtual on_destroy : 'v t -> unit

Destroy the primary selection device.

Destroy the primary selection device.

method dispatch : [Gtk_primary_selection_proto.Gtk_primary_selection_device.t, 'v, [ `Server ]] Proxy.Handler.t -> (Gtk_primary_selection_proto.Gtk_primary_selection_device.t, 'v, [ `Server ]) Proxy.t -> (Gtk_primary_selection_proto.Gtk_primary_selection_device.t, [ `R ]) Wayland.Msg.t -> unit