Class Gtk_primary_selection_device.v1

Handler for a proxy with version >= 1.

constraint 'b = [< `V1 ]

method private virtual on_set_selection : [> `V1 ] t -> source:([ `Gtk_primary_selection_source ], [> Wayland_protocols__Gtk_primary_selection_proto.Gtk_primary_selection_source.versions ], [ `Server ]) Wayland.Proxy.t option -> serial:int32 -> unit

Set the primary selection.

Replaces the current selection. The previous owner of the primary selection will receive a wp_primary_selection_source.cancelled event.

To unset the selection, set the source to NULL.

method private virtual on_destroy : [> `V1 ] t -> unit

Destroy the primary selection device.

Destroy the primary selection device.

method min_version : [Wayland_protocols.Gtk_primary_selection_proto.Gtk_primary_selection_device.t, 'b, [ `Server ]] {Proxy}2.Service_handler.t -> int32