Module Gtk_primary_selection_server.Gtk_primary_selection_device_manager

X primary selection emulation.

The primary selection device manager is a singleton global object that provides access to the primary selection. It allows to create wp_primary_selection_source objects, as well as retrieving the per-seat wp_primary_selection_device objects.

type 'v t = ([ `Gtk_primary_selection_device_manager ], 'v, [ `Server ]) Proxy.t

Version 1

class virtual +'a handlers : object ... end