Module Gtk_primary_selection_server.Gtk_primary_selection_offer

Offer to transfer primary selection contents.

A wp_primary_selection_offer represents an offer to transfer the contents of the primary selection clipboard to the client. Similar to wl_data_offer, the offer also describes the mime types that the source will transferthat the data can be converted to and provides the mechanisms for transferring the data directly to the client.

type 'v t = ([ `Gtk_primary_selection_offer ], 'v, [ `Server ]) Wayland.Proxy.t

Version 1

val offer : [< `V1 ] t -> mime_type:string -> unit

Advertise offered mime type.

Sent immediately after creating announcing the wp_primary_selection_offer through wp_primary_selection_device.data_offer. One event is sent per offered mime type.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.