Class Gtk_primary_selection_offer.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_receive : [> `V1 ] t -> mime_type:string -> fd:Unix.file_descr -> unit

Request that the data is transferred.

To transfer the contents of the primary selection clipboard, the client issues this request and indicates the mime type that it wants to receive. The transfer happens through the passed file descriptor (typically created with the pipe system call). The source client writes the data in the mime type representation requested and then closes the file descriptor.

The receiving client reads from the read end of the pipe until EOF and closes its end, at which point the transfer is complete.

method private virtual on_destroy : [> `V1 ] t -> unit

Destroy the primary selection offer.

Destroy the primary selection offer.

method min_version : int32