Module Gtk_primary_selection_server.Gtk_primary_selection_source

Offer to replace the contents of the primary selection.

The source side of a wp_primary_selection_offer, it provides a way to describe the offered data and respond to requests to transfer the requested contents of the primary selection clipboard.

type 'v t = ([ `Gtk_primary_selection_source ], 'v, [ `Server ]) Wayland.Proxy.t

Version 1

val cancelled : [< `V1 ] t -> unit

Request for primary selection contents was canceled.

This primary selection source is no longer valid. The client should clean up and destroy this primary selection source.

val send : [< `V1 ] t -> mime_type:string -> fd:Unix.file_descr -> unit

Send the primary selection contents.

Request for the current primary selection contents from the client. Send the specified mime type over the passed file descriptor, then close it.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.