Class Gtk_primary_selection_source.v1

Handler for a proxy with version >= 1.

constraint 'b = [< `V1 ]

method private virtual on_offer : [> `V1 ] t -> mime_type:string -> unit

Add an offered mime type.

This request adds a mime type to the set of mime types advertised to targets. Can be called several times to offer multiple types.

method private virtual on_destroy : [> `V1 ] t -> unit

Destroy the primary selection source.

Destroy the primary selection source.

method min_version : [Wayland_protocols.Gtk_primary_selection_proto.Gtk_primary_selection_source.t, 'b, [ `Server ]] {Proxy}2.Service_handler.t -> int32