Module Linux_dmabuf_unstable_v1_client.Zwp_linux_buffer_params_v1

Parameters for creating a dmabuf-based wl_buffer.

This temporary object is a collection of dmabufs and other parameters that together form a single logical buffer. The temporary object may eventually create one wl_buffer unless cancelled by destroying it before requesting 'create'.

Single-planar formats only require one dmabuf, however multi-planar formats may require more than one dmabuf. For all formats, an 'add' request must be called once per plane (even if the underlying dmabuf fd is identical).

You must use consecutive plane indices ('plane_idx' argument for 'add') from zero to the number of planes used by the drm_fourcc format code. All planes required by the format must be given exactly once, but can be given in any order. Each plane index can be set only once.

type 'v t = ([ `Zwp_linux_buffer_params_v1 ], 'v, [ `Client ]) Wayland.Proxy.t

module Error = Linux_dmabuf_unstable_v1_proto.Zwp_linux_buffer_params_v1.Error

module Flags = Linux_dmabuf_unstable_v1_proto.Zwp_linux_buffer_params_v1.Flags

Version 1

val create : [< `V1 | `V2 | `V3 | `V4 ] t -> width:int32 -> height:int32 -> format:int32 -> flags:int32 -> unit

Create a wl_buffer from the given dmabufs.

This asks for creation of a wl_buffer from the added dmabuf buffers. The wl_buffer is not created immediately but returned via the 'created' event if the dmabuf sharing succeeds. The sharing may fail at runtime for reasons a client cannot predict, in which case the 'failed' event is triggered.

The 'format' argument is a DRM_FORMAT code, as defined by the libdrm's drm_fourcc.h. The Linux kernel's DRM sub-system is the authoritative source on how the format codes should work.

The 'flags' is a bitfield of the flags defined in enum "flags". 'y_invert' means the that the image needs to be y-flipped.

Flag 'interlaced' means that the frame in the buffer is not progressive as usual, but interlaced. An interlaced buffer as supported here must always contain both top and bottom fields. The top field always begins on the first pixel row. The temporal ordering between the two fields is top field first, unless 'bottom_first' is specified. It is undefined whether 'bottom_first' is ignored if 'interlaced' is not set.

This protocol does not convey any information about field rate, duration, or timing, other than the relative ordering between the two fields in one buffer. A compositor may have to estimate the intended field rate from the incoming buffer rate. It is undefined whether the time of receiving wl_surface.commit with a new buffer attached, applying the wl_surface state, wl_surface.frame callback trigger, presentation, or any other point in the compositor cycle is used to measure the frame or field times. There is no support for detecting missed or late frames/fields/buffers either, and there is no support whatsoever for cooperating with interlaced compositor output.

The composited image quality resulting from the use of interlaced buffers is explicitly undefined. A compositor may use elaborate hardware features or software to deinterlace and create progressive output frames from a sequence of interlaced input buffers, or it may produce substandard image quality. However, compositors that cannot guarantee reasonable image quality in all cases are recommended to just reject all interlaced buffers.

Any argument errors, including non-positive width or height, mismatch between the number of planes and the format, bad format, bad offset or stride, may be indicated by fatal protocol errors: INCOMPLETE, INVALID_FORMAT, INVALID_DIMENSIONS, OUT_OF_BOUNDS.

Dmabuf import errors in the server that are not obvious client bugs are returned via the 'failed' event as non-fatal. This allows attempting dmabuf sharing and falling back in the client if it fails.

This request can be sent only once in the object's lifetime, after which the only legal request is destroy. This object should be destroyed after issuing a 'create' request. Attempting to use this object after issuing 'create' raises ALREADY_USED protocol error.

It is not mandatory to issue 'create'. If a client wants to cancel the buffer creation, it can just destroy this object.

val add : [< `V1 | `V2 | `V3 | `V4 ] t -> fd:Unix.file_descr -> plane_idx:int32 -> offset:int32 -> stride:int32 -> modifier_hi:int32 -> modifier_lo:int32 -> unit

Add a dmabuf to the temporary set.

This request adds one dmabuf to the set in this zwp_linux_buffer_params_v1.

The 64-bit unsigned value combined from modifier_hi and modifier_lo is the dmabuf layout modifier. DRM AddFB2 ioctl calls this the fb modifier, which is defined in drm_mode.h of Linux UAPI. This is an opaque token. Drivers use this token to express tiling, compression, etc. driver-specific modifications to the base format defined by the DRM fourcc code.

Starting from version 4, the invalid_format protocol error is sent if the format + modifier pair was not advertised as supported.

This request raises the PLANE_IDX error if plane_idx is too large. The error PLANE_SET is raised if attempting to set a plane that was already set.

val destroy : [< `V1 | `V2 | `V3 | `V4 ] t -> unit

Delete this object, used or not.

Cleans up the temporary data sent to the server for dmabuf-based wl_buffer creation.

Version 2, 3, 4

val create_immed : [< `V2 | `V3 | `V4 ] as 'a t -> [[ `Wl_buffer ], 'b, [ `Client ]] Wayland.Proxy.Handler.t -> width:int32 -> height:int32 -> format:int32 -> flags:int32 -> ([ `Wl_buffer ], 'c, [ `Client ]) Wayland.Proxy.t

Immediately create a wl_buffer from the given dmabufs.

This asks for immediate creation of a wl_buffer by importing the added dmabufs.

In case of import success, no event is sent from the server, and the wl_buffer is ready to be used by the client.

Upon import failure, either of the following may happen, as seen fit by the implementation:

the client is terminated with one of the following fatal protocol errors:

INCOMPLETE, INVALID_FORMAT, INVALID_DIMENSIONS, OUT_OF_BOUNDS, in case of argument errors such as mismatch between the number of planes and the format, bad format, non-positive width or height, or bad offset or stride.

INVALID_WL_BUFFER, in case the cause for failure is unknown or plaform specific.

the server creates an invalid wl_buffer, marks it as failed and sends a 'failed' event to the client. The result of using this invalid wl_buffer as an argument in any request by the client is defined by the compositor implementation.

This takes the same arguments as a 'create' request, and obeys the same restrictions.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.

class virtual +'a v4 : object ... end

Handler for a proxy with version >= 4.