Class Zwp_linux_buffer_params_v1.v4

Handler for a proxy with version >= 4.

constraint 'a = [< `V4 | `V5 ]

method private virtual on_created : [> `V4 | `V5 ] t -> ([ `Wl_buffer ], [ `V1 ], [ `Client ]) {Proxy}2.t -> unit

Buffer creation succeeded.

This event indicates that the attempted buffer creation was successful. It provides the new wl_buffer referencing the dmabuf(s).

Upon receiving this event, the client should destroy the zwp_linux_buffer_params_v1 object.

method private virtual on_failed : [> `V4 | `V5 ] t -> unit

Buffer creation failed.

This event indicates that the attempted buffer creation has failed. It usually means that one of the dmabuf constraints has not been fulfilled.

Upon receiving this event, the client should destroy the zwp_linux_buffer_params_v1 object.

method min_version : int32