Module Linux_dmabuf_unstable_v1_client.Zwp_linux_dmabuf_feedback_v1

Dmabuf feedback.

This object advertises dmabuf parameters feedback. This includes the preferred devices and the supported formats/modifiers.

The parameters are sent once when this object is created and whenever they change. The done event is always sent once after all parameters have been sent. When a single parameter changes, all parameters are re-sent by the compositor.

Compositors can re-send the parameters when the current client buffer allocations are sub-optimal. Compositors should not re-send the parameters if re-allocating the buffers would not result in a more optimal configuration. In particular, compositors should avoid sending the exact same parameters multiple times in a row.

The tranche_target_device and tranche_formats events are grouped by tranches of preference. For each tranche, a tranche_target_device, one tranche_flags and one or more tranche_formats events are sent, followed by a tranche_done event finishing the list. The tranches are sent in descending order of preference. All formats and modifiers in the same tranche have the same preference.

To send parameters, the compositor sends one main_device event, tranches (each consisting of one tranche_target_device event, one tranche_flags event, tranche_formats events and then a tranche_done event), then one done event.

type 'v t = ([ `Zwp_linux_dmabuf_feedback_v1 ], 'v, [ `Client ]) Wayland.Proxy.t

module Tranche_flags = Linux_dmabuf_unstable_v1_proto.Zwp_linux_dmabuf_feedback_v1.Tranche_flags

Version 1, 2, 3, 4

val destroy : [< `V1 | `V2 | `V3 | `V4 ] t -> unit

Destroy the feedback object.

Using this request a client can tell the server that it is not going to use the wp_linux_dmabuf_feedback object anymore.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.

class virtual +'a v4 : object ... end

Handler for a proxy with version >= 4.