Class Zwp_linux_dmabuf_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 | `V2 | `V3 | `V4 ]

method private virtual on_format : [> `V1 | `V2 | `V3 | `V4 ] t -> format:int32 -> unit

Supported buffer format.

This event advertises one buffer format that the server supports. All the supported formats are advertised once when the client binds to this interface. A roundtrip after binding guarantees that the client has received all supported formats.

For the definition of the format codes, see the zwp_linux_buffer_params_v1::create request.

Starting version 4, the format event is deprecated and must not be sent by compositors. Instead, use get_default_feedback or get_surface_feedback.

method private virtual on_modifier : [> `V3 | `V4 ] t -> format:int32 -> modifier_hi:int32 -> modifier_lo:int32 -> unit

Supported buffer format modifier.

This event advertises the formats that the server supports, along with the modifiers supported for each format. All the supported modifiers for all the supported formats are advertised once when the client binds to this interface. A roundtrip after binding guarantees that the client has received all supported format-modifier pairs.

For legacy support, DRM_FORMAT_MOD_INVALID (that is, modifier_hi == 0x00ffffff and modifier_lo == 0xffffffff) is allowed in this event. It indicates that the server can support the format with an implicit modifier. When a plane has DRM_FORMAT_MOD_INVALID as its modifier, it is as if no explicit modifier is specified. The effective modifier will be derived from the dmabuf.

A compositor that sends valid modifiers and DRM_FORMAT_MOD_INVALID for a given format supports both explicit modifiers and implicit modifiers.

For the definition of the format and modifier codes, see the zwp_linux_buffer_params_v1::create and zwp_linux_buffer_params_v1::add requests.

Starting version 4, the modifier event is deprecated and must not be sent by compositors. Instead, use get_default_feedback or get_surface_feedback.

method min_version : int32

method bind_version : [ `V1 ]