Class Zwp_linux_dmabuf_feedback_v1.v4

Handler for a proxy with version >= 4.

constraint 'a = [< `V4 ]

method private virtual on_destroy : [> `V4 ] t -> unit

Destroy the feedback object.

Using this request a client can tell the server that it is not going to use the wp_linux_dmabuf_feedback object anymore.

method min_version : int32