Class Zwp_linux_dmabuf_v1.v2

Handler for a proxy with version >= 2.

constraint 'a = [< `V2 | `V3 | `V4 ]

method private virtual on_destroy : [> `V2 | `V3 | `V4 ] t -> unit

Unbind the factory.

Objects created through this interface, especially wl_buffers, will remain valid.

method private virtual on_create_params : [> `V2 | `V3 | `V4 ] t -> ( [ `Zwp_linux_buffer_params_v1 ], 'b, [ `Server ] ) Wayland.Proxy.t -> unit

Create a temporary object for buffer parameters.

This temporary object is used to collect multiple dmabuf handles into a single batch to create a wl_buffer. It can only be used once and should be destroyed after a 'created' or 'failed' event has been received.

method private virtual on_get_default_feedback : [> `V4 ] t -> ( [ `Zwp_linux_dmabuf_feedback_v1 ], 'b, [ `Server ] ) Wayland.Proxy.t -> unit

Get default feedback.

This request creates a new wp_linux_dmabuf_feedback object not bound to a particular surface. This object will deliver feedback about dmabuf parameters to use if the client doesn't support per-surface feedback (see get_surface_feedback).

method private virtual on_get_surface_feedback : [> `V4 ] t -> ( [ `Zwp_linux_dmabuf_feedback_v1 ], 'b, [ `Server ] ) Wayland.Proxy.t -> surface: ( [ `Wl_surface ], [> Wayland__Wayland_proto.Wl_surface.versions ], [ `Server ] ) Wayland.Proxy.t -> unit

Get feedback for a surface.

This request creates a new wp_linux_dmabuf_feedback object for the specified wl_surface. This object will deliver feedback about dmabuf parameters to use for buffers attached to this surface.

If the surface is destroyed before the wp_linux_dmabuf_feedback object, the feedback object becomes inert.

method min_version : 'a v2 -> int32

method bind_version : 'c -> [ `V2 ]