Class Zwp_linux_dmabuf_v1.v5

Handler for a proxy with version >= 5.

constraint 'a = [< `V5 ]

method private virtual on_destroy : [> `V5 ] t -> unit

Unbind the factory.

Objects created through this interface, especially wl_buffers, will remain valid.

method private virtual on_create_params : [> `V5 ] t -> ([ `Zwp_linux_buffer_params_v1 ], 'b, [ `Server ]) {Proxy}2.t -> unit

Create a temporary object for buffer parameters.

This temporary object is used to collect multiple dmabuf handles into a single batch to create a wl_buffer. It can only be used once and should be destroyed after a 'created' or 'failed' event has been received.

method private virtual on_get_default_feedback : [> `V5 ] t -> ([ `Zwp_linux_dmabuf_feedback_v1 ], 'b, [ `Server ]) {Proxy}2.t -> unit

Get default feedback.

This request creates a new wp_linux_dmabuf_feedback object not bound to a particular surface. This object will deliver feedback about dmabuf parameters to use if the client doesn't support per-surface feedback (see get_surface_feedback).

method private virtual on_get_surface_feedback : [> `V5 ] t -> ([ `Zwp_linux_dmabuf_feedback_v1 ], 'b, [ `Server ]) {Proxy}2.t -> surface: ([ `Wl_surface ], [> {Imports}1.Wl_surface.versions ], [ `Server ]) {Proxy}2.t -> unit

Get feedback for a surface.

This request creates a new wp_linux_dmabuf_feedback object for the specified wl_surface. This object will deliver feedback about dmabuf parameters to use for buffers attached to this surface.

If the surface is destroyed before the wp_linux_dmabuf_feedback object, the feedback object becomes inert.

method min_version : int32

method bind_version : [ `V5 ]