Class Zwp_confined_pointer_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_confined : [> `V1 ] t -> unit

Pointer confined.

Notification that the pointer confinement of the seat's pointer is activated.

method private virtual on_unconfined : [> `V1 ] t -> unit

Pointer unconfined.

Notification that the pointer confinement of the seat's pointer is no longer active. If this is a oneshot pointer confinement (see wp_pointer_constraints.lifetime) this object is now defunct and should be destroyed. If this is a persistent pointer confinement (see wp_pointer_constraints.lifetime) this pointer confinement may again reactivate in the future.

method min_version : int32