Module Pointer_constraints_unstable_v1_client.Zwp_locked_pointer_v1

Receive relative pointer motion events.

The wp_locked_pointer interface represents a locked pointer state.

While the lock of this object is active, the wl_pointer objects of the associated seat will not emit any wl_pointer.motion events.

This object will send the event 'locked' when the lock is activated. Whenever the lock is activated, it is guaranteed that the locked surface will already have received pointer focus and that the pointer will be within the region passed to the request creating this object.

To unlock the pointer, send the destroy request. This will also destroy the wp_locked_pointer object.

If the compositor decides to unlock the pointer the unlocked event is sent. See wp_locked_pointer.unlock for details.

When unlocking, the compositor may warp the cursor position to the set cursor position hint. If it does, it will not result in any relative motion events emitted via wp_relative_pointer.

If the surface the lock was requested on is destroyed and the lock is not yet activated, the wp_locked_pointer object is now defunct and must be destroyed.

type 'v t = ([ `Zwp_locked_pointer_v1 ], 'v, [ `Client ]) Wayland.Proxy.t

Version 1

val set_region : [< `V1 ] t -> region:([ `Wl_region ], 'a, [ `Client ]) Wayland.Proxy.t option -> unit

Set a new lock region.

Set a new region used to lock the pointer.

The new lock region is double-buffered. The new lock region will only take effect when the associated surface gets its pending state applied. See wl_surface.commit for details.

For details about the lock region, see wp_locked_pointer.

val set_cursor_position_hint : [< `V1 ] t -> surface_x:Wayland.Fixed.t -> surface_y:Wayland.Fixed.t -> unit

Set the pointer cursor position hint.

Set the cursor position hint relative to the top left corner of the surface.

If the client is drawing its own cursor, it should update the position hint to the position of its own cursor. A compositor may use this information to warp the pointer upon unlock in order to avoid pointer jumps.

The cursor position hint is double buffered. The new hint will only take effect when the associated surface gets it pending state applied. See wl_surface.commit for details.

val destroy : [< `V1 ] t -> unit

Destroy the locked pointer object.

Destroy the locked pointer object. If applicable, the compositor will unlock the pointer.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.