Module Zwp_pointer_constraints_v1.Lifetime

Constraint lifetime.

These values represent different lifetime semantics. They are passed as arguments to the factory requests to specify how the constraint lifetimes should be managed.

type t =

| Oneshot : t
(*
The pointer constraint is defunct once deactivated.

A oneshot pointer constraint will never reactivate once it has been deactivated. See the corresponding deactivation event (wp_locked_pointer.unlocked and wp_confined_pointer.unconfined) for details.
*)

| Persistent : t
(*
The pointer constraint may reactivate.

A persistent pointer constraint may again reactivate once it has been deactivated. See the corresponding deactivation event (wp_locked_pointer.unlocked and wp_confined_pointer.unconfined) for details.
*)

val to_int32 : t -> int32

val of_int32 : int32 -> t