Class Zwp_confined_pointer_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_destroy : [> `V1 ] t -> unit

Destroy the confined pointer object.

Destroy the confined pointer object. If applicable, the compositor will unconfine the pointer.

method private virtual on_set_region : [> `V1 ] t -> region: ([ `Wl_region ], [> {Imports}1.Wl_region.versions ], [ `Server ]) {Proxy}2.t option -> unit

Set a new confine region.

Set a new region used to confine the pointer.

The new confine region is double-buffered. The new confine region will only take effect when the associated surface gets its pending state applied. See wl_surface.commit for details.

If the confinement is active when the new confinement region is applied and the pointer ends up outside of newly applied region, the pointer may warped to a position within the new confinement region. If warped, a wl_pointer.motion event will be emitted, but no wp_relative_pointer.relative_motion event.

The compositor may also, instead of using the new region, unconfine the pointer.

For details about the confine region, see wp_confined_pointer.

method min_version : int32