Module Pointer_constraints_unstable_v1_server.Zwp_locked_pointer_v1

Receive relative pointer motion events.

The wp_locked_pointer interface represents a locked pointer state.

While the lock of this object is active, the wl_pointer objects of the associated seat will not emit any wl_pointer.motion events.

This object will send the event 'locked' when the lock is activated. Whenever the lock is activated, it is guaranteed that the locked surface will already have received pointer focus and that the pointer will be within the region passed to the request creating this object.

To unlock the pointer, send the destroy request. This will also destroy the wp_locked_pointer object.

If the compositor decides to unlock the pointer the unlocked event is sent. See wp_locked_pointer.unlock for details.

When unlocking, the compositor may warp the cursor position to the set cursor position hint. If it does, it will not result in any relative motion events emitted via wp_relative_pointer.

If the surface the lock was requested on is destroyed and the lock is not yet activated, the wp_locked_pointer object is now defunct and must be destroyed.

type 'v t = ([ `Zwp_locked_pointer_v1 ], 'v, [ `Server ]) Wayland.Proxy.t

Version 1

val unlocked : [< `V1 ] t -> unit

Lock deactivation event.

Notification that the pointer lock of the seat's pointer is no longer active. If this is a oneshot pointer lock (see wp_pointer_constraints.lifetime) this object is now defunct and should be destroyed. If this is a persistent pointer lock (see wp_pointer_constraints.lifetime) this pointer lock may again reactivate in the future.

val locked : [< `V1 ] t -> unit

Lock activation event.

Notification that the pointer lock of the seat's pointer is activated.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.