Class Zwp_locked_pointer_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_destroy : [> `V1 ] t -> unit

Destroy the locked pointer object.

Destroy the locked pointer object. If applicable, the compositor will unlock the pointer.

method private virtual on_set_cursor_position_hint : [> `V1 ] t -> surface_x:{Fixed}4.t -> surface_y:{Fixed}4.t -> unit

Set the pointer cursor position hint.

Set the cursor position hint relative to the top left corner of the surface.

If the client is drawing its own cursor, it should update the position hint to the position of its own cursor. A compositor may use this information to warp the pointer upon unlock in order to avoid pointer jumps.

The cursor position hint is double buffered. The new hint will only take effect when the associated surface gets it pending state applied. See wl_surface.commit for details.

method private virtual on_set_region : [> `V1 ] t -> region: ([ `Wl_region ], [> {Imports}1.Wl_region.versions ], [ `Server ]) {Proxy}2.t option -> unit

Set a new lock region.

Set a new region used to lock the pointer.

The new lock region is double-buffered. The new lock region will only take effect when the associated surface gets its pending state applied. See wl_surface.commit for details.

For details about the lock region, see wp_locked_pointer.

method min_version : int32