Module Pointer_constraints_unstable_v1_server.Zwp_pointer_constraints_v1

Constrain the movement of a pointer.

The global interface exposing pointer constraining functionality. It exposes two requests: lock_pointer for locking the pointer to its position, and confine_pointer for locking the pointer to a region.

The lock_pointer and confine_pointer requests create the objects wp_locked_pointer and wp_confined_pointer respectively, and the client can use these objects to interact with the lock.

For any surface, only one lock or confinement may be active across all wl_pointer objects of the same seat. If a lock or confinement is requested when another lock or confinement is active or requested on the same surface and with any of the wl_pointer objects of the same seat, an 'already_constrained' error will be raised.

type 'v t = ([ `Zwp_pointer_constraints_v1 ], 'v, [ `Server ]) Wayland.Proxy.t

module Error = Pointer_constraints_unstable_v1_proto.Zwp_pointer_constraints_v1.Error

module Lifetime = Pointer_constraints_unstable_v1_proto.Zwp_pointer_constraints_v1.Lifetime

Version 1

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.