Module Relative_pointer_unstable_v1_client.Zwp_relative_pointer_manager_v1

Get relative pointer objects.

A global interface used for getting the relative pointer object for a given pointer.

type 'v t = ([ `Zwp_relative_pointer_manager_v1 ], 'v, [ `Client ]) Wayland.Proxy.t

Version 1

val get_relative_pointer : [< `V1 ] as 'a t -> [[ `Zwp_relative_pointer_v1 ], 'b, [ `Client ]] Wayland.Proxy.Handler.t -> pointer:([ `Wl_pointer ], 'c, [ `Client ]) Wayland.Proxy.t -> ([ `Zwp_relative_pointer_v1 ], 'd, [ `Client ]) Wayland.Proxy.t

Get a relative pointer object.

Create a relative pointer interface given a wl_pointer object. See the wp_relative_pointer interface for more details.

val destroy : [< `V1 ] t -> unit

Destroy the relative pointer manager object.

Used by the client to notify the server that it will no longer use this relative pointer manager object.

Handlers

Note: Servers will always want to use v1.

class +'a v1 : object ... end

Handler for a proxy with version >= 1.