Class Zwp_relative_pointer_manager_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_destroy : [> `V1 ] t -> unit

Destroy the relative pointer manager object.

Used by the client to notify the server that it will no longer use this relative pointer manager object.

method private virtual on_get_relative_pointer : [> `V1 ] t -> ([ `Zwp_relative_pointer_v1 ], 'b, [ `Server ]) {Proxy}2.t -> pointer: ([ `Wl_pointer ], [> {Imports}1.Wl_pointer.versions ], [ `Server ]) {Proxy}2.t -> unit

Get a relative pointer object.

Create a relative pointer interface given a wl_pointer object. See the wp_relative_pointer interface for more details.

method min_version : int32

method bind_version : [ `V1 ]