Module Server_decoration_client.Org_kde_kwin_server_decoration_manager

Server side window decoration manager.

This interface allows to coordinate whether the server should create a server-side window decoration around a wl_surface representing a shell surface (wl_shell_surface or similar). By announcing support for this interface the server indicates that it supports server side decorations.

type 'v t = ([ `Org_kde_kwin_server_decoration_manager ], 'v, [ `Client ]) $Proxy.t

module Mode = Server_decoration_proto.Org_kde_kwin_server_decoration_manager.Mode

Version 1

val create : [< `V1 ] as 'a t -> [[ `Org_kde_kwin_server_decoration ], 'a, [ `Client ]] $Proxy.Handler.t -> surface:([ `Wl_surface ], 'b, [ `Client ]) $Proxy.t -> ([ `Org_kde_kwin_server_decoration ], 'a, [ `Client ]) $Proxy.t

Create a server-side decoration object for a given surface.

When a client creates a server-side decoration object it indicates that it supports the protocol. The client is supposed to tell the server whether it wants server-side decorations or will provide client-side decorations.

If the client does not create a server-side decoration object for a surface the server interprets this as lack of support for this protocol and considers it as client-side decorated. Nevertheless a client-side decorated surface should use this protocol to indicate to the server that it does not want a server-side deco.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.