Class Org_kde_kwin_server_decoration_manager._handlers_unsafe

method user_data : [Server_decoration_proto.Org_kde_kwin_server_decoration_manager.t, 'v, [ `Client ]] Proxy.Handler.t -> (Server_decoration_proto.Org_kde_kwin_server_decoration_manager.t, [ `Client ]) Wayland.S.user_data

method metadata : [Server_decoration_proto.Org_kde_kwin_server_decoration_manager.t, 'v, [ `Client ]] Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Server_decoration_proto.Org_kde_kwin_server_decoration_manager.t)

method private virtual on_default_mode : 'v t -> mode:int32 -> unit

The default mode used on the server.

This event is emitted directly after binding the interface. It contains the default mode for the decoration. When a new server decoration object is created this new object will be in the default mode until the first request_mode is requested.

The server may change the default mode at any time.

method dispatch : [Server_decoration_proto.Org_kde_kwin_server_decoration_manager.t, 'v, [ `Client ]] Proxy.Handler.t -> (Server_decoration_proto.Org_kde_kwin_server_decoration_manager.t, 'v, [ `Client ]) Proxy.t -> (Server_decoration_proto.Org_kde_kwin_server_decoration_manager.t, [ `R ]) Wayland.Msg.t -> unit