Class Org_kde_kwin_server_decoration_manager.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_default_mode : [> `V1 ] t -> mode:int32 -> unit

The default mode used on the server.

This event is emitted directly after binding the interface. It contains the default mode for the decoration. When a new server decoration object is created this new object will be in the default mode until the first request_mode is requested.

The server may change the default mode at any time.

method min_version : 'a v1 -> int32

method bind_version : 'b -> [ `V1 ]