Module Server_decoration_server.Org_kde_kwin_server_decoration_manager

Server side window decoration manager.

This interface allows to coordinate whether the server should create a server-side window decoration around a wl_surface representing a shell surface (wl_shell_surface or similar). By announcing support for this interface the server indicates that it supports server side decorations.

type 'v t = ([ `Org_kde_kwin_server_decoration_manager ], 'v, [ `Server ]) $Proxy.t

module Mode = Server_decoration_proto.Org_kde_kwin_server_decoration_manager.Mode

Version 1

val default_mode : [< `V1 ] t -> mode:int32 -> unit

The default mode used on the server.

This event is emitted directly after binding the interface. It contains the default mode for the decoration. When a new server decoration object is created this new object will be in the default mode until the first request_mode is requested.

The server may change the default mode at any time.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.