Class Org_kde_kwin_server_decoration_manager.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_create : [> `V1 ] t -> ([ `Org_kde_kwin_server_decoration ], 'b, [ `Server ]) {Proxy}2.t -> surface: ([ `Wl_surface ], [> {Imports}1.Wl_surface.versions ], [ `Server ]) {Proxy}2.t -> unit

Create a server-side decoration object for a given surface.

When a client creates a server-side decoration object it indicates that it supports the protocol. The client is supposed to tell the server whether it wants server-side decorations or will provide client-side decorations.

If the client does not create a server-side decoration object for a surface the server interprets this as lack of support for this protocol and considers it as client-side decorated. Nevertheless a client-side decorated surface should use this protocol to indicate to the server that it does not want a server-side deco.

method min_version : int32

method bind_version : [ `V1 ]