Module Wlr_layer_shell_unstable_v1_client.Zwlr_layer_surface_v1

Layer metadata interface.

An interface that may be implemented by a wl_surface, for surfaces that are designed to be rendered as a layer of a stacked desktop-like environment.

Layer surface state (layer, size, anchor, exclusive zone, margin, interactivity) is double-buffered, and will be applied at the time wl_surface.commit of the corresponding wl_surface is called.

Attaching a null buffer to a layer surface unmaps it.

Unmapping a layer_surface means that the surface cannot be shown by the compositor until it is explicitly mapped again. The layer_surface returns to the state it had right after layer_shell.get_layer_surface. The client can re-map the surface by performing a commit without any buffer attached, waiting for a configure event and handling it as usual.

type 'v t = ([ `Zwlr_layer_surface_v1 ], 'v, [ `Client ]) Wayland.Proxy.t

module Keyboard_interactivity = Wlr_layer_shell_unstable_v1_proto.Zwlr_layer_surface_v1.Keyboard_interactivity

module Error = Wlr_layer_shell_unstable_v1_proto.Zwlr_layer_surface_v1.Error

module Anchor = Wlr_layer_shell_unstable_v1_proto.Zwlr_layer_surface_v1.Anchor

Version 1

val destroy : [< `V1 | `V2 | `V3 | `V4 ] t -> unit

Destroy the layer_surface.

This request destroys the layer surface.

val ack_configure : [< `V1 | `V2 | `V3 | `V4 ] t -> serial:int32 -> unit

Ack a configure event.

When a configure event is received, if a client commits the surface in response to the configure event, then the client must make an ack_configure request sometime before the commit request, passing along the serial of the configure event.

If the client receives multiple configure events before it can respond to one, it only has to ack the last configure event.

A client is not required to commit immediately after sending an ack_configure request - it may even ack_configure several times before its next surface commit.

A client may send multiple ack_configure requests before committing, but only the last request sent before a commit indicates which configure event the client really is responding to.

val get_popup : [< `V1 | `V2 | `V3 | `V4 ] t -> popup:([ `Xdg_popup ], 'a, [ `Client ]) Wayland.Proxy.t -> unit

Assign this layer_surface as an xdg_popup parent.

This assigns an xdg_popup's parent to this layer_surface. This popup should have been created via xdg_surface::get_popup with the parent set to NULL, and this request must be invoked before committing the popup's initial state.

See the documentation of xdg_popup for more details about what an xdg_popup is and how it is used.

val set_keyboard_interactivity : [< `V1 | `V2 | `V3 | `V4 ] t -> keyboard_interactivity: {Imports}1.Zwlr_layer_surface_v1.Keyboard_interactivity.t -> unit

Requests keyboard events.

Set how keyboard events are delivered to this surface. By default, layer shell surfaces do not receive keyboard events; this request can be used to change this.

This setting is inherited by child surfaces set by the get_popup request.

Layer surfaces receive pointer, touch, and tablet events normally. If you do not want to receive them, set the input region on your surface to an empty region.

Keyboard interactivity is double-buffered, see wl_surface.commit.

val set_margin : [< `V1 | `V2 | `V3 | `V4 ] t -> top:int32 -> right:int32 -> bottom:int32 -> left:int32 -> unit

Sets a margin from the anchor point.

Requests that the surface be placed some distance away from the anchor point on the output, in surface-local coordinates. Setting this value for edges you are not anchored to has no effect.

The exclusive zone includes the margin.

Margin is double-buffered, see wl_surface.commit.

val set_exclusive_zone : [< `V1 | `V2 | `V3 | `V4 ] t -> zone:int32 -> unit

Configures the exclusive geometry of this surface.

Requests that the compositor avoids occluding an area with other surfaces. The compositor's use of this information is implementation-dependent - do not assume that this region will not actually be occluded.

A positive value is only meaningful if the surface is anchored to one edge or an edge and both perpendicular edges. If the surface is not anchored, anchored to only two perpendicular edges (a corner), anchored to only two parallel edges or anchored to all edges, a positive value will be treated the same as zero.

A positive zone is the distance from the edge in surface-local coordinates to consider exclusive.

Surfaces that do not wish to have an exclusive zone may instead specify how they should interact with surfaces that do. If set to zero, the surface indicates that it would like to be moved to avoid occluding surfaces with a positive exclusive zone. If set to -1, the surface indicates that it would not like to be moved to accommodate for other surfaces, and the compositor should extend it all the way to the edges it is anchored to.

For example, a panel might set its exclusive zone to 10, so that maximized shell surfaces are not shown on top of it. A notification might set its exclusive zone to 0, so that it is moved to avoid occluding the panel, but shell surfaces are shown underneath it. A wallpaper or lock screen might set their exclusive zone to -1, so that they stretch below or over the panel.

The default value is 0.

Exclusive zone is double-buffered, see wl_surface.commit.

val set_anchor : [< `V1 | `V2 | `V3 | `V4 ] t -> anchor:int32 -> unit

Configures the anchor point of the surface.

Requests that the compositor anchor the surface to the specified edges and corners. If two orthogonal edges are specified (e.g. 'top' and 'left'), then the anchor point will be the intersection of the edges (e.g. the top left corner of the output); otherwise the anchor point will be centered on that edge, or in the center if none is specified.

Anchor is double-buffered, see wl_surface.commit.

val set_size : [< `V1 | `V2 | `V3 | `V4 ] t -> width:int32 -> height:int32 -> unit

Sets the size of the surface.

Sets the size of the surface in surface-local coordinates. The compositor will display the surface centered with respect to its anchors.

If you pass 0 for either value, the compositor will assign it and inform you of the assignment in the configure event. You must set your anchor to opposite edges in the dimensions you omit; not doing so is a protocol error. Both values are 0 by default.

Size is double-buffered, see wl_surface.commit.

Version 2, 3, 4

val set_layer : [< `V2 | `V3 | `V4 ] t -> layer:{Imports}1.Zwlr_layer_shell_v1.Layer.t -> unit

Change the layer of the surface.

Change the layer that the surface is rendered on.

Layer is double-buffered, see wl_surface.commit.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.

class virtual +'a v4 : object ... end

Handler for a proxy with version >= 4.