Module Zwlr_layer_shell_v1.Layer

Available layers for surfaces.

These values indicate which layers a surface can be rendered in. They are ordered by z depth, bottom-most first. Traditional shell surfaces will typically be rendered between the bottom and top layers. Fullscreen shell surfaces are typically rendered at the top layer. Multiple surfaces can share a single layer, and ordering within a single layer is undefined.

type t =

| Background : t

| Bottom : t

| Top : t

| Overlay : t

val to_int32 : t -> int32

val of_int32 : int32 -> t