Module Wlr_layer_shell_unstable_v1_server.Zwlr_layer_shell_v1

Create surfaces that are layers of the desktop.

Clients can use this interface to assign the surface_layer role to wl_surfaces. Such surfaces are assigned to a "layer" of the output and rendered with a defined z-depth respective to each other. They may also be anchored to the edges and corners of a screen and specify input handling semantics. This interface should be suitable for the implementation of many desktop shell components, and a broad number of other applications that interact with the desktop.

type 'v t = ([ `Zwlr_layer_shell_v1 ], 'v, [ `Server ]) Wayland.Proxy.t

module Error = Wlr_layer_shell_unstable_v1_proto.Zwlr_layer_shell_v1.Error

module Layer = Wlr_layer_shell_unstable_v1_proto.Zwlr_layer_shell_v1.Layer

Version 1, 2

Version 3, 4, 5

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.

class virtual +'a v4 : object ... end

Handler for a proxy with version >= 4.

class virtual +'a v5 : object ... end

Handler for a proxy with version >= 5.