Class Zwlr_screencopy_frame_v1.v3

Handler for a proxy with version >= 3.

constraint 'a = [< `V3 ]

method private virtual on_buffer : [> `V3 ] t -> format:{Imports}1.Wl_shm.Format.t -> width:int32 -> height:int32 -> stride:int32 -> unit

Wl_shm buffer information.

Provides information about wl_shm buffer parameters that need to be used for this frame. This event is sent once after the frame is created if wl_shm buffers are supported.

method private virtual on_flags : [> `V3 ] t -> flags:{Imports}1.Zwlr_screencopy_frame_v1.Flags.t -> unit

Frame flags.

Provides flags about the frame. This event is sent once before the "ready" event.

method private virtual on_ready : [> `V3 ] t -> tv_sec_hi:int32 -> tv_sec_lo:int32 -> tv_nsec:int32 -> unit

Indicates frame is available for reading.

Called as soon as the frame is copied, indicating it is available for reading. This event includes the time at which presentation happened at.

The timestamp is expressed as tv_sec_hi, tv_sec_lo, tv_nsec triples, each component being an unsigned 32-bit value. Whole seconds are in tv_sec which is a 64-bit value combined from tv_sec_hi and tv_sec_lo, and the additional fractional part in tv_nsec as nanoseconds. Hence, for valid timestamps tv_nsec must be in 0, 999999999. The seconds part may have an arbitrary offset at start.

After receiving this event, the client should destroy the object.

method private virtual on_failed : [> `V3 ] t -> unit

Frame copy failed.

This event indicates that the attempted frame copy has failed.

After receiving this event, the client should destroy the object.

method private virtual on_damage : [> `V3 ] t -> x:int32 -> y:int32 -> width:int32 -> height:int32 -> unit

Carries the coordinates of the damaged region.

This event is sent right before the ready event when copy_with_damage is requested. It may be generated multiple times for each copy_with_damage request.

The arguments describe a box around an area that has changed since the last copy request that was derived from the current screencopy manager instance.

The union of all regions received between the call to copy_with_damage and a ready event is the total damage since the prior ready event.

method private virtual on_linux_dmabuf : [> `V3 ] t -> format:int32 -> width:int32 -> height:int32 -> unit

Linux-dmabuf buffer information.

Provides information about linux-dmabuf buffer parameters that need to be used for this frame. This event is sent once after the frame is created if linux-dmabuf buffers are supported.

method private virtual on_buffer_done : [> `V3 ] t -> unit

All buffer types reported.

This event is sent once after all buffer events have been sent.

The client should proceed to create a buffer of one of the supported types, and send a "copy" request.

method min_version : int32