Module Wlr_screencopy_unstable_v1_client.Zwlr_screencopy_manager_v1

Manager to inform clients and begin capturing.

This object is a manager which offers requests to start capturing from a source.

type 'v t = ([ `Zwlr_screencopy_manager_v1 ], 'v, [ `Client ]) Proxy.t

Version 1, 2, 3

val destroy : [< `V1 | `V2 | `V3 ] t -> unit

Destroy the manager.

All objects created by the manager will still remain valid, until their appropriate destroy request has been called.

val capture_output_region : [< `V1 | `V2 | `V3 ] as 'a t -> [[ `Zwlr_screencopy_frame_v1 ], 'a, [ `Client ]] Proxy.Handler.t -> overlay_cursor:int32 -> output:([ `Wl_output ], 'b, [ `Client ]) Proxy.t -> x:int32 -> y:int32 -> width:int32 -> height:int32 -> ([ `Zwlr_screencopy_frame_v1 ], 'a, [ `Client ]) Proxy.t

Capture an output's region.

Capture the next frame of an output's region.

The region is given in output logical coordinates, see xdg_output.logical_size. The region will be clipped to the output's extents.

val capture_output : [< `V1 | `V2 | `V3 ] as 'a t -> [[ `Zwlr_screencopy_frame_v1 ], 'a, [ `Client ]] Proxy.Handler.t -> overlay_cursor:int32 -> output:([ `Wl_output ], 'b, [ `Client ]) Proxy.t -> ([ `Zwlr_screencopy_frame_v1 ], 'a, [ `Client ]) Proxy.t

Capture an output.

Capture the next frame of an entire output.

class +'a _handlers_unsafe : object ... end

Requesting a minimum version

class v1 : object ... end

class v2 : object ... end

class v3 : object ... end