Class Zwlr_screencopy_frame_v1.handlers

method user_data : [Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_frame_v1.t, 'v, [ `Server ]] Proxy.Handler.t -> (Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_frame_v1.t, [ `Server ]) Wayland.S.user_data

method metadata : [Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_frame_v1.t, 'v, [ `Server ]] Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_frame_v1.t)

method private virtual on_copy : 'v t -> buffer:([ `Wl_buffer ], 'v, [ `Server ]) Proxy.t -> unit

Copy the frame.

Copy the frame to the supplied buffer. The buffer must have a the correct size, see zwlr_screencopy_frame_v1.buffer and zwlr_screencopy_frame_v1.linux_dmabuf. The buffer needs to have a supported format.

If the frame is successfully copied, a "flags" and a "ready" events are sent. Otherwise, a "failed" event is sent.

method private virtual on_destroy : 'v t -> unit

Delete this object, used or not.

Destroys the frame. This request can be sent at any time by the client.

method private virtual on_copy_with_damage : 'v t -> buffer:([ `Wl_buffer ], 'v, [ `Server ]) Proxy.t -> unit

Copy the frame when it's damaged.

Same as copy, except it waits until there is damage to copy.

method dispatch : [Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_frame_v1.t, 'v, [ `Server ]] Proxy.Handler.t -> (Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_frame_v1.t, 'v, [ `Server ]) Proxy.t -> (Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_frame_v1.t, [ `R ]) Wayland.Msg.t -> unit