Class Zwlr_screencopy_frame_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 | `V2 | `V3 ]

method private virtual on_copy : [> `V1 | `V2 | `V3 ] t -> buffer: ( [ `Wl_buffer ], [> Wayland__Wayland_proto.Wl_buffer.versions ], [ `Server ] ) Wayland.Proxy.t -> unit

Copy the frame.

Copy the frame to the supplied buffer. The buffer must have a the correct size, see zwlr_screencopy_frame_v1.buffer and zwlr_screencopy_frame_v1.linux_dmabuf. The buffer needs to have a supported format.

If the frame is successfully copied, a "flags" and a "ready" events are sent. Otherwise, a "failed" event is sent.

method private virtual on_destroy : [> `V1 | `V2 | `V3 ] t -> unit

Delete this object, used or not.

Destroys the frame. This request can be sent at any time by the client.

method private virtual on_copy_with_damage : [> `V2 | `V3 ] t -> buffer: ( [ `Wl_buffer ], [> Wayland__Wayland_proto.Wl_buffer.versions ], [ `Server ] ) Wayland.Proxy.t -> unit

Copy the frame when it's damaged.

Same as copy, except it waits until there is damage to copy.

method min_version : 'a v1 -> int32