Class Zwlr_screencopy_manager_v1.handlers

constraint 'b = [> `V1 | `V2 | `V3 ]

method user_data : [Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> (Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t, [ `Server ]) Wayland.S.user_data

method metadata : [Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t)

method min_version : [Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> int32

method max_version : [Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> int32

method private virtual on_capture_output : 'b t -> ([ `Zwlr_screencopy_frame_v1 ], 'b, [ `Server ]) Proxy.t -> overlay_cursor:int32 -> output:([ `Wl_output ], 'b, [ `Server ]) Proxy.t -> unit

Capture an output.

Capture the next frame of an entire output.

method private virtual on_capture_output_region : 'b t -> ([ `Zwlr_screencopy_frame_v1 ], 'b, [ `Server ]) Proxy.t -> overlay_cursor:int32 -> output:([ `Wl_output ], 'b, [ `Server ]) Proxy.t -> x:int32 -> y:int32 -> width:int32 -> height:int32 -> unit

Capture an output's region.

Capture the next frame of an output's region.

The region is given in output logical coordinates, see xdg_output.logical_size. The region will be clipped to the output's extents.

method private virtual on_destroy : 'b t -> unit

Destroy the manager.

All objects created by the manager will still remain valid, until their appropriate destroy request has been called.

method dispatch : [Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> (Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t, 'b, [ `Server ]) Proxy.t -> (Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t, [ `R ]) Wayland.Msg.t -> unit