Class Zwlr_screencopy_manager_v1.v3

Handler for a proxy with version >= 3.

constraint 'b = [< `V3 ]

method private virtual on_capture_output : [> `V3 ] t -> ([ `Zwlr_screencopy_frame_v1 ], 'b, [ `Server ]) Wayland.Proxy.t -> overlay_cursor:int32 -> output:([ `Wl_output ], [> Wayland__Wayland_proto.Wl_output.versions ], [ `Server ]) Wayland.Proxy.t -> unit

Capture an output.

Capture the next frame of an entire output.

method private virtual on_capture_output_region : [> `V3 ] t -> ([ `Zwlr_screencopy_frame_v1 ], 'b, [ `Server ]) Wayland.Proxy.t -> overlay_cursor:int32 -> output:([ `Wl_output ], [> Wayland__Wayland_proto.Wl_output.versions ], [ `Server ]) Wayland.Proxy.t -> x:int32 -> y:int32 -> width:int32 -> height:int32 -> unit

Capture an output's region.

Capture the next frame of an output's region.

The region is given in output logical coordinates, see xdg_output.logical_size. The region will be clipped to the output's extents.

method private virtual on_destroy : [> `V3 ] t -> unit

Destroy the manager.

All objects created by the manager will still remain valid, until their appropriate destroy request has been called.

method min_version : [Wayland_protocols.Wlr_screencopy_unstable_v1_proto.Zwlr_screencopy_manager_v1.t, 'b, [ `Server ]] {Proxy}2.Service_handler.t -> int32

method bind_version : 'c -> [ `V3 ]