Module Wp_primary_selection_unstable_v1_client.Zwp_primary_selection_device_manager_v1

X primary selection emulation.

The primary selection device manager is a singleton global object that provides access to the primary selection. It allows to create wp_primary_selection_source objects, as well as retrieving the per-seat wp_primary_selection_device objects.

type 'v t = ( [ `Zwp_primary_selection_device_manager_v1 ], 'v, [ `Client ] ) Wayland.Proxy.t

Version 1

val destroy : [< `V1 ] t -> unit

Destroy the primary selection device manager.

Destroy the primary selection device manager.

val get_device : [< `V1 ] as 'a t -> [ [ `Zwp_primary_selection_device_v1 ], 'b, [ `Client ] ] Wayland.Proxy.Handler.t -> seat:( [ `Wl_seat ], 'c, [ `Client ] ) Wayland.Proxy.t -> ( [ `Zwp_primary_selection_device_v1 ], 'd, [ `Client ] ) Wayland.Proxy.t

Create a new primary selection device.

Create a new data device for a given seat.

val create_source : [< `V1 ] as 'a t -> [ [ `Zwp_primary_selection_source_v1 ], 'b, [ `Client ] ] Wayland.Proxy.Handler.t -> ( [ `Zwp_primary_selection_source_v1 ], 'c, [ `Client ] ) Wayland.Proxy.t

Create a new primary selection source.

Create a new primary selection source.

Handlers

Note: Servers will always want to use v1.

class +'a v1 : object ... end

Handler for a proxy with version >= 1.