Module Wp_primary_selection_unstable_v1_client.Zwp_primary_selection_device_v1

type 'v t = ([ `Zwp_primary_selection_device_v1 ], 'v, [ `Client ]) Wayland.Proxy.t

Version 1

val destroy : [< `V1 ] t -> unit

Destroy the primary selection device.

Destroy the primary selection device.

val set_selection : [< `V1 ] t -> source: ([ `Zwp_primary_selection_source_v1 ], 'a, [ `Client ]) Wayland.Proxy.t option -> serial:int32 -> unit

Set the primary selection.

Replaces the current selection. The previous owner of the primary selection will receive a wp_primary_selection_source.cancelled event.

To unset the selection, set the source to NULL.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.