Class Zwp_primary_selection_source_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_send : [> `V1 ] t -> mime_type:string -> fd:Unix.file_descr -> unit

Send the primary selection contents.

Request for the current primary selection contents from the client. Send the specified mime type over the passed file descriptor, then close it.

method private virtual on_cancelled : [> `V1 ] t -> unit

Request for primary selection contents was canceled.

This primary selection source is no longer valid. The client should clean up and destroy this primary selection source.

method min_version : int32