Module Wp_primary_selection_unstable_v1_server.Zwp_primary_selection_device_v1

type 'v t = ([ `Zwp_primary_selection_device_v1 ], 'v, [ `Server ]) Wayland.Proxy.t

Version 1

val selection : [< `V1 ] t -> ([ `Zwp_primary_selection_offer_v1 ], 'a, [ `Server ]) Wayland.Proxy.t option -> unit

Advertise a new primary selection.

The wp_primary_selection_device.selection event is sent to notify the client of a new primary selection. This event is sent after the wp_primary_selection.data_offer event introducing this object, and after the offer has announced its mimetypes through wp_primary_selection_offer.offer.

The data_offer is valid until a new offer or NULL is received or until the client loses keyboard focus. The client must destroy the previous selection data_offer, if any, upon receiving this event.

val data_offer : [< `V1 ] as 'a t -> [[ `Zwp_primary_selection_offer_v1 ], 'b, [ `Server ]] Wayland.Proxy.Handler.t -> ([ `Zwp_primary_selection_offer_v1 ], 'c, [ `Server ]) Wayland.Proxy.t

Introduce a new wp_primary_selection_offer.

Introduces a new wp_primary_selection_offer object that may be used to receive the current primary selection. Immediately following this event, the new wp_primary_selection_offer object will send wp_primary_selection_offer.offer events to describe the offered mime types.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.