Class Zxdg_toplevel_decoration_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_configure : [> `V1 ] t -> mode:{Imports}1.Zxdg_toplevel_decoration_v1.Mode.t -> unit

Suggest a surface change.

The configure event asks the client to change its decoration mode. The configured state should not be applied immediately. Clients must send an ack_configure in response to this event. See xdg_surface.configure and xdg_surface.ack_configure for details.

A configure event can be sent at any time. The specified mode must be obeyed by the client.

method min_version : int32