Class Zxdg_decoration_manager_v1.handlers

constraint 'b = [> `V1 ]

method user_data : [Xdg_decoration_unstable_v1_proto.Zxdg_decoration_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> (Xdg_decoration_unstable_v1_proto.Zxdg_decoration_manager_v1.t, [ `Server ]) Wayland.S.user_data

method metadata : [Xdg_decoration_unstable_v1_proto.Zxdg_decoration_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Xdg_decoration_unstable_v1_proto.Zxdg_decoration_manager_v1.t)

method min_version : [Xdg_decoration_unstable_v1_proto.Zxdg_decoration_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> int32

method max_version : [Xdg_decoration_unstable_v1_proto.Zxdg_decoration_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> int32

method private virtual on_destroy : 'b t -> unit

Destroy the decoration manager object.

Destroy the decoration manager. This doesn't destroy objects created with the manager.

method private virtual on_get_toplevel_decoration : 'b t -> ([ `Zxdg_toplevel_decoration_v1 ], 'b, [ `Server ]) Proxy.t -> toplevel:([ `Xdg_toplevel ], 'b, [ `Server ]) Proxy.t -> unit

Create a new toplevel decoration object.

Create a new decoration object associated with the given toplevel.

Creating an xdg_toplevel_decoration from an xdg_toplevel which has a buffer attached or committed is a client error, and any attempts by a client to attach or manipulate a buffer prior to the first xdg_toplevel_decoration.configure event must also be treated as errors.

method dispatch : [Xdg_decoration_unstable_v1_proto.Zxdg_decoration_manager_v1.t, 'b, [ `Server ]] Proxy.Handler.t -> (Xdg_decoration_unstable_v1_proto.Zxdg_decoration_manager_v1.t, 'b, [ `Server ]) Proxy.t -> (Xdg_decoration_unstable_v1_proto.Zxdg_decoration_manager_v1.t, [ `R ]) Wayland.Msg.t -> unit