Module Xdg_output_unstable_v1_client.Zxdg_output_manager_v1

Manage xdg_output objects.

A global factory interface for xdg_output objects.

type 'v t = ([ `Zxdg_output_manager_v1 ], 'v, [ `Client ]) Proxy.t

Version 1, 2, 3

val get_xdg_output : [< `V1 | `V2 | `V3 ] as 'a t -> [[ `Zxdg_output_v1 ], 'a, [ `Client ]] Proxy.Handler.t -> output:([ `Wl_output ], 'b, [ `Client ]) Proxy.t -> ([ `Zxdg_output_v1 ], 'a, [ `Client ]) Proxy.t

Create an xdg output from a wl_output.

This creates a new xdg_output object for the given wl_output.

val destroy : [< `V1 | `V2 | `V3 ] t -> unit

Destroy the xdg_output_manager object.

Using this request a client can tell the server that it is not going to use the xdg_output_manager object anymore.

Any objects already created through this instance are not affected.

class +'a _handlers_unsafe : object ... end

Requesting a minimum version

class v1 : object ... end

class v2 : object ... end

class v3 : object ... end