Module Xdg_output_unstable_v1_client.Zxdg_output_v1

Compositor logical output region.

An xdg_output describes part of the compositor geometry.

This typically corresponds to a monitor that displays part of the compositor space.

For objects version 3 onwards, after all xdg_output properties have been sent (when the object is created and when properties are updated), a wl_output.done event is sent. This allows changes to the output properties to be seen as atomic, even if they happen via multiple events.

type 'v t = ([ `Zxdg_output_v1 ], 'v, [ `Client ]) $Proxy.t

Version 1

val destroy : [< `V1 | `V2 | `V3 ] t -> unit

Destroy the xdg_output object.

Using this request a client can tell the server that it is not going to use the xdg_output object anymore.

Version 2, 3

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.