Class Zxdg_output_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 | `V2 | `V3 ]

method private virtual on_logical_position : [> `V1 | `V2 | `V3 ] t -> x:int32 -> y:int32 -> unit

Position of the output within the global compositor space.

The position event describes the location of the wl_output within the global compositor space.

The logical_position event is sent after creating an xdg_output (see xdg_output_manager.get_xdg_output) and whenever the location of the output changes within the global compositor space.

method private virtual on_logical_size : [> `V1 | `V2 | `V3 ] t -> width:int32 -> height:int32 -> unit

Size of the output in the global compositor space.

The logical_size event describes the size of the output in the global compositor space.

For example, a surface without any buffer scale, transformation nor rotation set, with the size matching the logical_size will have the same size as the corresponding output when displayed.

Most regular Wayland clients should not pay attention to the logical size and would rather rely on xdg_shell interfaces.

Some clients such as Xwayland, however, need this to configure their surfaces in the global compositor space as the compositor may apply a different scale from what is advertised by the output scaling property (to achieve fractional scaling, for example).

For example, for a wl_output mode 3840×2160 and a scale factor 2:

A compositor not scaling the surface buffers will advertise a logical size of 3840×2160,

A compositor automatically scaling the surface buffers will advertise a logical size of 1920×1080,

A compositor using a fractional scale of 1.5 will advertise a logical size of 2560×1440.

For example, for a wl_output mode 1920×1080 and a 90 degree rotation, the compositor will advertise a logical size of 1080x1920.

The logical_size event is sent after creating an xdg_output (see xdg_output_manager.get_xdg_output) and whenever the logical size of the output changes, either as a result of a change in the applied scale or because of a change in the corresponding output mode(see wl_output.mode) or transform (see wl_output.transform).

method private virtual on_done : [> `V1 | `V2 | `V3 ] t -> unit

All information about the output have been sent.

This event is sent after all other properties of an xdg_output have been sent.

This allows changes to the xdg_output properties to be seen as atomic, even if they happen via multiple events.

For objects version 3 onwards, this event is deprecated. Compositors are not required to send it anymore and must send wl_output.done instead.

method private virtual on_name : [> `V2 | `V3 ] t -> name:string -> unit

Name of this output.

Many compositors will assign names to their outputs, show them to the user, allow them to be configured by name, etc. The client may wish to know this name as well to offer the user similar behaviors.

The naming convention is compositor defined, but limited to alphanumeric characters and dashes (-). Each name is unique among all wl_output globals, but if a wl_output global is destroyed the same name may be reused later. The names will also remain consistent across sessions with the same hardware and software configuration.

Examples of names include 'HDMI-A-1', 'WL-1', 'X11-1', etc. However, do not assume that the name is a reflection of an underlying DRM connector, X11 connection, etc.

The name event is sent after creating an xdg_output (see xdg_output_manager.get_xdg_output). This event is only sent once per xdg_output, and the name does not change over the lifetime of the wl_output global.

method private virtual on_description : [> `V2 | `V3 ] t -> description:string -> unit

Human-readable description of this output.

Many compositors can produce human-readable descriptions of their outputs. The client may wish to know this description as well, to communicate the user for various purposes.

The description is a UTF-8 string with no convention defined for its contents. Examples might include 'Foocorp 11" Display' or 'Virtual X11 output via :1'.

The description event is sent after creating an xdg_output (see xdg_output_manager.get_xdg_output) and whenever the description changes. The description is optional, and may not be sent at all.

For objects of version 2 and lower, this event is only sent once per xdg_output, and the description does not change over the lifetime of the wl_output global. "

method min_version : 'a v1 -> int32