Class Zxdg_output_manager_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 | `V2 | `V3 ]

method private virtual on_destroy : [> `V1 | `V2 | `V3 ] t -> unit

Destroy the xdg_output_manager object.

Using this request a client can tell the server that it is not going to use the xdg_output_manager object anymore.

Any objects already created through this instance are not affected.

method private virtual on_get_xdg_output : [> `V1 | `V2 | `V3 ] t -> ([ `Zxdg_output_v1 ], 'b, [ `Server ]) {Proxy}2.t -> output: ([ `Wl_output ], [> {Imports}1.Wl_output.versions ], [ `Server ]) {Proxy}2.t -> unit

Create an xdg output from a wl_output.

This creates a new xdg_output object for the given wl_output.

method min_version : int32

method bind_version : [ `V1 ]