Class Zxdg_output_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 | `V2 | `V3 ]

method private virtual on_destroy : [> `V1 | `V2 | `V3 ] t -> unit

Destroy the xdg_output object.

Using this request a client can tell the server that it is not going to use the xdg_output object anymore.

method min_version : 'a v1 -> int32