Module Xdg_shell_client.Xdg_positioner

Child surface positioner.

The xdg_positioner provides a collection of rules for the placement of a child surface relative to a parent surface. Rules can be defined to ensure the child surface remains within the visible area's borders, and to specify how the child surface changes its position, such as sliding along an axis, or flipping around a rectangle. These positioner-created rules are constrained by the requirement that a child surface must intersect with or be at least partially adjacent to its parent surface.

See the various requests for details about possible rules.

At the time of the request, the compositor makes a copy of the rules specified by the xdg_positioner. Thus, after the request is complete the xdg_positioner object can be destroyed or reused; further changes to the object will have no effect on previous usages.

For an xdg_positioner object to be considered complete, it must have a non-zero size set by set_size, and a non-zero anchor rectangle set by set_anchor_rect. Passing an incomplete xdg_positioner object when positioning a surface raises an error.

type 'v t = ( [ `Xdg_positioner ], 'v, [ `Client ] ) Wayland.Proxy.t

module Error = Xdg_shell_proto.Xdg_positioner.Error

module Anchor = Xdg_shell_proto.Xdg_positioner.Anchor

module Gravity = Xdg_shell_proto.Xdg_positioner.Gravity

module Constraint_adjustment = Xdg_shell_proto.Xdg_positioner.Constraint_adjustment

Version 1, 2

val set_offset : [< `V1 | `V2 | `V3 | `V4 | `V5 ] t -> x:int32 -> y:int32 -> unit

Set surface position offset.

Specify the surface position offset relative to the position of the anchor on the anchor rectangle and the anchor on the surface. For example if the anchor of the anchor rectangle is at (x, y), the surface has the gravity bottom|right, and the offset is (ox, oy), the calculated surface position will be (x + ox, y + oy). The offset position of the surface is the one used for constraint testing. See set_constraint_adjustment.

An example use case is placing a popup menu on top of a user interface element, while aligning the user interface element of the parent surface with some user interface element placed somewhere in the popup surface.

val set_constraint_adjustment : [< `V1 | `V2 | `V3 | `V4 | `V5 ] t -> constraint_adjustment:int32 -> unit

Set the adjustment to be done when constrained.

Specify how the window should be positioned if the originally intended position caused the surface to be constrained, meaning at least partially outside positioning boundaries set by the compositor. The adjustment is set by constructing a bitmask describing the adjustment to be made when the surface is constrained on that axis.

If no bit for one axis is set, the compositor will assume that the child surface should not change its position on that axis when constrained.

If more than one bit for one axis is set, the order of how adjustments are applied is specified in the corresponding adjustment descriptions.

The default adjustment is none.

val set_gravity : [< `V1 | `V2 | `V3 | `V4 | `V5 ] t -> gravity:Wayland_protocols__Xdg_shell_proto.Xdg_positioner.Gravity.t -> unit

Set child surface gravity.

Defines in what direction a surface should be positioned, relative to the anchor point of the parent surface. If a corner gravity is specified (e.g. 'bottom_right' or 'top_left'), then the child surface will be placed towards the specified gravity; otherwise, the child surface will be centered over the anchor point on any axis that had no gravity specified.

val set_anchor : [< `V1 | `V2 | `V3 | `V4 | `V5 ] t -> anchor:Wayland_protocols__Xdg_shell_proto.Xdg_positioner.Anchor.t -> unit

Set anchor rectangle anchor.

Defines the anchor point for the anchor rectangle. The specified anchor is used derive an anchor point that the child surface will be positioned relative to. If a corner anchor is set (e.g. 'top_left' or 'bottom_right'), the anchor point will be at the specified corner; otherwise, the derived anchor point will be centered on the specified edge, or in the center of the anchor rectangle if no edge is specified.

val set_anchor_rect : [< `V1 | `V2 | `V3 | `V4 | `V5 ] t -> x:int32 -> y:int32 -> width:int32 -> height:int32 -> unit

Set the anchor rectangle within the parent surface.

Specify the anchor rectangle within the parent surface that the child surface will be placed relative to. The rectangle is relative to the window geometry as defined by xdg_surface.set_window_geometry of the parent surface.

When the xdg_positioner object is used to position a child surface, the anchor rectangle may not extend outside the window geometry of the positioned child's parent surface.

If a negative size is set the invalid_input error is raised.

val set_size : [< `V1 | `V2 | `V3 | `V4 | `V5 ] t -> width:int32 -> height:int32 -> unit

Set the size of the to-be positioned rectangle.

Set the size of the surface that is to be positioned with the positioner object. The size is in surface-local coordinates and corresponds to the window geometry. See xdg_surface.set_window_geometry.

If a zero or negative size is set the invalid_input error is raised.

val destroy : [< `V1 | `V2 | `V3 | `V4 | `V5 ] t -> unit

Destroy the xdg_positioner object.

Notify the compositor that the xdg_positioner will no longer be used.

Version 3, 4, 5

val set_parent_configure : [< `V3 | `V4 | `V5 ] t -> serial:int32 -> unit

Set parent configure this is a response to.

Set the serial of an xdg_surface.configure event this positioner will be used in response to. The compositor may use this information together with set_parent_size to determine what future state the popup should be constrained using.

val set_parent_size : [< `V3 | `V4 | `V5 ] t -> parent_width:int32 -> parent_height:int32 -> unit

.

Set the parent window geometry the compositor should use when positioning the popup. The compositor may use this information to determine the future state the popup should be constrained using. If this doesn't match the dimension of the parent the popup is eventually positioned against, the behavior is undefined.

The arguments are given in the surface-local coordinate space.

val set_reactive : [< `V3 | `V4 | `V5 ] t -> unit

Continuously reconstrain the surface.

When set reactive, the surface is reconstrained if the conditions used for constraining changed, e.g. the parent window moved.

If the conditions changed and the popup was reconstrained, an xdg_popup.configure event is sent with updated geometry, followed by an xdg_surface.configure event.

Handlers

Note: Servers will always want to use v1.

class +'a v1 : object ... end

Handler for a proxy with version >= 1.

class +'a v2 : object ... end

Handler for a proxy with version >= 2.

class +'a v3 : object ... end

Handler for a proxy with version >= 3.

class +'a v4 : object ... end

Handler for a proxy with version >= 4.

class +'a v5 : object ... end

Handler for a proxy with version >= 5.