Class Xdg_surface.v2

Handler for a proxy with version >= 2.

constraint 'b = [< `V2 | `V3 ]

method private virtual on_configure : [> `V2 | `V3 ] t -> serial:int32 -> unit

Suggest a surface change.

The configure event marks the end of a configure sequence. A configure sequence is a set of one or more events configuring the state of the xdg_surface, including the final xdg_surface.configure event.

Where applicable, xdg_surface surface roles will during a configure sequence extend this event as a latched state sent as events before the xdg_surface.configure event. Such events should be considered to make up a set of atomically applied configuration states, where the xdg_surface.configure commits the accumulated state.

Clients should arrange their surface for the new states, and then send an ack_configure request with the serial sent in this configure event at some point before committing the new surface.

If the client receives multiple configure events before it can respond to one, it is free to discard all but the last event it received.

method min_version : [Wayland_protocols.Xdg_shell_proto.Xdg_surface.t, 'b, [ `Client ]] {Proxy}2.Service_handler.t -> int32