Class Xdg_wm_base.v3

Handler for a proxy with version >= 3.

constraint 'b = [< `V3 ]

method private virtual on_ping : [> `V3 ] t -> serial:int32 -> unit

Check if the client is alive.

The ping event asks the client if it's still alive. Pass the serial specified in the event back to the compositor by sending a "pong" request back with the specified serial. See xdg_wm_base.pong.

Compositors can use this to determine if the client is still alive. It's unspecified what will happen if the client doesn't respond to the ping request, or in what timeframe. Clients should try to respond in a reasonable amount of time.

A compositor is free to ping in any way it wants, but a client must always respond to any xdg_wm_base object it created.

method min_version : [Wayland_protocols.Xdg_shell_proto.Xdg_wm_base.t, 'b, [ `Client ]] {Proxy}2.Service_handler.t -> int32

method bind_version : 'c -> [ `V3 ]