Module Xdg_toplevel.Resize_edge

Edge values for resizing.

These values are used to indicate which edge of a surface is being dragged in a resize operation.

type t =

| None : t

| Top : t

| Bottom : t

| Left : t

| Top_left : t

| Bottom_left : t

| Right : t

| Top_right : t

| Bottom_right : t

val to_int32 : t -> int32

val of_int32 : int32 -> t