Class Xdg_positioner.v6

Handler for a proxy with version >= 6.

constraint 'a = [< `V6 ]

method private virtual on_destroy : [> `V6 ] t -> unit

Destroy the xdg_positioner object.

Notify the compositor that the xdg_positioner will no longer be used.

method private virtual on_set_size : [> `V6 ] t -> width:int32 -> height:int32 -> unit

Set the size of the to-be positioned rectangle.

Set the size of the surface that is to be positioned with the positioner object. The size is in surface-local coordinates and corresponds to the window geometry. See xdg_surface.set_window_geometry.

If a zero or negative size is set the invalid_input error is raised.

method private virtual on_set_anchor_rect : [> `V6 ] t -> x:int32 -> y:int32 -> width:int32 -> height:int32 -> unit

Set the anchor rectangle within the parent surface.

Specify the anchor rectangle within the parent surface that the child surface will be placed relative to. The rectangle is relative to the window geometry as defined by xdg_surface.set_window_geometry of the parent surface.

When the xdg_positioner object is used to position a child surface, the anchor rectangle may not extend outside the window geometry of the positioned child's parent surface.

If a negative size is set the invalid_input error is raised.

method private virtual on_set_anchor : [> `V6 ] t -> anchor:{Imports}1.Xdg_positioner.Anchor.t -> unit

Set anchor rectangle anchor.

Defines the anchor point for the anchor rectangle. The specified anchor is used derive an anchor point that the child surface will be positioned relative to. If a corner anchor is set (e.g. 'top_left' or 'bottom_right'), the anchor point will be at the specified corner; otherwise, the derived anchor point will be centered on the specified edge, or in the center of the anchor rectangle if no edge is specified.

method private virtual on_set_gravity : [> `V6 ] t -> gravity:{Imports}1.Xdg_positioner.Gravity.t -> unit

Set child surface gravity.

Defines in what direction a surface should be positioned, relative to the anchor point of the parent surface. If a corner gravity is specified (e.g. 'bottom_right' or 'top_left'), then the child surface will be placed towards the specified gravity; otherwise, the child surface will be centered over the anchor point on any axis that had no gravity specified. If the gravity is not in the ‘gravity’ enum, an invalid_input error is raised.

method private virtual on_set_constraint_adjustment : [> `V6 ] t -> constraint_adjustment:int32 -> unit

Set the adjustment to be done when constrained.

Specify how the window should be positioned if the originally intended position caused the surface to be constrained, meaning at least partially outside positioning boundaries set by the compositor. The adjustment is set by constructing a bitmask describing the adjustment to be made when the surface is constrained on that axis.

If no bit for one axis is set, the compositor will assume that the child surface should not change its position on that axis when constrained.

If more than one bit for one axis is set, the order of how adjustments are applied is specified in the corresponding adjustment descriptions.

The default adjustment is none.

method private virtual on_set_offset : [> `V6 ] t -> x:int32 -> y:int32 -> unit

Set surface position offset.

Specify the surface position offset relative to the position of the anchor on the anchor rectangle and the anchor on the surface. For example if the anchor of the anchor rectangle is at (x, y), the surface has the gravity bottom|right, and the offset is (ox, oy), the calculated surface position will be (x + ox, y + oy). The offset position of the surface is the one used for constraint testing. See set_constraint_adjustment.

An example use case is placing a popup menu on top of a user interface element, while aligning the user interface element of the parent surface with some user interface element placed somewhere in the popup surface.

method private virtual on_set_reactive : [> `V6 ] t -> unit

Continuously reconstrain the surface.

When set reactive, the surface is reconstrained if the conditions used for constraining changed, e.g. the parent window moved.

If the conditions changed and the popup was reconstrained, an xdg_popup.configure event is sent with updated geometry, followed by an xdg_surface.configure event.

method private virtual on_set_parent_size : [> `V6 ] t -> parent_width:int32 -> parent_height:int32 -> unit

.

Set the parent window geometry the compositor should use when positioning the popup. The compositor may use this information to determine the future state the popup should be constrained using. If this doesn't match the dimension of the parent the popup is eventually positioned against, the behavior is undefined.

The arguments are given in the surface-local coordinate space.

method private virtual on_set_parent_configure : [> `V6 ] t -> serial:int32 -> unit

Set parent configure this is a response to.

Set the serial of an xdg_surface.configure event this positioner will be used in response to. The compositor may use this information together with set_parent_size to determine what future state the popup should be constrained using.

method min_version : int32