Class Xdg_surface.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 | `V2 | `V3 | `V4 | `V5 | `V6 ]

method private virtual on_destroy : [> `V1 | `V2 | `V3 | `V4 | `V5 | `V6 ] t -> unit

Destroy the xdg_surface.

Destroy the xdg_surface object. An xdg_surface must only be destroyed after its role object has been destroyed, otherwise a defunct_role_object error is raised.

method private virtual on_get_toplevel : [> `V1 | `V2 | `V3 | `V4 | `V5 | `V6 ] t -> ([ `Xdg_toplevel ], 'b, [ `Server ]) {Proxy}2.t -> unit

Assign the xdg_toplevel surface role.

This creates an xdg_toplevel object for the given xdg_surface and gives the associated wl_surface the xdg_toplevel role.

See the documentation of xdg_toplevel for more details about what an xdg_toplevel is and how it is used.

method private virtual on_get_popup : [> `V1 | `V2 | `V3 | `V4 | `V5 | `V6 ] t -> ([ `Xdg_popup ], 'b, [ `Server ]) {Proxy}2.t -> parent: ([ `Xdg_surface ], [> {Imports}1.Xdg_surface.versions ], [ `Server ]) {Proxy}2.t option -> positioner: ([ `Xdg_positioner ], [> {Imports}1.Xdg_positioner.versions ], [ `Server ]) {Proxy}2.t -> unit

Assign the xdg_popup surface role.

This creates an xdg_popup object for the given xdg_surface and gives the associated wl_surface the xdg_popup role.

If null is passed as a parent, a parent surface must be specified using some other protocol, before committing the initial state.

See the documentation of xdg_popup for more details about what an xdg_popup is and how it is used.

method private virtual on_set_window_geometry : [> `V1 | `V2 | `V3 | `V4 | `V5 | `V6 ] t -> x:int32 -> y:int32 -> width:int32 -> height:int32 -> unit

Set the new window geometry.

The window geometry of a surface is its "visible bounds" from the user's perspective. Client-side decorations often have invisible portions like drop-shadows which should be ignored for the purposes of aligning, placing and constraining windows.

The window geometry is double buffered, and will be applied at the time wl_surface.commit of the corresponding wl_surface is called.

When maintaining a position, the compositor should treat the (x, y) coordinate of the window geometry as the top left corner of the window. A client changing the (x, y) window geometry coordinate should in general not alter the position of the window.

Once the window geometry of the surface is set, it is not possible to unset it, and it will remain the same until set_window_geometry is called again, even if a new subsurface or buffer is attached.

If never set, the value is the full bounds of the surface, including any subsurfaces. This updates dynamically on every commit. This unset is meant for extremely simple clients.

The arguments are given in the surface-local coordinate space of the wl_surface associated with this xdg_surface, and may extend outside of the wl_surface itself to mark parts of the subsurface tree as part of the window geometry.

When applied, the effective window geometry will be the set window geometry clamped to the bounding rectangle of the combined geometry of the surface of the xdg_surface and the associated subsurfaces.

The effective geometry will not be recalculated unless a new call to set_window_geometry is done and the new pending surface state is subsequently applied.

The width and height of the effective window geometry must be greater than zero. Setting an invalid size will raise an invalid_size error.

method private virtual on_ack_configure : [> `V1 | `V2 | `V3 | `V4 | `V5 | `V6 ] t -> serial:int32 -> unit

Ack a configure event.

When a configure event is received, if a client commits the surface in response to the configure event, then the client must make an ack_configure request sometime before the commit request, passing along the serial of the configure event.

For instance, for toplevel surfaces the compositor might use this information to move a surface to the top left only when the client has drawn itself for the maximized or fullscreen state.

If the client receives multiple configure events before it can respond to one, it only has to ack the last configure event. Acking a configure event that was never sent raises an invalid_serial error.

A client is not required to commit immediately after sending an ack_configure request - it may even ack_configure several times before its next surface commit.

A client may send multiple ack_configure requests before committing, but only the last request sent before a commit indicates which configure event the client really is responding to.

Sending an ack_configure request consumes the serial number sent with the request, as well as serial numbers sent by all configure events sent on this xdg_surface prior to the configure event referenced by the committed serial.

It is an error to issue multiple ack_configure requests referencing a serial from the same configure event, or to issue an ack_configure request referencing a serial from a configure event issued before the event identified by the last ack_configure request for the same xdg_surface. Doing so will raise an invalid_serial error.

method min_version : int32