Class Xdg_wm_base.h1

method virtual on_pong : 'v t -> serial:int32 -> unit

Respond to a ping event.

A client must respond to a ping event with a pong request or the client may be deemed unresponsive. See xdg_wm_base.ping.

method virtual on_get_xdg_surface : 'v t -> int32 -> surface:int32 -> unit

Create a shell surface from a surface.

This creates an xdg_surface for the given surface. While xdg_surface itself is not a role, the corresponding surface may only be assigned a role extending xdg_surface, such as xdg_toplevel or xdg_popup.

This creates an xdg_surface for the given surface. An xdg_surface is used as basis to define a role to a given surface, such as xdg_toplevel or xdg_popup. It also manages functionality shared between xdg_surface based surface roles.

See the documentation of xdg_surface for more details about what an xdg_surface is and how it is used.

method virtual on_create_positioner : 'v t -> int32 -> unit

Create a positioner object.

Create a positioner object. A positioner object is used to position surfaces relative to some parent surface. See the interface description and xdg_surface.get_popup for details.

method virtual on_destroy : 'v t -> unit

Destroy xdg_wm_base.

Destroy this xdg_wm_base object.

Destroying a bound xdg_wm_base object while there are surfaces still alive created by this xdg_wm_base object instance is illegal and will result in a protocol error.