Class Gtk_primary_selection_device.handlers

method user_data : [Wayland_protocols.Gtk_primary_selection_proto.Gtk_primary_selection_device.t, 'v, [ `Client ]] Proxy.Handler.t -> (Wayland_protocols.Gtk_primary_selection_proto.Gtk_primary_selection_device.t, [ `Client ]) Wayland.S.user_data

method metadata : [Wayland_protocols.Gtk_primary_selection_proto.Gtk_primary_selection_device.t, 'v, [ `Client ]] Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wayland_protocols.Gtk_primary_selection_proto.Gtk_primary_selection_device.t)

method private virtual on_data_offer : 'v t -> ([ `Gtk_primary_selection_offer ], 'v, [ `Client ]) Proxy.t -> unit

Introduce a new wp_primary_selection_offer.

Introduces a new wp_primary_selection_offer object that may be used to receive the current primary selection. Immediately following this event, the new wp_primary_selection_offer object will send wp_primary_selection_offer.offer events to describe the offered mime types.

method private virtual on_selection : 'v t -> ([ `Gtk_primary_selection_offer ], 'v, [ `Client ]) Proxy.t option -> unit

Advertise a new primary selection.

The wp_primary_selection_device.selection event is sent to notify the client of a new primary selection. This event is sent after the wp_primary_selection.data_offer event introducing this object, and after the offer has announced its mimetypes through wp_primary_selection_offer.offer.

The data_offer is valid until a new offer or NULL is received or until the client loses keyboard focus. The client must destroy the previous selection data_offer, if any, upon receiving this event.

method dispatch : [Wayland_protocols.Gtk_primary_selection_proto.Gtk_primary_selection_device.t, 'v, [ `Client ]] Proxy.Handler.t -> (Wayland_protocols.Gtk_primary_selection_proto.Gtk_primary_selection_device.t, 'v, [ `Client ]) Proxy.t -> (Wayland_protocols.Gtk_primary_selection_proto.Gtk_primary_selection_device.t, [ `R ]) Wayland.Msg.t -> unit