Module Wayland_protocols__Gtk_primary_selection_client.Gtk_primary_selection_device_manager

X primary selection emulation.

The primary selection device manager is a singleton global object that provides access to the primary selection. It allows to create wp_primary_selection_source objects, as well as retrieving the per-seat wp_primary_selection_device objects.

type 'v t = ([ `Gtk_primary_selection_device_manager ], 'v, [ `Client ]) Proxy.t

Version 1

val destroy : [< `V1 ] t -> unit

Destroy the primary selection device manager.

Destroy the primary selection device manager.

val get_device : [< `V1 ] as 'a t -> [[ `Gtk_primary_selection_device ], 'a, [ `Client ]] Proxy.Handler.t -> seat:([ `Wl_seat ], 'b, [ `Client ]) Proxy.t -> ([ `Gtk_primary_selection_device ], 'a, [ `Client ]) Proxy.t

Create a new primary selection device.

Create a new data device for a given seat.

val create_source : [< `V1 ] as 'a t -> [[ `Gtk_primary_selection_source ], 'a, [ `Client ]] Proxy.Handler.t -> ([ `Gtk_primary_selection_source ], 'a, [ `Client ]) Proxy.t

Create a new primary selection source.

Create a new primary selection source.

class +'a _handlers_unsafe : object ... end

Requesting a minimum version

class v1 : object ... end