Class Xdg_wm_base._handlers_unsafe

method user_data : [Wayland_protocols.Xdg_shell_proto.Xdg_wm_base.t, 'v, [ `Client ]] Proxy.Handler.t -> (Wayland_protocols.Xdg_shell_proto.Xdg_wm_base.t, [ `Client ]) Wayland.S.user_data

method metadata : [Wayland_protocols.Xdg_shell_proto.Xdg_wm_base.t, 'v, [ `Client ]] Proxy.Handler.t -> (module Wayland.Metadata.S with type t = Wayland_protocols.Xdg_shell_proto.Xdg_wm_base.t)

method private virtual on_ping : 'v t -> serial:int32 -> unit

Check if the client is alive.

The ping event asks the client if it's still alive. Pass the serial specified in the event back to the compositor by sending a "pong" request back with the specified serial. See xdg_wm_base.pong.

Compositors can use this to determine if the client is still alive. It's unspecified what will happen if the client doesn't respond to the ping request, or in what timeframe. Clients should try to respond in a reasonable amount of time.

A compositor is free to ping in any way it wants, but a client must always respond to any xdg_wm_base object it created.

method dispatch : [Wayland_protocols.Xdg_shell_proto.Xdg_wm_base.t, 'v, [ `Client ]] Proxy.Handler.t -> (Wayland_protocols.Xdg_shell_proto.Xdg_wm_base.t, 'v, [ `Client ]) Proxy.t -> (Wayland_protocols.Xdg_shell_proto.Xdg_wm_base.t, [ `R ]) Wayland.Msg.t -> unit