Class Xdg_positioner.h1

method virtual on_set_offset : 'v t -> x:int32 -> y:int32 -> unit

Set surface position offset.

Specify the surface position offset relative to the position of the anchor on the anchor rectangle and the anchor on the surface. For example if the anchor of the anchor rectangle is at (x, y), the surface has the gravity bottom|right, and the offset is (ox, oy), the calculated surface position will be (x + ox, y + oy). The offset position of the surface is the one used for constraint testing. See set_constraint_adjustment.

An example use case is placing a popup menu on top of a user interface element, while aligning the user interface element of the parent surface with some user interface element placed somewhere in the popup surface.

method virtual on_set_constraint_adjustment : 'v t -> constraint_adjustment:int32 -> unit

Set the adjustment to be done when constrained.

Specify how the window should be positioned if the originally intended position caused the surface to be constrained, meaning at least partially outside positioning boundaries set by the compositor. The adjustment is set by constructing a bitmask describing the adjustment to be made when the surface is constrained on that axis.

If no bit for one axis is set, the compositor will assume that the child surface should not change its position on that axis when constrained.

If more than one bit for one axis is set, the order of how adjustments are applied is specified in the corresponding adjustment descriptions.

The default adjustment is none.

method virtual on_set_gravity : 'v t -> gravity:int32 -> unit

Set child surface gravity.

Defines in what direction a surface should be positioned, relative to the anchor point of the parent surface. If a corner gravity is specified (e.g. 'bottom_right' or 'top_left'), then the child surface will be placed towards the specified gravity; otherwise, the child surface will be centered over the anchor point on any axis that had no gravity specified.

method virtual on_set_anchor : 'v t -> anchor:int32 -> unit

Set anchor rectangle anchor.

Defines the anchor point for the anchor rectangle. The specified anchor is used derive an anchor point that the child surface will be positioned relative to. If a corner anchor is set (e.g. 'top_left' or 'bottom_right'), the anchor point will be at the specified corner; otherwise, the derived anchor point will be centered on the specified edge, or in the center of the anchor rectangle if no edge is specified.

method virtual on_set_anchor_rect : 'v t -> x:int32 -> y:int32 -> width:int32 -> height:int32 -> unit

Set the anchor rectangle within the parent surface.

Specify the anchor rectangle within the parent surface that the child surface will be placed relative to. The rectangle is relative to the window geometry as defined by xdg_surface.set_window_geometry of the parent surface.

When the xdg_positioner object is used to position a child surface, the anchor rectangle may not extend outside the window geometry of the positioned child's parent surface.

If a negative size is set the invalid_input error is raised.

method virtual on_set_size : 'v t -> width:int32 -> height:int32 -> unit

Set the size of the to-be positioned rectangle.

Set the size of the surface that is to be positioned with the positioner object. The size is in surface-local coordinates and corresponds to the window geometry. See xdg_surface.set_window_geometry.

If a zero or negative size is set the invalid_input error is raised.

method virtual on_destroy : 'v t -> unit

Destroy the xdg_positioner object.

Notify the compositor that the xdg_positioner will no longer be used.