Class Xdg_positioner.h3

inherit h1

method virtual on_set_parent_configure : 'v t -> serial:int32 -> unit

Set parent configure this is a response to.

Set the serial of an xdg_surface.configure event this positioner will be used in response to. The compositor may use this information together with set_parent_size to determine what future state the popup should be constrained using.

method virtual on_set_parent_size : 'v t -> parent_width:int32 -> parent_height:int32 -> unit

.

Set the parent window geometry the compositor should use when positioning the popup. The compositor may use this information to determine the future state the popup should be constrained using. If this doesn't match the dimension of the parent the popup is eventually positioned against, the behavior is undefined.

The arguments are given in the surface-local coordinate space.

method virtual on_set_reactive : 'v t -> unit

Continuously reconstrain the surface.

When set reactive, the surface is reconstrained if the conditions used for constraining changed, e.g. the parent window moved.

If the conditions changed and the popup was reconstrained, an xdg_popup.configure event is sent with updated geometry, followed by an xdg_surface.configure event.