Module Wayland

module Unix_transport : sig ... end

Send messages over a Unix-domain socket.

module Client : sig ... end

Connect to a Wayland server and interact with the initial display object.

module Registry : sig ... end

A simple wrapper for the Wayland_client.Wl_registry API.

val callback : (int32 -> unit) -> < dispatch : (Wayland_proto.Wl_callback.t, [< `V1 | `V2 | `V3 | `V4 ], [ `Client ]) Wayland__Proxy.t -> (Wayland_proto.Wl_callback.t, [ `R ]) Msg.t -> unit; max_version : int32; metadata : (module Metadata.S with type t = Wayland_proto.Wl_callback.t); min_version : int32; on_done : callback_data:int32 -> unit; user_data : (Wayland_proto.Wl_callback.t, [ `Client ]) S.user_data; >

callback fn is a simple wrapper for the Wayland_client.Wl_callback API. It calls fn data when the callback's "done" signal is received. Wl_callback seems to be an exception to the usual Wayland versioning rules (a wl_callback can be created by multiple objects).

module Server : sig ... end

Code for writing Wayland servers.

module Fixed : sig ... end

Wayland's 24.8 fixed-point type.

val src : Logs.src

Control the library's log level.

Generated bindings for the core Wayland protocol

module Wayland_client : sig ... end

Bindings for use by Wayland clients.

module Wayland_server : sig ... end

Bindings for use by Wayland servers.

module Wayland_proto : sig ... end

Type metadata for use by bindings.

Low-level API

module Proxy : sig ... end

Low-level operations on objects, mostly used by the generated wrappers.

module Metadata : sig ... end

Describing interfaces.

module Msg : sig ... end

Building and parsing messages.

module S : sig ... end

Type signatures.

module Iface_reg : sig ... end

Registry of known interfaces. Used by the generated code.