Class type Service_handler.t

An ('a, 'v, 'role) t handles incoming messages for a service object of type 'a.

This type is used when binding a service, to choose the version. All other handlers get the version from elsewhere (inherited from their parent object).

inherit ['a, 'v, 'role] Handler.t

method min_version : int32

method max_version : int32