Module Wayland_client.Wl_shell_surface

Desktop-style metadata interface.

An interface that may be implemented by a wl_surface, for implementations that provide a desktop-style user interface.

It provides requests to treat surfaces like toplevel, fullscreen or popup windows, move, resize or maximize them, associate metadata like title and class, etc.

On the server side the object is automatically destroyed when the related wl_surface is destroyed. On the client side, wl_shell_surface_destroy() must be called before destroying the wl_surface object.

type 'v t = ( [ `Wl_shell_surface ], 'v, [ `Client ] ) Proxy.t

module Resize = Wayland_proto.Wl_shell_surface.Resize

module Transient = Wayland_proto.Wl_shell_surface.Transient

module Fullscreen_method = Wayland_proto.Wl_shell_surface.Fullscreen_method

Version 1

val set_class : [< `V1 ] t -> class_:string -> unit

Set surface class.

Set a class for the surface.

The surface class identifies the general class of applications to which the surface belongs. A common convention is to use the file name (or the full path if it is a non-standard location) of the application's .desktop file as the class.

val set_title : [< `V1 ] t -> title:string -> unit

Set surface title.

Set a short title for the surface.

This string may be used to identify the surface in a task bar, window list, or other user interface elements provided by the compositor.

The string must be encoded in UTF-8.

val set_maximized : [< `V1 ] t -> output:( [ `Wl_output ], 'a, [ `Client ] ) Proxy.t option -> unit

Make the surface a maximized surface.

Map the surface as a maximized surface.

If an output parameter is given then the surface will be maximized on that output. If the client does not specify the output then the compositor will apply its policy - usually choosing the output on which the surface has the biggest surface area.

The compositor will reply with a configure event telling the expected new surface size. The operation is completed on the next buffer attach to this surface.

A maximized surface typically fills the entire output it is bound to, except for desktop elements such as panels. This is the main difference between a maximized shell surface and a fullscreen shell surface.

The details depend on the compositor implementation.

val set_popup : [< `V1 ] t -> seat:( [ `Wl_seat ], 'a, [ `Client ] ) Proxy.t -> serial:int32 -> parent:( [ `Wl_surface ], 'b, [ `Client ] ) Proxy.t -> x:int32 -> y:int32 -> flags:int32 -> unit

Make the surface a popup surface.

Map the surface as a popup.

A popup surface is a transient surface with an added pointer grab.

An existing implicit grab will be changed to owner-events mode, and the popup grab will continue after the implicit grab ends (i.e. releasing the mouse button does not cause the popup to be unmapped).

The popup grab continues until the window is destroyed or a mouse button is pressed in any other client's window. A click in any of the client's surfaces is reported as normal, however, clicks in other clients' surfaces will be discarded and trigger the callback.

The x and y arguments specify the location of the upper left corner of the surface relative to the upper left corner of the parent surface, in surface-local coordinates.

val set_fullscreen : [< `V1 ] t -> method_:Wayland__Wayland_proto.Wl_shell_surface.Fullscreen_method.t -> framerate:int32 -> output:( [ `Wl_output ], 'a, [ `Client ] ) Proxy.t option -> unit

Make the surface a fullscreen surface.

Map the surface as a fullscreen surface.

If an output parameter is given then the surface will be made fullscreen on that output. If the client does not specify the output then the compositor will apply its policy - usually choosing the output on which the surface has the biggest surface area.

The client may specify a method to resolve a size conflict between the output size and the surface size - this is provided through the method parameter.

The framerate parameter is used only when the method is set to "driver", to indicate the preferred framerate. A value of 0 indicates that the client does not care about framerate. The framerate is specified in mHz, that is framerate of 60000 is 60Hz.

A method of "scale" or "driver" implies a scaling operation of the surface, either via a direct scaling operation or a change of the output mode. This will override any kind of output scaling, so that mapping a surface with a buffer size equal to the mode can fill the screen independent of buffer_scale.

A method of "fill" means we don't scale up the buffer, however any output scale is applied. This means that you may run into an edge case where the application maps a buffer with the same size of the output mode but buffer_scale 1 (thus making a surface larger than the output). In this case it is allowed to downscale the results to fit the screen.

The compositor must reply to this request with a configure event with the dimensions for the output on which the surface will be made fullscreen.

val set_transient : [< `V1 ] t -> parent:( [ `Wl_surface ], 'a, [ `Client ] ) Proxy.t -> x:int32 -> y:int32 -> flags:int32 -> unit

Make the surface a transient surface.

Map the surface relative to an existing surface.

The x and y arguments specify the location of the upper left corner of the surface relative to the upper left corner of the parent surface, in surface-local coordinates.

The flags argument controls details of the transient behaviour.

val set_toplevel : [< `V1 ] t -> unit

Make the surface a toplevel surface.

Map the surface as a toplevel surface.

A toplevel surface is not fullscreen, maximized or transient.

val resize : [< `V1 ] t -> seat:( [ `Wl_seat ], 'a, [ `Client ] ) Proxy.t -> serial:int32 -> edges:int32 -> unit

Start an interactive resize.

Start a pointer-driven resizing of the surface.

This request must be used in response to a button press event. The server may ignore resize requests depending on the state of the surface (e.g. fullscreen or maximized).

val move : [< `V1 ] t -> seat:( [ `Wl_seat ], 'a, [ `Client ] ) Proxy.t -> serial:int32 -> unit

Start an interactive move.

Start a pointer-driven move of the surface.

This request must be used in response to a button press event. The server may ignore move requests depending on the state of the surface (e.g. fullscreen or maximized).

val pong : [< `V1 ] t -> serial:int32 -> unit

Respond to a ping event.

A client must respond to a ping event with a pong request or the client may be deemed unresponsive.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.