Module Wl_shell_surface.Fullscreen_method

Different method to set the surface fullscreen.

Hints to indicate to the compositor how to deal with a conflict between the dimensions of the surface and the dimensions of the output. The compositor is free to ignore this parameter.

type t =

| Default : t

| Scale : t

| Driver : t

| Fill : t

val to_int32 : t -> int32

val of_int32 : int32 -> t