Module Wl_shell_surface.Resize

Edge values for resizing.

These values are used to indicate which edge of a surface is being dragged in a resize operation. The server may use this information to adapt its behavior, e.g. choose an appropriate cursor image.

type t = int32

val none : int32

val top : int32

val bottom : int32

val left : int32

val top_left : int32

val bottom_left : int32

val right : int32

val top_right : int32

val bottom_right : int32

val to_int32 : 'a -> 'a

val of_int32 : 'a -> 'a