Module Wl_shm.Format

Pixel formats.

This describes the memory layout of an individual pixel.

All renderers should support argb8888 and xrgb8888 but any other formats are optional and may not be supported by the particular renderer in use.

The drm format codes match the macros defined in drm_fourcc.h, except argb8888 and xrgb8888. The formats actually supported by the compositor will be reported by the format event.

type t =

| Argb8888 : t

| Xrgb8888 : t

| C8 : t

| Rgb332 : t

| Bgr233 : t

| Xrgb4444 : t

| Xbgr4444 : t

| Rgbx4444 : t

| Bgrx4444 : t

| Argb4444 : t

| Abgr4444 : t

| Rgba4444 : t

| Bgra4444 : t

| Xrgb1555 : t

| Xbgr1555 : t

| Rgbx5551 : t

| Bgrx5551 : t

| Argb1555 : t

| Abgr1555 : t

| Rgba5551 : t

| Bgra5551 : t

| Rgb565 : t

| Bgr565 : t

| Rgb888 : t

| Bgr888 : t

| Xbgr8888 : t

| Rgbx8888 : t

| Bgrx8888 : t

| Abgr8888 : t

| Rgba8888 : t

| Bgra8888 : t

| Xrgb2101010 : t

| Xbgr2101010 : t

| Rgbx1010102 : t

| Bgrx1010102 : t

| Argb2101010 : t

| Abgr2101010 : t

| Rgba1010102 : t

| Bgra1010102 : t

| Yuyv : t

| Yvyu : t

| Uyvy : t

| Vyuy : t

| Ayuv : t

| Nv12 : t

| Nv21 : t

| Nv16 : t

| Nv61 : t

| Yuv410 : t

| Yvu410 : t

| Yuv411 : t

| Yvu411 : t

| Yuv420 : t

| Yvu420 : t

| Yuv422 : t

| Yvu422 : t

| Yuv444 : t

| Yvu444 : t

| R8 : t

| R16 : t

| Rg88 : t

| Gr88 : t

| Rg1616 : t

| Gr1616 : t

| Xrgb16161616f : t

| Xbgr16161616f : t

| Argb16161616f : t

| Abgr16161616f : t

| Xyuv8888 : t

| Vuy888 : t

| Vuy101010 : t

| Y210 : t

| Y212 : t

| Y216 : t

| Y410 : t

| Y412 : t

| Y416 : t

| Xvyu2101010 : t

| Xvyu12_16161616 : t

| Xvyu16161616 : t

| Y0l0 : t

| X0l0 : t

| Y0l2 : t

| X0l2 : t

| Yuv420_8bit : t

| Yuv420_10bit : t

| Xrgb8888_a8 : t

| Xbgr8888_a8 : t

| Rgbx8888_a8 : t

| Bgrx8888_a8 : t

| Rgb888_a8 : t

| Bgr888_a8 : t

| Rgb565_a8 : t

| Bgr565_a8 : t

| Nv24 : t

| Nv42 : t

| P210 : t

| P010 : t

| P012 : t

| P016 : t

| Axbxgxrx106106106106 : t

| Nv15 : t

| Q410 : t

| Q401 : t

val to_int32 : t -> int32

val of_int32 : int32 -> t