Module Wayland_server.Wl_callback

Callback object.

Clients can handle the 'done' event to get notified when the related request is done.

type 'v t = ([ `Wl_callback ], 'v, [ `Server ]) Proxy.t

Version 1, 2, 3, 4

val done_ : [< `V1 | `V2 | `V3 | `V4 ] t -> callback_data:int32 -> unit

Done event.

Notify the client when the related request is done.

class +'a handlers : object ... end