Module Wayland_server.Wl_callback

Callback object.

Clients can handle the 'done' event to get notified when the related request is done.

type 'v t = ([ `Wl_callback ], 'v, [ `Server ]) Proxy.t

Version 1, 2, 3, 4

val done_ : [< `V1 | `V2 | `V3 | `V4 ] t -> callback_data:int32 -> unit

Done event.

Notify the client when the related request is done.

Handlers

Note: Servers will always want to use v1.

class +'a v1 : object ... end

Handler for a proxy with version >= 1.

class +'a v2 : object ... end

Handler for a proxy with version >= 2.

class +'a v3 : object ... end

Handler for a proxy with version >= 3.

class +'a v4 : object ... end

Handler for a proxy with version >= 4.