Module Wayland_server.Wl_shell_surface

Desktop-style metadata interface.

An interface that may be implemented by a wl_surface, for implementations that provide a desktop-style user interface.

It provides requests to treat surfaces like toplevel, fullscreen or popup windows, move, resize or maximize them, associate metadata like title and class, etc.

On the server side the object is automatically destroyed when the related wl_surface is destroyed. On the client side, wl_shell_surface_destroy() must be called before destroying the wl_surface object.

type 'v t = ([ `Wl_shell_surface ], 'v) Proxy.t

Version 1

val popup_done : 'v t -> unit

Popup interaction is done.

The popup_done event is sent out when a popup grab is broken, that is, when the user clicks a surface that doesn't belong to the client owning the popup surface.

val configure : 'v t -> edges:int32 -> width:int32 -> height:int32 -> unit

Suggest resize.

The configure event asks the client to resize its surface.

The size is a hint, in the sense that the client is free to ignore it if it doesn't resize, pick a smaller size (to satisfy aspect ratio or resize in steps of NxM pixels).

The edges parameter provides a hint about how the surface was resized. The client may use this information to decide how to adjust its content to the new size (e.g. a scrolling area might adjust its content position to leave the viewable content unmoved).

The client is free to dismiss all but the last configure event it received.

The width and height arguments specify the size of the window in surface-local coordinates.

val ping : 'v t -> serial:int32 -> unit

Ping client.

Ping a client to check if it is receiving events and sending requests. A client is expected to reply with a pong request.

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Wl_shell_surface ], 'v, [< `V1 ], [> `V1 ]) Proxy.handler