Module Wayland_client.Wl_shell

Create desktop-style surfaces.

This interface is implemented by servers that provide desktop-style user interfaces.

It allows clients to associate a wl_shell_surface with a basic surface.

Note! This protocol is deprecated and not intended for production use. For desktop-style user interfaces, use xdg_shell.

type 'v t = ([ `Wl_shell ], 'v, [ `Client ]) Wayland.Proxy.t

module Error = Wayland.Wayland_proto.Wl_shell.Error

Version 1

val get_shell_surface : [< `V1 ] as 'a t -> [[ `Wl_shell_surface ], 'a, [ `Client ]] Wayland.Proxy.Handler.t -> surface:([ `Wl_surface ], 'b, [ `Client ]) Wayland.Proxy.t -> ([ `Wl_shell_surface ], 'a, [ `Client ]) Wayland.Proxy.t

Create a shell surface from a surface.

Create a shell surface for an existing surface. This gives the wl_surface the role of a shell surface. If the wl_surface already has another role, it raises a protocol error.

Only one shell surface can be associated with a given surface.

class +'a _handlers_unsafe : object ... end

Requesting a minimum version

class v1 : object ... end