Module Wl_data_device_manager.Dnd_action

Drag and drop actions.

This is a bitmask of the available/preferred actions in a drag-and-drop operation.

In the compositor, the selected action is a result of matching the actions offered by the source and destination sides. "action" events with a "none" action will be sent to both source and destination if there is no match. All further checks will effectively happen on (source actions ∩ destination actions).

In addition, compositors may also pick different actions in reaction to key modifiers being pressed. One common design that is used in major toolkits (and the behavior recommended for compositors) is:

If no modifiers are pressed, the first match (in bit order) will be used.

Pressing Shift selects "move", if enabled in the mask.

Pressing Control selects "copy", if enabled in the mask.

Behavior beyond that is considered implementation-dependent. Compositors may for example bind other modifiers (like Alt/Meta) or drags initiated with other buttons than BTN_LEFT to specific actions (e.g. "ask").

type t = int32

val none : int32

val copy : int32

val move : int32

val ask : int32

val to_int32 : 'a -> 'a

val of_int32 : 'a -> 'a