Module Wayland_server.Wl_shm

Shared memory support.

A singleton global object that provides support for shared memory.

Clients can create wl_shm_pool objects using the create_pool request.

At connection setup time, the wl_shm object emits one or more format events to inform clients about the valid pixel formats that can be used for buffers.

type 'v t = ([ `Wl_shm ], 'v, [ `Server ]) Wayland.Proxy.t

module Error = Wayland.Wayland_proto.Wl_shm.Error

module Format = Wayland.Wayland_proto.Wl_shm.Format

Version 1

val format : [< `V1 ] t -> format:$Imports.Wl_shm.Format.t -> unit

Pixel format description.

Informs the client about a valid pixel format that can be used for buffers. Known formats include argb8888 and xrgb8888.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.