Class Gtk_primary_selection_device.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_set_selection : [> `V1 ] t -> source: ([ `Gtk_primary_selection_source ], [> {Imports}1.Gtk_primary_selection_source.versions ], [ `Server ]) {Proxy}2.t option -> serial:int32 -> unit

Set the primary selection.

Replaces the current selection. The previous owner of the primary selection will receive a wp_primary_selection_source.cancelled event.

To unset the selection, set the source to NULL.

method private virtual on_destroy : [> `V1 ] t -> unit

Destroy the primary selection device.

Destroy the primary selection device.

method min_version : int32