Module Gtk_primary_selection_server.Gtk_primary_selection_device_manager

X primary selection emulation.

The primary selection device manager is a singleton global object that provides access to the primary selection. It allows to create wp_primary_selection_source objects, as well as retrieving the per-seat wp_primary_selection_device objects.

type 'v t = ([ `Gtk_primary_selection_device_manager ], 'v, [ `Server ]) Wayland.Proxy.t

Version 1

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.