Module Linux_dmabuf_unstable_v1_server.Zwp_linux_buffer_params_v1

Parameters for creating a dmabuf-based wl_buffer.

This temporary object is a collection of dmabufs and other parameters that together form a single logical buffer. The temporary object may eventually create one wl_buffer unless cancelled by destroying it before requesting 'create'.

Single-planar formats only require one dmabuf, however multi-planar formats may require more than one dmabuf. For all formats, an 'add' request must be called once per plane (even if the underlying dmabuf fd is identical).

You must use consecutive plane indices ('plane_idx' argument for 'add') from zero to the number of planes used by the drm_fourcc format code. All planes required by the format must be given exactly once, but can be given in any order. Each plane index can be set only once.

type 'v t = ([ `Zwp_linux_buffer_params_v1 ], 'v, [ `Server ]) Wayland.Proxy.t

module Error = Linux_dmabuf_unstable_v1_proto.Zwp_linux_buffer_params_v1.Error

module Flags = Linux_dmabuf_unstable_v1_proto.Zwp_linux_buffer_params_v1.Flags

Version 1

val failed : [< `V1 | `V2 | `V3 | `V4 ] t -> unit

Buffer creation failed.

This event indicates that the attempted buffer creation has failed. It usually means that one of the dmabuf constraints has not been fulfilled.

Upon receiving this event, the client should destroy the zwp_linux_buffer_params_v1 object.

val created : [< `V1 | `V2 | `V3 | `V4 ] as 'a t -> [[ `Wl_buffer ], 'b, [ `Server ]] Wayland.Proxy.Handler.t -> ([ `Wl_buffer ], 'c, [ `Server ]) Wayland.Proxy.t

Buffer creation succeeded.

This event indicates that the attempted buffer creation was successful. It provides the new wl_buffer referencing the dmabuf(s).

Upon receiving this event, the client should destroy the zwp_linux_buffer_params_v1 object.

Version 2, 3, 4

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.

class virtual +'a v4 : object ... end

Handler for a proxy with version >= 4.