Module Server_decoration_server.Org_kde_kwin_server_decoration

type 'v t = ([ `Org_kde_kwin_server_decoration ], 'v, [ `Server ]) Wayland.Proxy.t

module Mode = Server_decoration_proto.Org_kde_kwin_server_decoration.Mode

Version 1

val mode : [< `V1 ] t -> mode:int32 -> unit

The new decoration mode applied by the server.

This event is emitted directly after the decoration is created and represents the base decoration policy by the server. E.g. a server which wants all surfaces to be client-side decorated will send Client, a server which wants server-side decoration will send Server.

The client can request a different mode through the decoration request. The server will acknowledge this by another event with the same mode. So even if a server prefers server-side decoration it's possible to force a client-side decoration.

The server may emit this event at any time. In this case the client can again request a different mode. It's the responsibility of the server to prevent a feedback loop.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.