Module Xdg_decoration_unstable_v1_client.Zxdg_decoration_manager_v1

Window decoration manager.

This interface allows a compositor to announce support for server-side decorations.

A window decoration is a set of window controls as deemed appropriate by the party managing them, such as user interface components used to move, resize and change a window's state.

A client can use this protocol to request being decorated by a supporting compositor.

If compositor and client do not negotiate the use of a server-side decoration using this protocol, clients continue to self-decorate as they see fit.

Warning! The protocol described in this file is experimental and backward incompatible changes may be made. Backward compatible changes may be added together with the corresponding interface version bump. Backward incompatible changes are done by bumping the version number in the protocol and interface names and resetting the interface version. Once the protocol is to be declared stable, the 'z' prefix and the version number in the protocol and interface names are removed and the interface version number is reset.

type 'v t = ([ `Zxdg_decoration_manager_v1 ], 'v, [ `Client ]) Proxy.t

Version 1

val get_toplevel_decoration : [< `V1 ] as 'a t -> [[ `Zxdg_toplevel_decoration_v1 ], 'a, [ `Client ]] Proxy.Handler.t -> toplevel:([ `Xdg_toplevel ], 'b, [ `Client ]) Proxy.t -> ([ `Zxdg_toplevel_decoration_v1 ], 'a, [ `Client ]) Proxy.t

Create a new toplevel decoration object.

Create a new decoration object associated with the given toplevel.

Creating an xdg_toplevel_decoration from an xdg_toplevel which has a buffer attached or committed is a client error, and any attempts by a client to attach or manipulate a buffer prior to the first xdg_toplevel_decoration.configure event must also be treated as errors.

val destroy : [< `V1 ] t -> unit

Destroy the decoration manager object.

Destroy the decoration manager. This doesn't destroy objects created with the manager.

class +'a _handlers_unsafe : object ... end

Requesting a minimum version

class v1 : object ... end