Class Zxdg_decoration_manager_v1.v1

Handler for a proxy with version >= 1.

constraint 'a = [< `V1 ]

method private virtual on_destroy : [> `V1 ] t -> unit

Destroy the decoration manager object.

Destroy the decoration manager. This doesn't destroy objects created with the manager.

method private virtual on_get_toplevel_decoration : [> `V1 ] t -> ( [ `Zxdg_toplevel_decoration_v1 ], 'b, [ `Server ] ) Wayland.Proxy.t -> toplevel: ( [ `Xdg_toplevel ], [> Wayland_protocols__Xdg_shell_proto.Xdg_toplevel.versions ], [ `Server ] ) Wayland.Proxy.t -> unit

Create a new toplevel decoration object.

Create a new decoration object associated with the given toplevel.

Creating an xdg_toplevel_decoration from an xdg_toplevel which has a buffer attached or committed is a client error, and any attempts by a client to attach or manipulate a buffer prior to the first xdg_toplevel_decoration.configure event must also be treated as errors.

method min_version : 'a v1 -> int32

method bind_version : 'c -> [ `V1 ]