Module Xdg_decoration_unstable_v1_server.Zxdg_toplevel_decoration_v1

Decoration object for a toplevel surface.

The decoration object allows the compositor to toggle server-side window decorations for a toplevel surface. The client can request to switch to another mode.

The xdg_toplevel_decoration object must be destroyed before its xdg_toplevel.

type 'v t = ([ `Zxdg_toplevel_decoration_v1 ], 'v, [ `Server ]) $Proxy.t

module Error = Xdg_decoration_unstable_v1_proto.Zxdg_toplevel_decoration_v1.Error

module Mode = Xdg_decoration_unstable_v1_proto.Zxdg_toplevel_decoration_v1.Mode

Version 1

val configure : [< `V1 ] t -> mode:$Imports.Zxdg_toplevel_decoration_v1.Mode.t -> unit

Suggest a surface change.

The configure event asks the client to change its decoration mode. The configured state should not be applied immediately. Clients must send an ack_configure in response to this event. See xdg_surface.configure and xdg_surface.ack_configure for details.

A configure event can be sent at any time. The specified mode must be obeyed by the client.

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.