Module Xdg_toplevel.State

Types of state on the surface.

The different state values used on the surface. This is designed for state values like maximized, fullscreen. It is paired with the configure event to ensure that both the client and the compositor setting the state can be synchronized.

States set in this way are double-buffered. They will get applied on the next commit.

type t =

| Maximized : t	(*
The surface is maximized.

The surface is maximized. The window geometry specified in the configure event must be obeyed by the client.

The client should draw without shadow or other decoration outside of the window geometry.
*)

| Fullscreen : t	(*
The surface is fullscreen.

The surface is fullscreen. The window geometry specified in the configure event is a maximum; the client cannot resize beyond it. For a surface to cover the whole fullscreened area, the geometry dimensions must be obeyed by the client. For more details, see xdg_toplevel.set_fullscreen.
*)

| Resizing : t	(*
The surface is being resized.

The surface is being resized. The window geometry specified in the configure event is a maximum; the client cannot resize beyond it. Clients that have aspect ratio or cell sizing configuration can use a smaller size, however.
*)

| Activated : t	(*
The surface is now activated.

Client window decorations should be painted as if the window is active. Do not assume this means that the window actually has keyboard or pointer focus.
*)

| Tiled_left : t	(*
The surface is tiled.

The window is currently in a tiled layout and the left edge is considered to be adjacent to another part of the tiling grid.
*)

| Tiled_right : t	(*
The surface is tiled.

The window is currently in a tiled layout and the right edge is considered to be adjacent to another part of the tiling grid.
*)

| Tiled_top : t	(*
The surface is tiled.

The window is currently in a tiled layout and the top edge is considered to be adjacent to another part of the tiling grid.
*)

| Tiled_bottom : t	(*
The surface is tiled.

The window is currently in a tiled layout and the bottom edge is considered to be adjacent to another part of the tiling grid.
*)

val to_int32 : t -> int32

val of_int32 : int32 -> t