Module Xdg_shell_server.Xdg_positioner

Child surface positioner.

The xdg_positioner provides a collection of rules for the placement of a child surface relative to a parent surface. Rules can be defined to ensure the child surface remains within the visible area's borders, and to specify how the child surface changes its position, such as sliding along an axis, or flipping around a rectangle. These positioner-created rules are constrained by the requirement that a child surface must intersect with or be at least partially adjacent to its parent surface.

See the various requests for details about possible rules.

At the time of the request, the compositor makes a copy of the rules specified by the xdg_positioner. Thus, after the request is complete the xdg_positioner object can be destroyed or reused; further changes to the object will have no effect on previous usages.

For an xdg_positioner object to be considered complete, it must have a non-zero size set by set_size, and a non-zero anchor rectangle set by set_anchor_rect. Passing an incomplete xdg_positioner object when positioning a surface raises an invalid_positioner error.

type 'v t = ([ `Xdg_positioner ], 'v, [ `Server ]) Wayland.Proxy.t

module Error = Xdg_shell_proto.Xdg_positioner.Error

module Anchor = Xdg_shell_proto.Xdg_positioner.Anchor

module Gravity = Xdg_shell_proto.Xdg_positioner.Gravity

module Constraint_adjustment = Xdg_shell_proto.Xdg_positioner.Constraint_adjustment

Version 1, 2

Version 3, 4, 5, 6

Handlers

Note: Servers will always want to use v1.

class virtual +'a v1 : object ... end

Handler for a proxy with version >= 1.

class virtual +'a v2 : object ... end

Handler for a proxy with version >= 2.

class virtual +'a v3 : object ... end

Handler for a proxy with version >= 3.

class virtual +'a v4 : object ... end

Handler for a proxy with version >= 4.

class virtual +'a v5 : object ... end

Handler for a proxy with version >= 5.

class virtual +'a v6 : object ... end

Handler for a proxy with version >= 6.