Module Xdg_shell_server.Xdg_wm_base

Create desktop-style surfaces.

The xdg_wm_base interface is exposed as a global object enabling clients to turn their wl_surfaces into windows in a desktop environment. It defines the basic functionality needed for clients and the compositor to create windows that can be dragged, resized, maximized, etc, as well as creating transient windows such as popup menus.

type 'v t = ([ `Xdg_wm_base ], 'v) Proxy.t

Version 1, 2, 3

val ping : 'v t -> serial:int32 -> unit

Check if the client is alive.

The ping event asks the client if it's still alive. Pass the serial specified in the event back to the compositor by sending a "pong" request back with the specified serial. See xdg_wm_base.pong.

Compositors can use this to determine if the client is still alive. It's unspecified what will happen if the client doesn't respond to the ping request, or in what timeframe. Clients should try to respond in a reasonable amount of time.

A compositor is free to ping in any way it wants, but a client must always respond to any xdg_wm_base object it created.

class virtual 'v h1 : object ... end

val v1 : 'v h1 -> ([ `Xdg_wm_base ], 'v, [< `V1 ], [> `V1 ]) Proxy.handler

val v2 : 'v h1 -> ([ `Xdg_wm_base ], 'v, [< `V1 | `V2 ], [> `V2 ]) Proxy.handler

val v3 : 'v h1 -> ([ `Xdg_wm_base ], 'v, [< `V1 | `V2 | `V3 ], [> `V3 ]) Proxy.handler